% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute_angle.R
\name{absolute_angle}
\alias{absolute_angle}
\title{Return the smallest possible angle between two compass bearings}
\usage{
absolute_angle(from, to)
}
\arguments{
\item{from}{The start bearing, counted clockwise from North}

\item{to}{The end bearing, counted clockwise from North}
}
\value{
The smallest possible angle between the two bearings as decimal degrees
within \verb{[0..360[}.
}
\description{
Return the smallest possible angle between two compass bearings
}
\examples{
testthat::expect_equal(absolute_angle(NA, 10), NA)
testthat::expect_equal(absolute_angle(10, NA), NA)
testthat::expect_equal(absolute_angle(0, 20), 20)
testthat::expect_equal(absolute_angle(20, 0), -20)
testthat::expect_equal(absolute_angle(10, 0), -10)
testthat::expect_equal(absolute_angle(20, 60), 40)
testthat::expect_equal(absolute_angle(350, 10), 20)
testthat::expect_equal(absolute_angle(10, 350), -20)
}
\seealso{
Other helpers: 
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
