% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.R
\name{add_dates}
\alias{add_dates}
\title{Add turtle dates to a dataframe with datetime col `observation_start_time`.}
\usage{
add_dates(data)
}
\arguments{
\item{data}{a dataframe with datetime col `observation_start_time`, e.g. the
output of ODKC form "turtle track or nest", "predator or disturbance", or
"marine wildlife incident".}
}
\value{
The initial dataframe plus new columns

  * calendar_date_awst
  * turtle_date
  * season
  * season_week
  * iso_week
}
\description{
\lifecycle{stable}
}
