% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.R
\name{add_dates_sve}
\alias{add_dates_sve}
\title{Add turtle dates to a dataframe with datetime col `survey_end_time`.}
\usage{
add_dates_sve(data)
}
\arguments{
\item{data}{a dataframe with datetime col `survey_end_time`, e.g. the
output of ODKC form "site visit end".}
}
\value{
The initial dataframe plus new columns

  * calendar_date_awst
  * turtle_date
  * season
  * season_week
  * iso_week
}
\description{
\lifecycle{stable}
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
