% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks_odkc.R
\name{add_hatching_emergence_success_odkc}
\alias{add_hatching_emergence_success_odkc}
\title{Add calculated egg count, hatching and emergence success to
\code{odkc_data$tracks_egg}.}
\usage{
add_hatching_emergence_success_odkc(data)
}
\arguments{
\item{data}{ODKC Track or Nest data annotated with egg counts, e.g.
\code{odkc_data$tracks_egg}.}
}
\description{
Hatching success as percentage (0..100) after Miller 1999:
Hatching success = 100 * no_egg_shells / (
no_egg_shells + no_undeveloped_eggs + no_unhatched_eggs +
no_unhatched_term + no_depredated_eggs)
}
\details{
Emergence success as percentage (0..100) after Miller 1999:
Emergence success = 100 *
(no_egg_shells - no_live_hatchlings - no_dead_hatchlings) / (
no_egg_shells + no_undeveloped_eggs + no_unhatched_eggs +
no_unhatched_term + no_depredated_eggs)
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{hatching_emergence_success_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{map_tt_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
