% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nest_labels.R
\name{add_nest_labels}
\alias{add_nest_labels}
\title{Add labels for species and nest/track type to parsed \code{turtle-nest-encounters}.}
\usage{
add_nest_labels(nests)
}
\arguments{
\item{nests}{(tibble) A dataframe or tibble of parsed
\code{turtle-nest-encounters}.}
}
\value{
The dataframe with two new columns, \code{species_colours} and
  \code{nest_type_text}.
}
\description{
Adds two columns, two new columns, \code{species_colours} and
\code{nest_type_text}, to parsed \code{turtle-nest-encounters} containing
\code{WAStD} \code{species} and \code{nest_type}. The new columns can be used
to create \code{leaflet} markers with
\code{leaflet::makeAwesomeIcon(
  text = ~nest_type_text, markerColor = ~species_colours)}.
}
