% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nest_labels.R
\name{add_nest_labels_odkc}
\alias{add_nest_labels_odkc}
\title{Add labels for species and nest/track type to parsed
\code{turtle-nest-encounters} from ODK Central.}
\usage{
add_nest_labels_odkc(nests)
}
\arguments{
\item{nests}{(tibble) A dataframe or tibble of parsed
\code{turtle-nest-encounters}.}
}
\value{
The dataframe with two new columns, \code{species_colours} and
  \code{nest_type_text}.
}
\description{
Adds two columns, two new columns, \code{species_colours} and
\code{nest_type_text}, to parsed \code{turtle-nest-encounters} containing
\code{WAStD} \code{species} and \code{nest_type}. The new columns can be used
to create \code{leaflet} markers with
\code{leaflet::makeAwesomeIcon(
  text = ~nest_type_text, markerColor = ~species_colours)}.
}
\seealso{
Other odkc: 
\code{\link{disturbance_by_season_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_dead_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{filter_predation_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
