% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_encounters}
\alias{animal_encounters}
\title{AnimalEncounter WAStD API response}
\format{\preformatted{List of 3
 $ content   :List of 10
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-255155"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4482
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2013 Tagging "
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "public_report"
  .. .. .. ..$ name    : chr "Public Report"
  .. .. .. ..$ email   : chr ""
  .. .. .. ..$ phone   : NULL
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2013-08-04T02:00:00Z"
  .. .. ..$ name                     : NULL
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "eretmochelys-imbricata"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "floating"
  .. .. ..$ nesting_event            : chr "na"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4482/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WA14951"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "removed"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WA14772"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "removed"
  .. .. .. .. ..$ comments        : chr "Tags returnedNA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267172"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4647
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7309"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-23T14:17:00Z"
  .. .. ..$ name                     : chr "WB7309"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "beach-edge-of-vegetation"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4647/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7310"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7309"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267173"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4648
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7351"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-24T16:25:00Z"
  .. .. ..$ name                     : chr "WB7351"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4648/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7352"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7351"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267175"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4649
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7369"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-25T13:40:00Z"
  .. .. ..$ name                     : chr "WB7369"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4649/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7370"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7369"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267176"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4650
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7335"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-25T14:10:00Z"
  .. .. ..$ name                     : chr "WB7335"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4650/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7336"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7335"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267177"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4651
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7309"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-26T10:25:00Z"
  .. .. ..$ name                     : chr "WB7309"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : chr ""
  .. .. ..$ habitat                  : chr "beach-edge-of-vegetation"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "present"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4651/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7310"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7309"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267178"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4652
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7335"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-26T10:43:00Z"
  .. .. ..$ name                     : chr "WB7335"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "beach-above-high-water"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "present"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4652/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7336"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7335"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267181"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4655
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7382"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-26T15:35:00Z"
  .. .. ..$ name                     : chr "WB7382"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : NULL
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "na"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4655/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7433"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7382"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267182"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4656
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7337"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-26T15:40:00Z"
  .. .. ..$ name                     : chr "WB7337"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : chr ""
  .. .. ..$ habitat                  : chr "na"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4656/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7338"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7337"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  ..$ :List of 4
  .. ..$ id        : chr "wamtram-observation-id-267386"
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 31
  .. .. ..$ pk                       : int 4842
  .. .. ..$ site_visit               : NULL
  .. .. ..$ source                   : chr "wamtram"
  .. .. ..$ encounter_type           : chr "tagging"
  .. .. ..$ leaflet_title            : chr "2016 Tagging WB7311"
  .. .. ..$ status                   : chr "new"
  .. .. ..$ observer                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ reporter                 :List of 4
  .. .. .. ..$ username: chr "admin"
  .. .. .. ..$ name    : chr "Florian Mayer (Admin)"
  .. .. .. ..$ email   : chr "florian.wendelin.mayer@gmail.com"
  .. .. .. ..$ phone   : chr "+61414239161"
  .. .. ..$ latitude                 : num -21.5
  .. .. ..$ longitude                : num 115
  .. .. ..$ crs                      : chr "WGS 84"
  .. .. ..$ location_accuracy        : chr "10"
  .. .. ..$ when                     : chr "2016-11-27T12:08:00Z"
  .. .. ..$ name                     : chr "WB7311"
  .. .. ..$ taxon                    : chr "Cheloniidae"
  .. .. ..$ species                  : chr "natator-depressus"
  .. .. ..$ health                   : chr "na"
  .. .. ..$ sex                      : chr "female"
  .. .. ..$ behaviour                : chr ""
  .. .. ..$ habitat                  : chr "in-dune-vegetation"
  .. .. ..$ activity                 : chr "returning-to-water"
  .. .. ..$ nesting_event            : chr "absent"
  .. .. ..$ checked_for_injuries     : chr "na"
  .. .. ..$ scanned_for_pit_tags     : chr "na"
  .. .. ..$ checked_for_flipper_tags : chr "na"
  .. .. ..$ cause_of_death           : chr "na"
  .. .. ..$ cause_of_death_confidence: chr "na"
  .. .. ..$ absolute_admin_url       : chr "/admin/observations/animalencounter/4842/change/"
  .. .. ..$ photographs              : list()
  .. .. ..$ tx_logs                  : list()
  .. .. ..$ observation_set          :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7312"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-right-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "flipper-tag"
  .. .. .. .. ..$ name            : chr "WB7311"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left-1"
  .. .. .. .. ..$ status          : chr "resighted"
  .. .. .. .. ..$ comments        : chr "NANA"
 $ serializer: chr "animal-encounters"
 $ response  :List of 10
  ..$ url        : chr "https://strandings.dpaw.wa.gov.au/api/1/animal-encounters/?taxon=Cheloniidae&limit=10&format=json&observer=1"
  ..$ status_code: int 200
  ..$ headers    :List of 13
  .. ..$ date                     : chr "Mon, 01 May 2017 23:21:00 GMT"
  .. ..$ content-type             : chr "application/json"
  .. ..$ transfer-encoding        : chr "chunked"
  .. ..$ connection               : chr "keep-alive"
  .. ..$ server                   : chr "nginx"
  .. ..$ x-xss-protection         : chr "1; mode=block"
  .. ..$ x-content-type-options   : chr "nosniff"
  .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. ..$ vary                     : chr "Accept, Cookie"
  .. ..$ allow                    : chr "GET, POST, OPTIONS"
  .. ..$ x-frame-options          : chr "DENY"
  .. ..$ x-ua-compatible          : chr "chrome=1"
  .. ..$ content-encoding         : chr "gzip"
  .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  ..$ all_headers:List of 2
  .. ..$ :List of 3
  .. .. ..$ status : int 301
  .. .. ..$ version: chr "HTTP/1.1"
  .. .. ..$ headers:List of 11
  .. .. .. ..$ date                     : chr "Mon, 01 May 2017 23:20:59 GMT"
  .. .. .. ..$ content-type             : chr "text/html; charset=utf-8"
  .. .. .. ..$ transfer-encoding        : chr "chunked"
  .. .. .. ..$ connection               : chr "keep-alive"
  .. .. .. ..$ server                   : chr "nginx"
  .. .. .. ..$ x-xss-protection         : chr "1; mode=block"
  .. .. .. ..$ x-content-type-options   : chr "nosniff"
  .. .. .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. .. .. ..$ location                 : chr "/api/1/animal-encounters/?taxon=Cheloniidae&limit=10&format=json&observer=1"
  .. .. .. ..$ x-frame-options          : chr "DENY"
  .. .. .. ..$ x-ua-compatible          : chr "chrome=1"
  .. .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  .. ..$ :List of 3
  .. .. ..$ status : int 200
  .. .. ..$ version: chr "HTTP/1.1"
  .. .. ..$ headers:List of 13
  .. .. .. ..$ date                     : chr "Mon, 01 May 2017 23:21:00 GMT"
  .. .. .. ..$ content-type             : chr "application/json"
  .. .. .. ..$ transfer-encoding        : chr "chunked"
  .. .. .. ..$ connection               : chr "keep-alive"
  .. .. .. ..$ server                   : chr "nginx"
  .. .. .. ..$ x-xss-protection         : chr "1; mode=block"
  .. .. .. ..$ x-content-type-options   : chr "nosniff"
  .. .. .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. .. .. ..$ vary                     : chr "Accept, Cookie"
  .. .. .. ..$ allow                    : chr "GET, POST, OPTIONS"
  .. .. .. ..$ x-frame-options          : chr "DENY"
  .. .. .. ..$ x-ua-compatible          : chr "chrome=1"
  .. .. .. ..$ content-encoding         : chr "gzip"
  .. .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  ..$ cookies    :'data.frame':	0 obs. of  7 variables:
  .. ..$ domain    : logi(0) 
  .. ..$ flag      : logi(0) 
  .. ..$ path      : logi(0) 
  .. ..$ secure    : logi(0) 
  .. ..$ expiration:Classes 'POSIXct', 'POSIXt'  num(0) 
  .. ..$ name      : logi(0) 
  .. ..$ value     : logi(0) 
  ..$ content    : raw [1:14181] 7b 22 74 79 ...
  ..$ date       : POSIXct[1:1], format: "2017-05-01 23:21:00"
  ..$ times      : Named num [1:6] 0.062893 0.000055 0.00006 0.000167 0.230716 ...
  .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
  ..$ request    :List of 7
  .. ..$ method    : chr "GET"
  .. ..$ url       : chr "https://strandings.dpaw.wa.gov.au/api/1/animal-encounters?taxon=Cheloniidae&limit=10&format=json&observer=1"
  .. ..$ headers   : Named chr [1:2] "application/json, text/xml, application/xml, */*" "Token c5575c09f6a0171668f31d5dd5013f02658668bd"
  .. .. ..- attr(*, "names")= chr [1:2] "Accept" "Authorization"
  .. ..$ fields    : NULL
  .. ..$ options   :List of 2
  .. .. ..$ useragent: chr "http://github.com/parksandwildlife/turtle-scripts"
  .. .. ..$ httpget  : logi TRUE
  .. ..$ auth_token: NULL
  .. ..$ output    : list()
  .. .. ..- attr(*, "class")= chr [1:2] "write_memory" "write_function"
  .. ..- attr(*, "class")= chr "request"
  ..$ handle     :Class 'curl_handle' <externalptr> 
  ..- attr(*, "class")= chr "response"
 - attr(*, "class")= chr "wastd_api_response"
}}
\source{
https://strandings.dpaw.wa.gov.au/api/1/animal-encounters/?taxon=Cheloniidae&limit=10&format=json
}
\usage{
animal_encounters
}
\description{
This API response is parsed into wastr's data "tags"
}
\examples{
# Prove that animal_encounters parses to animals

library(dplyr)
data("animal_encounters")
data("animals")
fresh_animals <- parse_animal_encounters(animal_encounters)
testthat::expect_equal(nrow(fresh_animals), nrow(animals))
# Compare pickled and fresh animals excluding list columns (like obs)
testthat::expect_equal(fresh_animals \%>\% dplyr::select(-obs),
                       animals \%>\% dplyr::select(-obs))
}
\keyword{datasets}

