% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_user_mapping.R
\name{annotate_user_mapping}
\alias{annotate_user_mapping}
\title{Annotate the output of make_user_mapping with Markdown links to WAStD Users.}
\usage{
annotate_user_mapping(user_mapping)
}
\arguments{
\item{user_mapping}{The output of \code{\link{user_mapping}}}
}
\value{
The annotated user_mapping tibble
}
\description{
Annotate the output of make_user_mapping with Markdown links to WAStD Users.
}
