% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wastd_api.R
\name{check_wastd_api}
\alias{check_wastd_api}
\title{Check whether API token is set and WAStD is online, else abort}
\usage{
check_wastd_api()
}
\description{
Check whether API token is set and WAStD is online, else abort
}
