% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chr2int.R
\name{chr2int}
\alias{chr2int}
\title{Return multiple lists of int as chr as one list of int and discard NA.}
\usage{
chr2int(data)
}
\arguments{
\item{data}{One or several lists of integer numbers as character}
}
\value{
One list containing just the integer numbers
}
\description{
Return multiple lists of int as chr as one list of int and discard NA.
}
\examples{
list("NA,2,234,NA", "NA", "NA,NA,1") \%>\% chr2int()
list("NA,2,234,NA", "NA", "5,NA,NA,1") \%>\% chr2int()
}
