% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_as_isoweek.R
\name{datetime_as_isoweek}
\alias{datetime_as_isoweek}
\title{Calculate the isoweek from a given datetime}
\usage{
datetime_as_isoweek(datetime)
}
\arguments{
\item{datetime}{(dttm) A datetime}
}
\value{
The ISO week as int, e.g. 26
}
\description{
Calculate the isoweek from a given datetime
}
\details{
Return the isoweek of a given datetime.

The isoweek is useful as a grouping variable for weekly summaries.
}
\examples{
# These datetimes are turtle season 2017:
datetime_as_isoweek(httpdate_as_gmt08("2017-06-30T15:59:59Z")) # 26
datetime_as_isoweek(httpdate_as_gmt08("2017-06-30T16:00:00Z")) # 26
datetime_as_isoweek(httpdate_as_gmt08("2017-08-30T06:38:43Z")) # 35
datetime_as_isoweek(httpdate_as_gmt08("2017-11-01T22:00:00Z")) # 44
datetime_as_isoweek(httpdate_as_gmt08("2018-11-01T22:00:00Z")) # 44
}
