% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_as_season.R
\name{datetime_as_season}
\alias{datetime_as_season}
\title{Calculate the "turtle season" (FY) from a given datetime}
\usage{
datetime_as_season(datetime)
}
\arguments{
\item{datetime}{(dttm) A datetime}
}
\value{
The season as int, e.g. 2017
}
\description{
Calculate the "turtle season" (FY) from a given datetime
}
\details{
Return the earliest year of the season start.

In practice, the turtle date is the date component (year, month, day) of the
datetime minus 12 hours.
}
\examples{
# These datetimes are turtle season 2017:
datetime_as_season(httpdate_as_gmt08("2017-06-30T15:59:59Z")) # 2016
datetime_as_season(httpdate_as_gmt08("2017-06-30T16:00:00Z")) # 2017
datetime_as_season(httpdate_as_gmt08("2017-08-30T06:38:43Z")) # 2017
}
