% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_as_turtle_date.R
\name{datetime_as_turtle_date}
\alias{datetime_as_turtle_date}
\title{Calculate the "turtle date" from a given datetime}
\usage{
datetime_as_turtle_date(datetime)
}
\arguments{
\item{datetime}{(dttm) A datetime}
}
\value{
A POSIXct date
}
\description{
Calculate the "turtle date" from a given datetime
}
\details{
Return the actual date for afternoon and evening datetimes
(noon to midnight) or the date of the previous day for morning datetimes
(midnight to noon).
The turtle date keeps observations made between sunset and sunrise,
as well as the "morning after" together by assigning the date of the start
of observations.

In practice, the turtle date is the date component (year, month, day) of the
datetime minus 12 hours.

This date convention is widely used in the turtle research community.
}
\examples{
# noon AWST is turtle date "2016-11-20":
datetime_as_turtle_date(
  httpdate_as_gmt08("2016-11-20T04:00:00Z")
)

# 1 sec before noon AWST is turtle date "2016-11-19":
datetime_as_turtle_date(
  httpdate_as_gmt08_turtle_date("2016-11-21T03:59:59Z")
)
}
