% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{disturbance_by_season}
\alias{disturbance_by_season}
\title{Tally WAStD disturbances by season, cause, and encounter type}
\usage{
disturbance_by_season(value)
}
\arguments{
\item{value}{The ouput of
\code{
\link{wastd_GET}("turtle-nest-disturbance-observations") %>%
\link{parse_encounterobservations}()
} or \code{data("wastd_data"); wastd_data$nest_dist}}
}
\value{
A tibble with columns `season`, `disturbance_cause`,
 `encounter_type`, and a tally `n`.
}
\description{
Tally WAStD disturbances by season, cause, and encounter type
}
\examples{
data("wastd_data")
wastd_data$nest_dist \%>\% disturbance_by_season()
}
\concept{wastd}
