% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{disturbance_by_season_odkc}
\alias{disturbance_by_season_odkc}
\title{Summarise disturbance by season and cause for ODKC data}
\usage{
disturbance_by_season_odkc(value)
}
\arguments{
\item{value}{The ouput of \code{wastd_GET("disturbance-observations") %>%
parse_disturbance_observations()}}
}
\description{
Summarise disturbance by season and cause for ODKC data
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_dead_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{filter_predation_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
