% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_minimal_tsc_turtledata.R
\name{download_minimal_tsc_turtledata}
\alias{download_minimal_tsc_turtledata}
\title{Download a minimal dataset of turtle observations from TSC}
\usage{
download_minimal_tsc_turtledata(source = "odk", year = NULL)
}
\arguments{
\item{source}{The TSC Encounter source, default: "odk".}

\item{year}{The earliest calendar year to be returned. Default: NULL.
This value, if not NULL, will be used to construct a query parameter
\code{when_year_gte=2019}}
}
\value{
A list of tibbles:
\itemize{
\item users A tibble of user names, roles, and contact details which can be
  used to resolve submitted user names to TSC user IDs
\item enc A tibble of encounters (source, source ID, QA status) which can be
  used to determine which encounters to create / update / skip.
\item surveys A tibble of surveys reconstructed from source and year.
}
}
\description{
Download a minimal dataset of turtle observations from TSC
}
