% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_odkc_turtledata_2020.R
\name{download_odkc_turtledata_2020}
\alias{download_odkc_turtledata_2020}
\title{Download all turtle data from DBCA's ODK Central 2020}
\usage{
download_odkc_turtledata_2020(
  local_dir = here::here("media"),
  prod = "https://odkc.dbca.wa.gov.au",
  tz = ruODK::get_default_tz(),
  download = TRUE,
  odkc_version = 1,
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{local_dir}{A local directory to which to save the attachment files.
Attachment filepaths will be relative to the directory.
The directory and its parent folders will be created if not existing.}

\item{prod}{The ODKC PROD server,
default: "https://odkc.dbca.wa.gov.au".}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{download}{(lgl) Whether to download attachments to \code{local_dir} or not,}

\item{odkc_version}{The numeric ODK Central version, e.g. 0.7 or 0.8.}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 class "odkc_turtledata" with items:
\itemize{
\item downloaded_on An UTC POSIXct timestamp of the data snapshot.
\item tracks The turtle tracks and nests from form
"Turtle Track or Nest 1.0".
\item tracks_dist Individual disturbances recorded against tracks,
one record per disturbance.
\item tracks_log Individual nest tags recorded against nests,
one record per tag.
\item tracks_egg Next excavation photos, one record per photo.
\item tracks_hatch Turtle hatchling morphometrics, one record per measured
hatchling.
\item tracks_fan_outlier Individual hatchling track outliers recorded
against hatched nests, one record per outlier.
\item tracks_light Individual light sources known at hatchling emergence,
one record per light source.
\item track_tally A line transect tally of turtle tracks from form
"Track Tally 0.6".
\item dist The disturbance and predation records from form
"Predator or Disturbance 1.1".
\item mwi Strandings and rescues from the form
"Marine Wildlife Incident 0.6 ".
\item mwi_dmg Individual injuries recorded against mwi,
one record per injury.
\item mwi_tag Individual tags sighted during an mwi, one record per tag.
\item tsi Turtle Sightings from form \code{Turtle-Sighting},
one record per sighted turtle.
\item tt Individual turtles encountered and tagged after nesting
during night time turtle tagging from form \code{Turtle-Tagging-3-0},
one record per encountered turtle.
\item tt_dmg Characteristic damages or injuries recorded during turtle
tagging, one record per damage.
\item tt_tag Individual tags recorded during turtle tagging,
one record per tag. Tags can be of different type,
re-sighted, applied, or removed.
\item tt_log Individual loggers recorded during turtle tagging,
one record per logger. Loggers can be of different type,
re-sighted, deployed, or removed.
\item svs Survey start points from form \code{Site-Visit-Start}.
\item sve Survey end points from form \code{Site-Visit-End}.
\item sites An sf object of known WAStD sites.
\item areas An sf object of known WAStD localities.
}
}
\description{
\lifecycle{maturing}
}
\seealso{
Other odkc: 
\code{\link{add_hatching_emergence_success_odkc}()},
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{hatching_emergence_success_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
