% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_w2_data.R
\name{download_w2_data}
\alias{download_w2_data}
\title{Download Turtle Tagging data from WAMTRAM2.}
\usage{
download_w2_data(
  ord = c("YmdHMS", "Ymd"),
  tz = "Australia/Perth",
  db_drv = Sys.getenv("W2_DRV"),
  db_srv = Sys.getenv("W2_SRV"),
  db_name = Sys.getenv("W2_DB"),
  db_user = Sys.getenv("W2_UN"),
  db_pass = Sys.getenv("W2_PW"),
  db_port = Sys.getenv("W2_PT"),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{ord}{Lubridate orders, default: `c("YmdHMS", "Ymd")`.}

\item{tz}{Timezone, default: `"Australia/Perth"`.}

\item{db_drv}{Database driver, default: `Sys.getenv("W2_DRV")` which should
resolve to e.g. `"/usr/lib/x86_64-linux-gnu/odbc/libtdsodbc.so"`.}

\item{db_srv}{Database server, default: `Sys.getenv("W2_SRV")`, which should
resolve to a valid server hostname, e.g. `"myserver.corporateict.domain"`.}

\item{db_name}{The database name, default: `Sys.getenv("W2_DB")`, which should
resolve to a valid database name, e.g. `turtle_tagging`.}

\item{db_user}{The read-permitted database user,
default: `Sys.getenv("W2_UN")`.}

\item{db_pass}{The database user's password, default: `Sys.getenv("W2_PW")`.}

\item{db_port}{The database port, default: `Sys.getenv("W2_PT")`, which
should resolve to a numeric port, e.g. `1234`.}

\item{verbose}{Whether to print verbose help messages,
default: `wastdr::get_wastdr_verbose()`.}
}
\value{
A named list of sanitised tables from the Turtle Tagging DB:

  # Metadata
  downloaded_on
  w2_tables

  # personnel
  persons

  # Data entry
  data_entry_batches
  data_entry
  data_entry_operators

  # Lookups
  lookup_beach_positions
  lookup_body_parts
  lookup_conditions
  lookup_damage_causes
  lookup_damage_codes
  lookup_datum_codes
  lookup_egg_count_methods
  lookup_id_types
  lookup_measurement_types
  lookup_pit_tag_states
  lookup_sample_tissue_type
  lookup_tag_states

  # Sites
  sites

  # Encounters
  enc
  enc_qa

  # Observations
  obs_flipper_tags
  obs_pit_tags
  obs_damages
  obs_measurements
  obs_samples

  # Reconstructed
  reconstructed_pit_tags
  reconstructed_tags
  reconstructed_turtles
}
\description{
This function requires an installed ODBC driver for MS SQL Server 2012.
The database credentials are handled via environment variables.
}
