% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wastd_sites.R
\name{download_wastd_sites}
\alias{download_wastd_sites}
\title{Download Areas of type Site from WAStD}
\usage{
download_wastd_sites(
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}
}
\value{
An `sf` dataframe of WAStD Sites joined to Locality details.
}
\description{
Download Areas of type Site from WAStD
}
