% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wastd_turtledata.R
\name{download_wastd_turtledata}
\alias{download_wastd_turtledata}
\title{Download all turtle data from WAStD}
\usage{
download_wastd_turtledata(max_records = NULL, verbose = get_wastdr_verbose())
}
\arguments{
\item{max_records}{<int> The max number of records to retrieve,
default: NULL (all).}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 class "wastd_data" with items:

  * downloaded_on An UTC POSIXct timestamp of the data snapshot generated by
    \code{\link{Sys.time()}}.
  * sites An sf object of known WAStD sites.
  * areas An sf object of known WAStD localities.
  * surveys A tibble of surveys, reconstructed by WAStD from Site Visit Start
    and End records (where given) and (TurtleNest)Encounters.
  * animals A tibble of AnimalEncounters, which includes live sightings,
    rescues, strandings, and other incidents, and turtle tagging.
  * tracks Turtle tracks and nests.
  * tracks_dist Individual disturbances recorded against tracks, one row per
    disturbance.
  * nest_dist Turtle Nest Disturbances
  * nest_tags Turtle Nest Tags
  * nest_excavations Turtle Nest Excavations - hatching and emergence success
  * hatchling_morph Hatchling morphometrics from next exacavations
  * nest_fans Hatchling emergence tracks - main fan
  * nest_fan_outliers Hatchling emergence tracks - individual outliers
  * nest_lightsources - Light sources present durign hatchling emergence
}
\description{
Download all turtle data from WAStD
}
