% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wastd_users.R
\name{download_wastd_users}
\alias{download_wastd_users}
\title{Download a minimal dataset of turtle observations from WAStD}
\usage{
download_wastd_users(
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A tibble of user names, roles, and contact details which can be
  used to resolve submitted user names to WAStD user IDs
}
\description{
Download a minimal dataset of turtle observations from WAStD
}
