% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{duplicate_surveys}
\alias{duplicate_surveys}
\title{List sites with more than one production survey on a given date}
\usage{
duplicate_surveys(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\description{
Sites are expected to be surveyed not more than once daily with the exception
of turtle tagging, where a morning survey and night tagging can happen
legitimately on the same calendar date.
}
\details{
QA operators will want to open each link from this list of potential duplicate
surveys, inspect their details, then decide on whether to make one the main
survey and close others as duplicates with the "make production" button.

\lifecycle{stable}
}
\seealso{
Other wastd_surveys: 
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()}
}
\concept{wastd_surveys}
