% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{exclude_training_species}
\alias{exclude_training_species}
\title{Exclude records with \code{species} "corolla-corolla".}
\usage{
exclude_training_species(value)
}
\arguments{
\item{value}{The output of \code{parse_disturbance_observations()}.}
}
\description{
Exclude records with \code{species} "corolla-corolla".
}
