% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{exclude_training_surveys}
\alias{exclude_training_surveys}
\title{Exclude training surveys.}
\usage{
exclude_training_surveys(value)
}
\arguments{
\item{value}{The output of \code{parse_surveys}}
}
\description{
Exclude training surveys.
}
