% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_wastd_turtledata.R
\name{export_wastd_turtledata}
\alias{export_wastd_turtledata}
\title{Export all WAStD turtledata to CSV and GeoJSON.}
\usage{
export_wastd_turtledata(x, outdir = here::here())
}
\arguments{
\item{x}{An object of class `wastd_data` as returned by
\code{\link{download_wastd_turtledata}}. This data can be filtered to
an area_name (WAStD Area of type Locality).}

\item{outdir}{The destination to write exported data to,
default: \code{here::here()}.}

\item{area_name}{The name of a WAStD Area of type Locality to filter the
data.}
}
\description{
Export all WAStD turtledata to CSV and GeoJSON.
}
\examples{
\dontrun{
data("wastd_data", package = "wastdr")
an <- "Ningaloo"
wastd_data \%>\%
filter_wastd_turtledata(area_name = an) \%>\%
export_wastd_turtledata(outdir = here::here("inst/export", urlize(an)))
}
}
