% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_possibly.R
\name{extract_possibly}
\alias{extract_possibly}
\title{Wrap magrittr::extract to return field value or -1}
\usage{
extract_possibly(...)
}
\arguments{
\item{...}{A list and a field name}
}
\description{
\lifecycle{stable}
}
\examples{
lol <- list(field1 = 1, field2 = 2, field3 = 3, field4 = 4, field5 = 5)
extract_possibly(lol, "field1")
extract_possibly(lol, "field8")
}
\seealso{
\url{https://rdrr.io/cran/purrr/man/safely.html}

Other helpers: 
\code{\link{absolute_angle}()},
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{leaflet_basemap}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{read_table}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{verify_odkc_turtledata}()},
\code{\link{verify_wamtram_data}()},
\code{\link{verify_wastd_data}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
