% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdate_as_tdate.R
\name{fdate_as_tdate}
\alias{fdate_as_tdate}
\title{Days since fiscal year start as true date}
\usage{
fdate_as_tdate(value)
}
\arguments{
\item{value}{A date}
}
\description{
Days since fiscal year start as true date
}
