% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mwi.R
\name{filter_alive}
\alias{filter_alive}
\title{Filter strandings to live rescues}
\usage{
filter_alive(data)
}
\arguments{
\item{data}{A dataframe with column "health", e.g. data from form
"Marine Wildlife Incident 0.6".}
}
\value{
The dataframe with rows matching live outcomes.
}
\description{
Filter strandings to live rescues
}
