% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mwi.R
\name{filter_dead}
\alias{filter_dead}
\title{Filter strandings to mortalities.}
\usage{
filter_dead(data)
}
\arguments{
\item{data}{A dataframe with column "health", e.g. data from form
"Marine Wildlife Incident 0.6".}
}
\value{
The dataframe with rows matching mortal outcomes.
}
\description{
Filter strandings to mortalities.
}
