% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{filter_disturbance}
\alias{filter_disturbance}
\title{Filter disturbance data to disturbances}
\usage{
filter_disturbance(data)
}
\arguments{
\item{data}{A dataframe with column "disturbance_cause".}
}
\value{
The dataframe with rows matching disturbance causes.
}
\description{
Filter disturbance data to disturbances
}
