% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_missing_survey}
\alias{filter_missing_survey}
\title{Filter records with missing \code{survey_id}}
\usage{
filter_missing_survey(value)
}
\arguments{
\item{value}{The output of \code{\link{parse_surveys}}}
}
\description{
Filter records with missing \code{survey_id}
}
