% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_odkc_turtledata.R
\name{filter_odkc_turtledata}
\alias{filter_odkc_turtledata}
\title{Filter odkc_data to a given area_name.}
\usage{
filter_odkc_turtledata(
  data,
  area_name = NULL,
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{<list> The output of data("odkc_data"), a list of tibbles and sf}

\item{area_name}{<chr> The name of the area to filter the data by. Options:
* NULL (default): do not filter data, return unfiltered data.
* "All turtle programs": do not filter data, return unfiltered data.
* "Other": Filter data to area_name NA.
* Any name in `unique(odkc_data$areas$area_name)`: return data filtered to
  this area_name.}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
<list> The input data, optionally filtered to a subset of records.
}
\description{
Filter odkc_data to a given area_name.
}
\examples{
data(odkc_data)

# All data
data_all <- filter_odkc_turtledata(odkc_data)
nrow(data_all$tracks)

# One area
area_names <- unique(odkc_data$areas$area_name)
area_names
data_area_1 <- filter_odkc_turtledata(odkc_data, area_name = "Cape Dommett")
nrow(data_area_1$tracks)
data_area_2 <- filter_odkc_turtledata(odkc_data, area_name = area_names[[2]])
nrow(data_area_2$tracks)


# Orphans
data_orphans <- filter_odkc_turtledata(odkc_data, area_name = "Other")
nrow(data_orphans)
}
