% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{filter_predation}
\alias{filter_predation}
\title{Filter disturbance data to predator presences}
\usage{
filter_predation(data)
}
\arguments{
\item{data}{A dataframe with column "disturbance_cause".}
}
\value{
The dataframe with rows matching predator presences.
}
\description{
Filter disturbance data to predator presences
}
