% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{filter_predation_odkc}
\alias{filter_predation_odkc}
\title{Filter disturbance data to predator presences for ODKC data}
\usage{
filter_predation_odkc(data)
}
\arguments{
\item{data}{A dataframe with column "disturbance_cause".}
}
\value{
The dataframe with rows matching predator presences.
}
\description{
\lifecycle{stable}
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_dead_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
