% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_surveys_missing_end}
\alias{filter_surveys_missing_end}
\title{Filter surveys with a missing \code{end_source_id}}
\usage{
filter_surveys_missing_end(value)
}
\arguments{
\item{value}{The output of \code{parse_surveys}}
}
\description{
Filter surveys with a missing \code{end_source_id}
}
