% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_surveys_requiring_qa}
\alias{filter_surveys_requiring_qa}
\title{Filter surveys with "NEEDS QA" in \code{start_comments} or
\code{end_comments}.}
\usage{
filter_surveys_requiring_qa(value)
}
\arguments{
\item{value}{The output of \code{parse_surveys}}
}
\description{
Filter surveys with "NEEDS QA" in \code{start_comments} or
\code{end_comments}.
}
