% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filter_wastd_season}
\alias{filter_wastd_season}
\title{Filter a dataframe of tracks, disturbance, incidents, or surveys to season}
\usage{
filter_wastd_season(data, season_start_year)
}
\arguments{
\item{data}{A dataframe of tracks, disturbance, incidents, or surveys
containing a column "season" (int) with the season start year, e.g. 2019.}

\item{season_start_year}{The desired season's start year, e.g. 2019.}
}
\value{
The dataframe filtered to rows from the desired season.
}
\description{
Filter a dataframe of tracks, disturbance, incidents, or surveys to season
}
