% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_wastd_turtledata.R
\name{filter_wastd_turtledata}
\alias{filter_wastd_turtledata}
\title{Filter WAStD turtledata to an area_name}
\usage{
filter_wastd_turtledata(x, area_name)
}
\arguments{
\item{x}{An object of class `wastd_data` as returned by
\code{\link{download_wastd_turtledata}}. This data can be filtered to
an area_name (WAStD Area of type Locality).}

\item{area_name}{The name of a WAStD Area of type Locality to filter the
data.}
}
\value{
An object of class `wastd_data` filtered to records within
  `area_name`.
}
\description{
Filter WAStD turtledata to an area_name
}
