% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_wastd_turtledata.R
\name{filter_wastd_turtledata_area}
\alias{filter_wastd_turtledata_area}
\title{Filter WAStD turtledata to an area_name}
\usage{
filter_wastd_turtledata_area(
  x,
  area_name = NULL,
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{x}{An object of class \code{wastd_data} as returned by
\code{\link{download_wastd_turtledata}}. This data can be filtered to
an area_name (WAStD Area of type Locality).}

\item{area_name}{(chr) The name of the area to filter the data by. Options:
\itemize{
\item NULL (default): do not filter data, return unfiltered data.
\item "All turtle programs": do not filter data, return unfiltered data.
\item "Other": Filter data to area_name NA.
\item Any name in \code{unique(wastd_data$areas$area_name)}: return data filtered to
this area_name.
}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An object of class \code{wastd_data} filtered to records within
\code{area_name} and the list of \code{seasons}.
}
\description{
Filter WAStD turtledata to an area_name
}
\examples{
\dontrun{
data(wastd_data)
first_area <- wastd_data$areas$area_name[[1]]
wastd_data_filtered <- wastd_data \%>\%
  filter_wastd_turtledata_arera(area_name = first_area)
wastd_data
wastd_data_filtered
unique(wastd_data_filtered$areas$area_name)
}
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_sites}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata_seasons}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{handle_http_status}()},
\code{\link{wastd_GET}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_bulk_post}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_create_update_skip}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
