% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_more.R
\name{get_more}
\alias{get_more}
\title{Decide whether the given download limit has been reached}
\usage{
get_more(total_count, limit)
}
\arguments{
\item{total_count}{<int> The total count of downloaded records}

\item{limit}{<int> or NULL The max count of records to download}
}
\value{
<lgl> TRUE if limit has not been reached or limit is NULL, else FALSE
}
\description{
Decide whether the given download limit has been reached
}
\examples{
testthat::expect_true(get_more(0L,NULL))
testthat::expect_true(get_more(1000L,NULL))
testthat::expect_true(get_more(9L,10))
testthat::expect_false(get_more(10L,10))
}
