% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_field.R
\name{get_num_field}
\alias{get_num_field}
\title{From a list of unnamed lists, or the data.frame equivalent, extract a field
or -1.}
\usage{
get_num_field(lol, field)
}
\arguments{
\item{lol}{(list) A list of lists}

\item{field}{(chr) The name of a field inside the list of lists}
}
\value{
The field value or -1
}
\description{
The given list of lists is unlisted,
  \code{\link{extract_possibly}} extracts the value of the field or -1,
  another call to unlist deduplicates, and finally the result is cast to
  numeric.
}
\details{
\lifecycle{stable}
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
