% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{ggplot_disturbance_by_season}
\alias{ggplot_disturbance_by_season}
\title{Plot disturbance by season}
\usage{
ggplot_disturbance_by_season(x)
}
\arguments{
\item{x}{An object of S3 class \code{wastd_data}, e.g. the output of
\code{wastdr::download_wastd_turtledata()}, optionally filtered to a locality
by \code{wastdr::filter_wastd_turtledata(area_name="Thevenard Island")}.}
}
\value{
A ggplot2 object
}
\description{
Plot disturbance by season
}
\examples{
data(wastd_data)
wastd_data \%>\% ggplot_disturbance_by_season()
}
