% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_fanangles.R
\name{ggplot_hatchling_misorientation}
\alias{ggplot_hatchling_misorientation}
\title{Return a ggplot violin plot of hatchling emergence misorientation}
\usage{
ggplot_hatchling_misorientation(x)
}
\arguments{
\item{x}{An object of S3 class \code{wastd_data}, e.g. the output of
\code{wastdr::download_wastd_turtledata()}, optionally filtered to a locality
by \code{wastdr::filter_wastd_turtledata(area_name="Thevenard Island")}.}
}
\value{
A ggplot figure
}
\description{
Facets: species
}
\examples{
data(wastd_data)
wastd_data \%>\%
  filter_wastd_turtledata(area_name = "Delambre Island") \%>\%
  ggplot_hatchling_misorientation()

wastd_data \%>\%
  filter_wastd_turtledata(area_name = "Delambre Island") \%>\%
  ggplot_hatchling_misorientation() \%>\%
  plotly::ggplotly()
}
