% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{ggplot_nesting_success_per_area_season_species_pct}
\alias{ggplot_nesting_success_per_area_season_species_pct}
\title{Return a ggplot of nesting success}
\usage{
ggplot_nesting_success_per_area_season_species_pct(data)
}
\arguments{
\item{data}{The output of \code{nesting_success_per_area_season_species}, a
summary of \code{wastd_data}.}
}
\value{
A ggplot figure
}
\description{
Facets: species
}
\examples{
data(wastd_data)
wastd_data \%>\%
  filter_wastd_turtledata(area_name = "Delambre Island") \%>\%
  nesting_success_per_area_season_species() \%>\%
  ggplot_nesting_success_per_area_season_species_pct()

wastd_data \%>\%
  filter_wastd_turtledata(area_name = "Delambre Island") \%>\%
  nesting_success_per_area_season_species() \%>\%
  ggplot_nesting_success_per_area_season_species_pct() \%>\%
  plotly::ggplotly()
}
