% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{ggplot_sighting_status_per_area_season_species}
\alias{ggplot_sighting_status_per_area_season_species}
\title{Return a stacked ggplot barchart of processed emergences by recapture status}
\usage{
ggplot_sighting_status_per_area_season_species(data)
}
\arguments{
\item{data}{The output of \code{total_emergences_per_area_season_species}, a
summary of \code{wastd_data}.}
}
\value{
A ggplot2 figure
}
\description{
Facets: species
}
\examples{
data(wastd_data)
wastd_data \%>\%
  sighting_status_per_area_season_species() \%>\%
  ggplot_sighting_status_per_area_season_species()

wastd_data \%>\%
  sighting_status_per_area_season_species() \%>\%
  ggplot_sighting_status_per_area_season_species() \%>\%
  plotly::ggplotly()
}
