% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{ggplot_track_successrate_by_date}
\alias{ggplot_track_successrate_by_date}
\title{Plot the track success rate (relative numbers) of a given species as time
series.}
\usage{
ggplot_track_successrate_by_date(
  data,
  speciesname,
  placename = "",
  prefix = ""
)
}
\arguments{
\item{data}{The output of \code{parse_turtle_nest_encounters()}}

\item{speciesname}{The species name, e.g. "natator-depressus"}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}
}
\description{
Plot the track success rate (relative numbers) of a given species as time
series.
}
