% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs_getFeature.R
\name{gs_getFeature}
\alias{gs_getFeature}
\title{Return a GeoJSON Featurecollection from an OGC WFS 2.0.0 endpoint}
\usage{
gs_getFeature(
  layer_name = "public:herbie_hbvsupra_public",
  url = "https://kmi.dbca.wa.gov.au/geoserver/dpaw/ows"
)
}
\arguments{
\item{layer_name}{The GeoServer layer name, default: "public:herbie_hbvsupra_public"}

\item{url}{The GeoServer URL, default: "https://kmi.dbca.wa.gov.au/geoserver/dpaw/ows"}
}
\value{
The httr::content of a GET request to the WFS GetFeature endpoint
}
\description{
Defaults to DBCA's KMI GeoServer offering public WACensus views
}
