% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_http_status.R
\name{handle_http_status}
\alias{handle_http_status}
\title{Emit a specific warning messages depending on HTTP status}
\usage{
handle_http_status(response)
}
\arguments{
\item{response}{An HTTP response.}
}
\description{
Status 401: Instruct how to authenticate.
Status http_error: warn and show status.
}
