% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{hatching_emergence_success}
\alias{hatching_emergence_success}
\title{Sumarizes HS and ES for tracks of type `hatched-nest`.}
\usage{
hatching_emergence_success(value)
}
\arguments{
\item{value}{The output of \code{parse_turtle_nest_encounters()}}
}
\description{
Sumarizes HS and ES for tracks of type `hatched-nest`.
}
