% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{hatching_emergence_success_area}
\alias{hatching_emergence_success_area}
\title{Sumarizes HS and ES for tracks of type `hatched-nest`, grouped by `area_name`.}
\usage{
hatching_emergence_success_area(value)
}
\arguments{
\item{value}{The output of \code{parse_turtle_nest_encounters()}}
}
\description{
Sumarizes HS and ES for tracks of type `hatched-nest`, grouped by `area_name`.
}
