% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks_odkc.R
\name{hatching_emergence_success_odkc}
\alias{hatching_emergence_success_odkc}
\title{Summarize HS and ES for Nest excavations}
\usage{
hatching_emergence_success_odkc(data)
}
\arguments{
\item{data}{The output of \code{wastd_data$nest_excavations}}
}
\description{
\lifecycle{stable}
}
\examples{
data("odkc_data")
odkc_data$tracks_egg \%>\%
  wastdr::sf_as_tbl() \%>\%
  wastdr::add_hatching_emergence_success_odkc() \%>\%
  wastdr::hatching_emergence_success_odkc()
}
\seealso{
Other odkc: 
\code{\link{add_hatching_emergence_success_odkc}()},
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
