% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{hatching_emergence_success_site}
\alias{hatching_emergence_success_site}
\title{Sumarizes HS and ES for tracks of type `hatched-nest`
grouped by `site_name`.}
\usage{
hatching_emergence_success_site(value)
}
\arguments{
\item{value}{The output of \code{parse_turtle_nest_encounters()}}
}
\description{
Sumarizes HS and ES for tracks of type `hatched-nest`
grouped by `site_name`.
}
