% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpdate_as_gmt08.R
\name{httpdate_as_gmt08}
\alias{httpdate_as_gmt08}
\title{Return a UTC HTTP date string as POSIXct object in GMT+08}
\usage{
httpdate_as_gmt08(datestring)
}
\arguments{
\item{datestring}{(character) A UTC HTTP date as character string,
e.g. '2016-11-20T21:46:57.163000Z' with time zone included}
}
\value{
A POSIXct date object in GMT+08, e.g. '2016-11-21 05:46:57 AWST'
}
\description{
Return a UTC HTTP date string as POSIXct object in GMT+08
}
\examples{
httpdate_as_gmt08("2016-11-20T21:46:57.163000Z")
}
