% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize}
\alias{humanize}
\title{Convert between url-safe dash-separated-names and Human Readable Title Case}
\usage{
humanize(urlsafe_string)
}
\arguments{
\item{urlsafe_string}{(chr) A url-safe-string}
}
\value{
a Human Readable Title Cased String
}
\description{
Convert between url-safe dash-separated-names and Human Readable Title Case
}
\examples{
humanize("chelonia-mydas")
humanize("natator-depressus")
humanize("successful-crawl")
}
