% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_tsc_sites.R
\name{join_tsc_sites}
\alias{join_tsc_sites}
\title{Join a dataframe with lat/lon to sf objects of sites and areas.}
\usage{
join_tsc_sites(data, sites, prefix = "observed_at_")
}
\arguments{
\item{data}{A dataframe with an optionally prefixed latitude and longitude}

\item{sites}{An sf object of TSC sites}

\item{prefix}{An optional prefix to the longitude/latitude field names,
default: "observed_at_".}
}
\value{
The dataframe as sf with site/area joined.
}
\description{
Join a dataframe with lat/lon to sf objects of sites and areas.
}
