% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet_basemap.R
\name{leaflet_basemap}
\alias{leaflet_basemap}
\title{Create a Leaflet basemap for Western Australia}
\usage{
leaflet_basemap(l_width = NULL, l_height = NULL)
}
\arguments{
\item{l_width}{The leaflet map width, default: NULL}

\item{l_height}{The leaflet map height, default: NULL}
}
\value{
A Leaflet map focused on WA
}
\description{
Create a Leaflet basemap for Western Australia
}
\examples{
leaflet_basemap()
}
