% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{list_survey_count}
\alias{list_survey_count}
\title{Create a table of survey counts from the output of \code{parse_surveys}}
\usage{
list_survey_count(surveys, placename = "")
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{placename}{(string) The place name, used in labels.}
}
\description{
\lifecycle{stable}
}
\seealso{
Other wastd_surveys: 
\code{\link{duplicate_surveys}()},
\code{\link{list_survey_effort}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()}
}
\concept{wastd_surveys}
