% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qa_user_mapping.R
\name{make_search_links}
\alias{make_search_links}
\title{Turn a string of names into individual Markdown search links to WAStD Users.}
\usage{
make_search_links(s)
}
\arguments{
\item{s}{A string, e.g. "Florian Mayer" or "Flo and Joe".}
}
\value{
A single string of whitespace separated Markdown links to search
  WAStD Users.
}
\description{
Turn a string of names into individual Markdown search links to WAStD Users.
}
