% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dist_odkc.R
\name{map_dist_odkc}
\alias{map_dist_odkc}
\title{Map Disturbance observations interactively from ODK Central data.}
\usage{
map_dist_odkc(
  dist,
  tracks = NULL,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  tz = "Australia/Perth",
  cluster = FALSE
)
}
\arguments{
\item{dist}{The output of \code{parse_disturbance_observations()}.}

\item{tracks}{The output of `turtleviewer::turtledata$tracks_dist` -
Turtle nest disturbance obs joined to turtle nests.}

\item{sites}{An sf object of sites with `site_name` and polygon geom, e.g.
`turtleviewer::turtledata$sites`.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}
}
\value{
A leaflet map
}
\description{
\lifecycle{maturing}
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle
each disturbance class separately. The maps auto-zooms to the extent of data
given.
}
\examples{
\dontrun{
data("odkc_data")
map_dist_odkc(
  odkc_data$dist,
  tracks = odkc_data$tracks_dist, sites = odkc_data$sites
)
map_dist_odkc(NULL, tracks = odkc_data$tracks_dist, sites = odkc_data$sites)
map_dist_odkc(odkc_data$dist, tracks = NULL, sites = odkc_data$sites)
map_dist_odkc(odkc_data$dist, tracks = odkc_data$tracks_dist, sites = NULL)
}
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
