% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_mwi.R
\name{map_mwi}
\alias{map_mwi}
\title{Map Marine Wildlife Incident 0.6}
\usage{
map_mwi(
  data,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  tz = "Australia/Perth",
  cluster = FALSE
)
}
\arguments{
\item{data}{AnimalEncounters from WAStD.}

\item{sites}{An sf object of sites with `site_name` and polygon geom, e.g.
`turtleviewer::turtledata$sites`.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}
}
\value{
A leaflet map
}
\description{
Map Marine Wildlife Incident 0.6
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle
each species separately. The maps auto-zooms to the extent of data given.
}
