% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nests.R
\name{map_nests}
\alias{map_nests}
\title{Map tagged Turtle Nests interactively.}
\usage{
map_nests(data, fmt = "\%d/\%m/\%Y \%H:\%M")
}
\arguments{
\item{data}{The output of \code{parse_nesttag_observations}.}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}
}
\value{
A leaflet map
}
\description{
Map tagged Turtle Nests interactively.
}
\details{
Creates a Leaflet map with an interactive legend.
The maps auto-zooms to the extent of data given.
}
