% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_sitevisit.R
\name{map_sv_odkc}
\alias{map_sv_odkc}
\title{Map Site Visit Start and End from ODK Central data over TSC Sites.}
\usage{
map_sv_odkc(
  svs,
  sve,
  sites,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  tz = "Australia/Perth",
  cluster = FALSE
)
}
\arguments{
\item{svs}{The output of `turtleviewer::turtledata$svs` - data from ODK form
`Site Visit Start 0.3`.}

\item{sve}{The output of `turtleviewer::turtledata$sve` - data from ODK form
`Site Visit End 0.2`.}

\item{sites}{An sf object of sites with `site_name` and polygon geom, e.g.
`turtleviewer::turtledata$sites`.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}
}
\value{
A leaflet map
}
\description{
Map Site Visit Start and End from ODK Central data over TSC Sites.
}
\details{
Creates a Leaflet map of site visit start and end points.
The maps auto-zooms to the extent of data given.
}
