% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tracks.R
\name{map_tracks}
\alias{map_tracks}
\title{Map TurtleNestEncounters interactively.}
\usage{
map_tracks(tracks)
}
\arguments{
\item{tracks}{The output of \code{parse_turtle_nest_encounters} and \code{add_nest_labels}.}
}
\value{
A leaflet map
}
\description{
Map TurtleNestEncounters interactively.
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle each species
separately. The maps auto-zooms to the extent of data given.
}
