% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tt_odkc.R
\name{map_tt_odkc}
\alias{map_tt_odkc}
\title{Map Turtle Tagging 3.0 parsed with ruODK from ODK Central.}
\usage{
map_tt_odkc(
  data,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  tz = "Australia/Perth",
  cluster = FALSE
)
}
\arguments{
\item{data}{The output of \code{odkc_data$tt} from ODK form \verb{Turtle Tagging 3.0}.}

\item{sites}{An sf object of sites with \code{site_name} and polygon geom, e.g.
\code{turtleviewer::turtledata$sites}.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}
}
\value{
A leaflet map
}
\description{
\lifecycle{maturing}
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle
each species separately. The maps auto-zooms to the extent of data given.

This map function uses data from ODK Central / ruODK using form
"Turtle Tagging 3.0".
}
\examples{
\dontrun{
data("odkc_data")
map_tt_odkc(odkc_data$tt, sites = odkc_data$sites)
map_tt_odkc(data = NULL, sites = odkc_data$sites)
}
}
\seealso{
Other odkc: 
\code{\link{add_hatching_emergence_success_odkc}()},
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{hatching_emergence_success_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
