% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wamtram.R
\name{map_wamtram}
\alias{map_wamtram}
\title{Map WAMTRAM data}
\usage{
map_wamtram(data, location = NULL, place = NULL)
}
\arguments{
\item{data}{The output of \code{wastdr::download_w2_data()}.}

\item{location}{A W2 location code, e.g. "TH", to filter the data by.}

\item{place}{A W2 place code, e.g. "THEE", to filter the data by. Providing
\code{place} overrides \code{location}.}
}
\value{
A leaflet map showing W2 sites and encounters.
}
\description{
Map WAMTRAM data
}
\examples{
\dontrun{
data("w2_data", package="etlTurtleNesting")
map_wamtram(w2_data)

map_wamtram(w2_data, location="DH")
map_wamtram(w2_data, place="DHTB")
}
}
