% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wastd_wamtram_sites.R
\name{map_wastd_wamtram_sites}
\alias{map_wastd_wamtram_sites}
\title{Map WAStD and WAMTRAM sites}
\usage{
map_wastd_wamtram_sites(wastd_areas, wastd_sites, wamtram_sites)
}
\arguments{
\item{wastd_areas}{Areas from \code{wastd_data}.}

\item{wastd_sites}{Sites from \code{wastd_data}.}

\item{wamtram_sites}{Sites from \code{wamtram_data}.}
}
\value{
A leaflet map
}
\description{
This map is useful to compare and update WAStD with and from WAMTRAM sites.
}
\examples{
\dontrun{
library(wastdr)

# From canned data
data("wastd_data")
data("wamtram_data") # canned data for package
w2_data = readRDS(here::here("inst/w2.rds")) # full snapshot for dev
map_wastd_wamtram_sites(wastd_data$areas, wastd_data$sites, wamtram_data$sites)


# From live data
areas_sf <- wastd_GET("area") \%>\% magrittr::extract2("data") \%>\%
  geojsonio::as.json() \%>\% geojsonsf::geojson_sf()

areas <- areas_sf \%>\%
  dplyr::filter(area_type == "Locality") \%>\%
  dplyr::transmute(area_id = pk, area_name = name, w2_location_code = w2_location_code)

sites <- areas_sf \%>\%
  dplyr::filter(area_type == "Site") \%>\%
  dplyr::transmute(site_id = pk, site_name = name, w2_place_code = w2_place_code) \%>\%
  sf::st_join(areas)

# split rows by place code, map only what place codes are not in wastd yet
map_wastd_wamtram_sites(areas, sites, w2_data$sites)
}
}
