% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_bearing.R
\name{mean_bearing}
\alias{mean_bearing}
\title{Calculate the mean bearing of two bearings (from, to)}
\usage{
mean_bearing(from, to)
}
\arguments{
\item{from}{The start bearing, counted clockwise from North}

\item{to}{The end bearing, counted clockwise from North}
}
\value{
The mean of the two bearings as decimal degrees within \verb{[0..360[}
}
\description{
Calculate the mean bearing of two bearings (from, to)
}
\examples{
testthat::expect_equal(mean_bearing(NA, 10), NA)
testthat::expect_equal(mean_bearing(10, NA), NA)
testthat::expect_equal(mean_bearing(0, 20), 10)
testthat::expect_equal(mean_bearing(20, 60), 40)
testthat::expect_equal(mean_bearing(60, 0), 210)
}
\seealso{
Other helpers: 
\code{\link{absolute_angle}()},
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
