% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mis_bearing.R
\name{mis_bearing}
\alias{mis_bearing}
\title{Return the smallest possible angle for a sector's mean bearing and a given direction}
\usage{
mis_bearing(from, to, water)
}
\arguments{
\item{from}{The smallest bearing of the sector, measured clockwise from North}

\item{to}{The largest bearing of the sector, measured clockwise from North}

\item{water}{The bearing of a direction}
}
\description{
\code{dplyr::rowwise} is required to apply this function in \code{dplyr::mutate}.
}
\examples{
testthat::expect_equal(mis_bearing(10, 50, 40), c(30, 40))
testthat::expect_equal(mis_bearing(10, 50, 340), c(340, 30))
testthat::expect_equal(mis_bearing(30, 50, 230), c(230, 40))
testthat::expect_equal(mis_bearing(30, 50, 210), c(40, 210))
}
\seealso{
Other helpers: 
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
