% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nest_json}
\alias{nest_json}
\title{300 nest encounters}
\format{A \code{wastd_api_response} with 300 encounters
\itemize{
  \item TODO
}}
\usage{
nest_json
}
\description{
A \code{wastd_api_response} with 300 nest encounters from
\code{\link{get_wastd("turtle-nest-encounters")}}
}
\keyword{datasets}
