% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{nesting_success_per_area_season_species}
\alias{nesting_success_per_area_season_species}
\title{Calculate nesting success for emergences per area, season, species}
\usage{
nesting_success_per_area_season_species(x)
}
\arguments{
\item{x}{An object of S3 class \code{wastd_data}, e.g. the output of
\code{wastdr::download_wastd_turtledata()}, optionally filtered to a locality
by \code{wastdr::filter_wastd_turtledata(area_name="Thevenard Island")}.}
}
\value{
A tibble with the summary data
}
\description{
Break up total emergences into tagged and non_tagged animals
}
\examples{
data(wastd_data)
wastd_data \%>\%
  nesting_success_per_area_season_species() \%>\%
  View()
}
