% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{nesting_type_by_season_species}
\alias{nesting_type_by_season_species}
\title{Pivot table of nesting type by season and species}
\usage{
nesting_type_by_season_species(value)
}
\arguments{
\item{value}{The output of \code{parse_turtle_nest_encounters()}}
}
\description{
Pivot table of nesting type by season and species
}
