% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks_odkc.R
\name{nesting_type_by_site_season_age_species_odkc}
\alias{nesting_type_by_site_season_age_species_odkc}
\title{Pivot table of nesting type by site, season, track age, and species for ODKC data}
\usage{
nesting_type_by_site_season_age_species_odkc(value)
}
\arguments{
\item{value}{Submissions to ODK Central from form Track or Nest 1.0,
retrieved and parsed with \code{ruODK::odata_submission_get()}.}
}
\description{
Pivot table of nesting type by site, season, track age, and species for ODKC data
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_dead_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{filter_predation_odkc}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
