% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nests}
\alias{nests}
\title{Example turtle nest encounters (only nests, excluding tracks)}
\format{A \code{tbl_df} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item datetime <dttm>
  \item longitude <dbl>
  \item latitude <dbl>
  \item crs <chr>
  \item location_accuracy <dbl>
  \item date <date>
  \item species <chr>
  \item nest_age <chr>
  \item nest_type <chr>
  \item name <chr>
  \item habitat <chr>
  \item disturbance <chr>
  \item absolute_admin_url <chr>
  \item obs <list>
  \item photos <list>
  \item hatching_success <dbl>
  \item emergence_success <dbl>
  \item clutch_size <dbl>
  \item source <chr>
  \item source_id <chr>
  \item encounter_type <chr>
  \item status <chr>
  \item observer <chr>
  \item reporter <chr>
}}
\source{
https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=10&format=json&nest_type=hatched-nest

https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=100&format=json
}
\usage{
nests

nests
}
\description{
A parsed \code{wastd_api_response} with 10 turtle nest encounters from
\code{get_wastd("turtle-nest-encounters")}

A parsed \code{wastd_api_response} with 100 turtle nest encounters from
\code{get_wastd("turtle-nest-encounters")}
}
\examples{
data("nests")
head(nests)
head(nests)
}
\keyword{datasets}
