% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_as_tsc.R
\name{odkc_as_tsc}
\alias{odkc_as_tsc}
\title{Top level helper combining all mapped data ODKC to TSC}
\usage{
odkc_as_tsc(odkc_data, user_mapping)
}
\arguments{
\item{odkc_data}{.}

\item{user_mapping}{.}
}
\description{
Top level helper combining all mapped data ODKC to TSC
}
