% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_as_wastd.R
\name{odkc_as_wastd}
\alias{odkc_as_wastd}
\title{Top level helper combining all mapped data ODKC to WAStD}
\usage{
odkc_as_wastd(odkc_data, user_mapping)
}
\arguments{
\item{odkc_data}{.}

\item{user_mapping}{.}
}
\description{
Top level helper combining all mapped data ODKC to WAStD
}
