% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_dist_as_distenc.R
\name{odkc_dist_as_distenc}
\alias{odkc_dist_as_distenc}
\title{Transform odkc_data$dist into WAStD Encounters.}
\usage{
odkc_dist_as_distenc(data, user_mapping)
}
\arguments{
\item{data}{A tibble of parsed ODKC Disturbance encounters,
e.g. \code{odkc_data$dist}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("encounters")}
}
\description{
Related TurtleNestDisturbanceObservations are generated with
\code{\link{odkc_dist_as_tndo}}.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping <- NULL # see odkc_plan for actual user mapping

odkc_data$tracks \%>\%
  odkc_dist_as_distenc(user_mapping) \%>\%
  head(1) \%>\%
  jsonlite::toJSON()
}
}
