% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_dist_as_tndo.R
\name{odkc_dist_as_tndo}
\alias{odkc_dist_as_tndo}
\title{Transform odkc_data$dist into WAStD TurtleNestDisturbanceObservations.}
\usage{
odkc_dist_as_tndo(data)
}
\arguments{
\item{data}{A tibble of dist,  e.g. \code{odkc_data$dist}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-disturbance-observations")}
}
\description{
The corresponding Encounters are created through
\code{\link{odkc_dist_as_distenc}}.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_dist_as_tndo(odkc_ex$tracks_dist)
x \%>\% wastd_POST("turtle-nest-disturbance-observations",
  api_url = au, api_token = at
)
}
}
