% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_mwi_as_wastd_turtlemorph.R
\name{odkc_mwi_as_wastd_turtlemorph}
\alias{odkc_mwi_as_wastd_turtlemorph}
\title{Transform odkc_data$mwi into WAStD TurtleMorphometricObservations.}
\usage{
odkc_mwi_as_wastd_turtlemorph(data, user_mapping)
}
\arguments{
\item{data}{The tibble of Marine Wildlife Incidents,
e.g. \code{odkc_data$mwi}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-morphometrics")}
}
\description{
Transform odkc_data$mwi into WAStD TurtleMorphometricObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping = NULL # see odkc_plan for actual user mapping
tmorph <- odkc_mwi_as_wastd_turtlemorph(odkc_ex$mwi, user_mapping)

tmorph \%>\%
  wastd_POST("turtle-morphometrics",
    api_url = Sys.getenv("WASTDR_API_DEV_URL"),
    api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
}
}
