% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_mwi_dmg_as_wastd_turtledmg.R
\name{odkc_mwi_dmg_as_wastd_turtledmg}
\alias{odkc_mwi_dmg_as_wastd_turtledmg}
\title{Transform odkc_data$mwi_dmg into WAStD TurtleDamageObservations.}
\usage{
odkc_mwi_dmg_as_wastd_turtledmg(data)
}
\arguments{
\item{data}{A tibble of turtle damage obs, e.g. \code{odkc_data$mwi_dmg}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-damage-observations")}
}
\description{
Transform odkc_data$mwi_dmg into WAStD TurtleDamageObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_as_wastd_nesttagobs(odkc_ex$mwi_dmg)
x \%>\% wastd_POST("turtle-damage-observations", api_url = au, api_token = at)
}
}
