% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_mwi_tag_as_wastd_tagobs.R
\name{odkc_mwi_tag_as_wastd_tagobs}
\alias{odkc_mwi_tag_as_wastd_tagobs}
\title{Transform odkc_data$mwi_tag into WAStD TagObservations.}
\usage{
odkc_mwi_tag_as_wastd_tagobs(data, user_mapping)
}
\arguments{
\item{data}{A tibble of tags, e.g. \code{odkc_data$mwi_tag}.}

\item{user_mapping}{The output of \code{\link{make_user_mapping}()},
containing at least columns \code{pk} (WAStD User PK) and
\code{odkc_username} (as written in odkc data field \code{reporter}).}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("tag-observations")}
}
\description{
Transform odkc_data$mwi_tag into WAStD TagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
user_mapping <- NULL # see odkc_plan for actual user mapping
x <- odkc_mwi_tag_as_wastd_tagobs(odkc_ex$mwi_tag, user_mapping)
x \%>\% wastd_POST("tag-observations", api_url = au, api_token = at)
}
}
