% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_svs_sve_as_wastd_surveys.R
\name{odkc_svs_sve_as_wastd_surveys}
\alias{odkc_svs_sve_as_wastd_surveys}
\title{Reconstruct WAStD Surveys from ODKC Site Visit Start/End}
\usage{
odkc_svs_sve_as_wastd_surveys(svs, sve, user_mapping)
}
\arguments{
\item{svs}{ODKC Site Visit Start, e.g. `data("odkc_data"); odkc_data$svs`}

\item{sve}{ODKC Site Visit End, e.g. `data("odkc_data"); odkc_data$sve`}
}
\value{

}
\description{
Reconstruct WAStD Surveys from ODKC Site Visit Start/End
}
\details{
Surveys are reconstructed from Site Visit Start records.
Survey end details are populated from Site Visit End records, where matching
by `device_id`, `calendar_date_awst`, and `site_id`.

Sources of error:
* SVS forgotten to record
* SVE forgotten to record
* SVS recorded outside known site boundaries (WAStD Area),
  `site_id` mismatch, svs$site_id empty
* SVE recorded outside known site boundaries (WAStD Area),
  `site_id` mismatch, sve$site_id empty
* Changed devices mid survey (`device_id` mismatch, likely reporter mismatch,
  see end_comments)

Things to provide as QA reports outside this function:
* Stray SVE without matching SVS
* SVS joined to SVE only by device_id and calendar_date_awst minus "good"
  surveys (location out of bounds)
* SVS joined to SVE only by calendar_date_awst and site_id minus "good"
  surveys (device swapped?)
}
\examples{
wastdr_msg_info("examples go here")
}
