% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_as_wastd_nestobs.R
\name{odkc_tracks_as_wastd_nestobs}
\alias{odkc_tracks_as_wastd_nestobs}
\title{Transform odkc_data$tracks into WAStD TurtleNestObservations (egg counts).}
\usage{
odkc_tracks_as_wastd_nestobs(data)
}
\arguments{
\item{data}{A tibble of tracks,  e.g. \code{odkc_data$tracks}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-excavations")}
}
\description{
Transform odkc_data$tracks into WAStD TurtleNestObservations (egg counts).
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_as_wastd_nestobs(odkc_ex$tracks)
x \%>\% wastd_POST("turtle-nest-excavations", api_url = au, api_token = at)
}
}
