% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_as_wastd_nesttagobs.R
\name{odkc_tracks_as_wastd_nesttagobs}
\alias{odkc_tracks_as_wastd_nesttagobs}
\title{Transform odkc_data$tracks into WAStD NestTagObservations.}
\usage{
odkc_tracks_as_wastd_nesttagobs(data, user_mapping)
}
\arguments{
\item{data}{A tibble of tracks,  e.g. \code{odkc_data$tracks}.}

\item{user_mapping}{The output of \code{\link{make_user_mapping}()},
containing at least columns \code{pk} (WAStD User PK) and
\code{odkc_username} (as written in odkc data field \code{reporter}).}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("nest-tag-observations")}
}
\description{
Transform odkc_data$tracks into WAStD NestTagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
user_mapping <- NULL # see odkc_plan for actual user mapping
x <- odkc_tracks_as_wastd_nesttagobs(odkc_ex$tracks, user_mapping)
x \%>\% wastd_POST("nest-tag-observations", api_url = au, api_token = at)
}
}
