% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_as_wastd_tne.R
\name{odkc_tracks_as_wastd_tne}
\alias{odkc_tracks_as_wastd_tne}
\title{Transform odkc_data$tracks into WAStD TurtleNestEncounter}
\usage{
odkc_tracks_as_wastd_tne(data)
}
\arguments{
\item{data}{The tibble of parsed WAStD TurtleNestEncounters,
e.g. \code{odkc_data$tracks}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_bulk_post}("turtle-nest-encounters")}
}
\description{
Transform odkc_data$tracks into WAStD TurtleNestEncounter
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne() \%>\%
  head(1) \%>\%
  jsonlite::toJSON()

odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne() \%>\%
  wastd_bulk_post("turtle-nest-encounters",
    api_url = "http://localhost:8220/api/1/"
  )
}
}
