% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_dist_as_wastd_tndistobs.R
\name{odkc_tracks_dist_as_wastd_tndistobs}
\alias{odkc_tracks_dist_as_wastd_tndistobs}
\title{Transform odkc_data$tracks_dist into WAStD TurtleNestDisturbanceObservations.}
\usage{
odkc_tracks_dist_as_wastd_tndistobs(data)
}
\arguments{
\item{data}{A tibble of tracks_dist,  e.g. \code{odkc_data$tracks_dist}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-disturbance-observations")}
}
\description{
Transform odkc_data$tracks_dist into WAStD TurtleNestDisturbanceObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_dist_as_wastd_tndistobs(odkc_ex$tracks_dist)
x \%>\% wastd_POST("turtle-nest-disturbance-observations",
  api_url = au, api_token = at)
}
}
