% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_fan_outlier_as_tnheo.R
\name{odkc_tracks_fan_outlier_as_tnheo}
\alias{odkc_tracks_fan_outlier_as_tnheo}
\title{Transform odkc_data$tracks into WAStD TNHE Outliers.}
\usage{
odkc_tracks_fan_outlier_as_tnheo(data)
}
\arguments{
\item{data}{A tibble of tracks,  e.g. \code{odkc_data$tracks_fan_outlier}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-hatchling-emergence-outliers")}
}
\description{
Transform odkc_data$tracks into WAStD TNHE Outliers.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_fan_outlier_as_tnheo(odkc_ex$tracks_fan_outlier)
x \%>\% wastd_POST("turtle-nest-hatchling-emergence-outliers",
                 api_url = au, api_token = at)
}
}
