% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_hatch_as_wastd_thmorph.R
\name{odkc_tracks_hatch_as_wastd_thmorph}
\alias{odkc_tracks_hatch_as_wastd_thmorph}
\title{Transform odkc_data$tracks into WAStD NestTagObservations.}
\usage{
odkc_tracks_hatch_as_wastd_thmorph(data)
}
\arguments{
\item{data}{A tibble of tracks,  e.g. \code{odkc_data$tracks_hatch}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-hatchling-morphometrics")}
}
\description{
Transform odkc_data$tracks into WAStD NestTagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_as_wastd_nesttagobs(odkc_ex$tracks_hatch)
x \%>\%
  wastd_POST("turtle-hatchling-morphometrics", api_url = au, api_token = at)
}
}
