% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_light_as_wastd_tnhels.R
\name{odkc_tracks_light_as_wastd_tnhels}
\alias{odkc_tracks_light_as_wastd_tnhels}
\title{Transform odkc_data$tracks into WAStD TurtleNestHatchlingEmergenceLightSources}
\usage{
odkc_tracks_light_as_wastd_tnhels(data)
}
\arguments{
\item{data}{A tibble of nest_lightsources,  e.g. \code{odkc_data$tracks_light}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-hatchling-emergence-light-sources")}
}
\description{
Transform odkc_data$tracks into WAStD TurtleNestHatchlingEmergenceLightSources
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tracks_light_as_wastd_tnhels(odkc_ex$tracks_light)
x \%>\% wastd_POST("turtle-nest-hatchling-emergence-light-sources",
  api_url = au, api_token = at
)
}
}
