% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_log_as_loggerenc.R
\name{odkc_tracks_log_as_loggerenc}
\alias{odkc_tracks_log_as_loggerenc}
\title{Transform odkc_data$tracks into WAStD TurtleNestEncounters.}
\usage{
odkc_tracks_log_as_loggerenc(data, user_mapping)
}
\arguments{
\item{data}{A tibble of parsed ODKC Logger encounters,
e.g. \code{odkc_data$tracks_log}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("logger-encounters")}
}
\description{
Transform odkc_data$tracks into WAStD TurtleNestEncounters.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping <- NULL # see odkc_plan for actual user mapping

odkc_data$tracks \%>\%
  odkc_tracks_log_as_loggerenc(user_mapping) \%>\%
  head(1) \%>\%
  jsonlite::toJSON()
}
}
