% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tsi_as_wastd_ae.R
\name{odkc_tsi_as_wastd_ae}
\alias{odkc_tsi_as_wastd_ae}
\title{Transform odkc_data$tsi into WAStD AnimalEncounters.}
\usage{
odkc_tsi_as_wastd_ae(data, user_mapping)
}
\arguments{
\item{data}{The tibble of Turtle Sightings,
e.g. \code{odkc_data$tsi}.}

\item{user_mapping}{The output of \code{\link{make_user_mapping}()},
containing at least columns \code{pk} (WAStD User PK) and
\code{odkc_username} (as written in odkc data field \code{reporter}).}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("animal-encounters")}
}
\description{
Transform odkc_data$tsi into WAStD AnimalEncounters.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping <- NULL # see odkc_plan for actual user mapping
ae <- odkc_tsi_as_wastd_ae(odkc_ex$tsi, user_mapping)

ae \%>\%
  wastd_POST("animal-encounters",
    api_url = Sys.getenv("WASTDR_API_DEV_URL"),
    api_token = Sys.getenv("WASTDR_API_DEV_TOKEN")
  )
}
}
