% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_lte.R
\name{odkc_tt_as_wastd_lte}
\alias{odkc_tt_as_wastd_lte}
\title{Transform odkc_data$track_tally into WAStD LineTransectEncounters.}
\usage{
odkc_tt_as_wastd_lte(data, user_mapping)
}
\arguments{
\item{data}{A tibble of parsed ODKC Track Tallies,
e.g. \code{odkc_data$track_tally}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("line-transect-encounters")}
}
\description{
Transform odkc_data$track_tally into WAStD LineTransectEncounters.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping <- NULL # see odkc_plan for actual user mapping

odkc_data$track_tally \%>\%
  odkc_tt_as_wastd_lte(user_mapping) \%>\%
  head(1) \%>\%
  jsonlite::toJSON()
}
}
