% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_tndto.R
\name{odkc_tt_as_wastd_tndto}
\alias{odkc_tt_as_wastd_tndto}
\title{Transform odkc_data$track_tally_dist into WAStD
TurtleNestDisturbanceTallyObservation.}
\usage{
odkc_tt_as_wastd_tndto(data)
}
\arguments{
\item{data}{A tibble of track_tally_dist,
e.g. \code{odkc_data$track_tally_dist}.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-disturbance-tally")}
}
\description{
Transform odkc_data$track_tally_dist into WAStD
TurtleNestDisturbanceTallyObservation.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_data$track_tally_dist \%>\%
  odkc_tt_as_wastd_tndto \%>\%
  wastd_POST("turtle-nest-disturbance-tally", api_url = au, api_token = at)
}
}
