% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_animal_encounters.R
\name{parse_animal_encounters}
\alias{parse_animal_encounters}
\title{Parse a \code{wastd_api_response} of \code{animal-encounters} to tbl_df}
\usage{
parse_animal_encounters(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{animal-encounters}, e.g. \code{get_wastd("animal-encounters")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item survey_id <int>
  \item survey_start_time <dttm>
  \item survey_end_time <dttm>
  \item survey_start_comments <chr>
  \item survey_end_comments <chr>
  \item datetime <dttm>
  \item turtle_date <date>
  \item season <int>
  \item week <int>
  \item longitude <chr>
  \item latitude <chr>
  \item crs <chr>
  \item location_accuracy <dbl>
  \item name <chr>
  \item species <chr>
  \item health <chr>
  \item sex <chr>
  \item maturity <chr>
  \item behaviour <chr>
  \item habitat <chr>
  \item activity <chr>
  \item nesting_event <chr>
  \item checked_for_injuries <chr>
  \item scanned_for_pit_tags <chr>
  \item checked_for_flipper_tags <chr>
  \item cause_of_death <chr>
  \item cause_of_death_confidence <chr>
  \item absolute_admin_url <chr>
  \item obs <list>
  \item source <chr>
  \item source_id <chr>
  \item encounter_type <chr>
  \item status <chr>
}
}
\description{
Parse a \code{wastd_api_response} of \code{animal-encounters} to tbl_df
}
