% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_area_sf.R
\name{parse_area_sf}
\alias{parse_area_sf}
\title{Parse a \code{wastd_api_response} of \code{area} to tbl_df and sf}
\usage{
parse_area_sf(wastd_api_response, wastd_url = wastdr::get_wastd_url())
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{area}, e.g. \code{get_wastd("area")}}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}
}
\value{
A \code{tbl_df} and \code{sf} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item northern_extent <num>
  \item length_surveyed_m <int>
  \item length_survey_roundtrip_m <int>
}
}
\description{
Parse a \code{wastd_api_response} of \code{area} to tbl_df and sf
}
