% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_community-conservationlisting.R
\name{parse_community_conservationlisting}
\alias{parse_community_conservationlisting}
\title{Parse a \code{wastd_api_response} of \code{community-conservationlisting} to tbl_df}
\usage{
parse_community_conservationlisting(
  wastd_api_response,
  wastd_url = wastdr::get_wastd_url()
)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{taxon-conservationlisting},
e.g. \code{get_wastd("community-conservationlisting")}.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item id <int> The TSC ID of the Conservation Listing
  \item community <int> The TSC Community code.
  \item source <int> The TSC ID for the Source of the Conservation Listing.
  \item source_id <chr> The original ID of the source as chr. Most sources
        will be an integer, uuid, or any other unique ID.
  \item scope <int> The TSC ID for the legal scope of this Conservation
        Listing: State of Western Australia (0),
        Commonwealth of Austalia (1), International (2), or Action Plan (3).
  \item status <int> The TSC ID of the approval status, e.g. listed (80),
        de-listed (90), rejected (100), or in stages of approval (0..70).
  \item proposed_on <dttm> The date on which this Conservation Listing was
        proposed on (GMT+08).
  \item effective_from <dttm> The date on which this Conservation Listing was
        effective from (GMT+08).
  \item effective_to <dttm> The date on which this Conservation Listing was
        effective_to (GMT+08).
  \item last_reviewed_on <dttm> The date on which this Conservation Listing was
        last reviewed on (GMT+08).
  \item review_due <dttm> The date on which this Conservation Listing will be
        due for reviewed (GMT+08).
  \item comments <chr> Any comments. Worth a read!
  \item category_cache <chr> A human-readable representation of the
        conservation category (or categories).
  \item criteria_cache <chr> A human-readable representation of the
        conservation criteria.
  \item label <chr> An auto-generated label for the Conservation Listing.
  \item category <lst> A list if TSC conservation category IDs.
  \item criteria <lst> A list if TSC conservation criterion IDs.
}
}
\description{
Parse a \code{wastd_api_response} of \code{community-conservationlisting} to tbl_df
}
