% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_disturbance_observations.R
\name{parse_disturbance_observations}
\alias{parse_disturbance_observations}
\title{Parse a \code{wastd_api_response} of \code{disturbance-observations} to tbl_df}
\usage{
parse_disturbance_observations(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{disturbance-observations}, e.g. \code{wastd_GET("disturbance-observations")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item survey_id <int>
  \item survey_start_time <dttm>
  \item survey_end_time <dttm>
  \item survey_start_comments <chr>
  \item survey_end_comments <chr>
  \item encounter_id <chr>
  \item datetime <dttm>
  \item longitude <dbl>
  \item latitude <dbl>
  \item crs <chr>
  \item location_accuracy <dbl>
  \item turtle_date <date>
  \item season <int>
  \item comments <chr>
  \item absolute_admin_url <chr>
  \item photos <list>
  \item source <chr>
  \item source_id <chr>
  \item encounter_type <chr>
  \item status <chr>
  \item observer <chr>
  \item reporter <chr>
  \item disturbance_cause <chr>
  \item disturbance_cause_confidence <chr>
  \item disturbance_severity <chr>
  \item disturbance_comments <chr>
}
}
\description{
Parse a \code{wastd_api_response} of \code{disturbance-observations} to tbl_df
}
