% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fanangles.R
\name{parse_encounterobservations}
\alias{parse_encounterobservations}
\title{Parse a \code{wastd_api_response} of Encounter observations to tbl_df}
\usage{
parse_encounterobservations(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{turtle-nest-*}, e.g.
\code{wastd_GET("turtle-nest-encounters")}}
}
\value{
A \code{tbl_df} with non-repetitive columns of:
\itemize{
  \item Encounter
  \item Survey
  \item Area
  \item Site
  \item Observation
}
}
\description{
WAStD provides all subclasses of `wastd.observations.observation` together
with their encounter (survey, site, area, reporter, observer) via its API
as GeoJSON FeatureCollections.
}
\details{
This function acts as a generic parser for

* "turtle-nest-hatchling-emergences"
* "turtle-nest-hatchling-emergence-outliers"
* "turtle-nest-hatchling-emergence-light-sources"
* "turtle-nest-excavations"
* "turtle-nest-tag-observations"
* "turtle-nest-disturbance-observations"
}
