% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nesttag_observations.R
\name{parse_nesttag_observations}
\alias{parse_nesttag_observations}
\title{Parse a \code{wastd_api_response} of \code{nesttag_observations} to tbl_df}
\usage{
parse_nesttag_observations(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{nesttag_observations}, e.g. \code{wastd_GET("nesttag_observations")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item survey_id <int>
  \item survey_start_time <dttm>
  \item survey_end_time <dttm>
  \item survey_start_comments <chr>
  \item survey_end_comments <chr>
  \item datetime <dttm>
  \item turtle_date <date>
  \item season <int>
  \item season_week <int> Number of completed weeks since fiscal year start
  \item iso_week <int> Number of completed weeks since calendar year start
  \item longitude <dbl>
  \item latitude <dbl>
  \item crs <chr>
  \item location_accuracy <dbl>
  \item source <chr>
  \item source_id <chr>
  \item encounter_id <chr>
  \item absolute_admin_url <chr>
  \item observer <chr>
  \item reporter <chr>
  \item status <chr>
  \item photos <list>
  \item encounter_comments <chr>
  \item tag_status <chr>
  \item flipper_tag_id <chr>
  \item date_nest_laid <chr>
  \item tag_label <chr>
}
}
\description{
Parse a \code{wastd_api_response} of \code{nesttag_observations} to tbl_df
}
