% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveys.R
\name{parse_surveys}
\alias{parse_surveys}
\title{Parse a \code{wastd_api_response} of \code{animal-encounters} to tbl_df}
\usage{
parse_surveys(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{animal-encounters}, e.g. \code{get_wastd("animal-encounters")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item start_time <dttm>
  \item end_time <dttm>
  \item source <chr>
  \item source_id <chr>
  \item status <chr>
}
}
\description{
Parse a \code{wastd_api_response} of \code{animal-encounters} to tbl_df
}
