% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{plot_survey_count}
\alias{plot_survey_count}
\title{Plot the surveyed hours from the output of \code{parse_surveys}}
\usage{
plot_survey_count(surveys, placename = "", prefix = "")
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{placename}{(string) The place name, used in labels. Default: ""}

\item{prefix}{(string) A prefix for the filename. Default: ""}
}
\description{
Plot the surveyed hours from the output of \code{parse_surveys}
}
