% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{plotly_nesting_success_per_area_season_species}
\alias{plotly_nesting_success_per_area_season_species}
\title{Return a stacked plotly barchart of emergences}
\usage{
plotly_nesting_success_per_area_season_species(data)
}
\arguments{
\item{data}{The output of \code{ggplot_nesting_success_per_area_season_species},
a visualisation of a summary of \code{wastd_data}.}
}
\value{
A plotly figure
}
\description{
Return a stacked plotly barchart of emergences
}
\examples{
data(wastd_data)
wastd_data \%>\%
  nesting_success_per_area_season_species() \%>\%
  ggplot_nesting_success_per_area_season_species() \%>\%
  plotly_nesting_success_per_area_season_species()
}
