% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_odkc_turtledata_2019.R
\name{print.odkc_turtledata}
\alias{print.odkc_turtledata}
\title{S3 print method for 'odkc_turtledata'.}
\usage{
\method{print}{odkc_turtledata}(x, ...)
}
\arguments{
\item{x}{An object of class `wastd_data` as returned by
\code{\link{download_odkc_turtledata_2019}}.}

\item{...}{Extra parameters for `print`}
}
\description{
Prints a short representation of data returned by
\code{\link{download_odkc_turtledata_2019}}.
}
