% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_w2_data.R
\name{print.wamtram_data}
\alias{print.wamtram_data}
\title{S3 print method for 'wamtram_data'.}
\usage{
\method{print}{wamtram_data}(x, ...)
}
\arguments{
\item{x}{An object of class \code{wamtram_data} as returned by
\code{\link{download_w2_data}}.}

\item{...}{Extra parameters for \code{print}}
}
\description{
Prints a short representation of data returned by
\code{\link{download_w2_data}}.
}
\seealso{
Other included: 
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
