% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_api.R
\name{print.wastd_api}
\alias{print.wastd_api}
\title{S3 print method for 'wastd_api'.}
\usage{
\method{print}{wastd_api}(x, ...)
}
\arguments{
\item{x}{An object of class `wastd_api` as returned by `wastd_api`.}

\item{...}{Extra parameters for `print`}
}
\description{
Prints a short representation of data returned by `wastd_api`.
}
