% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{sf_as_tbl}
\alias{sf_as_tbl}
\title{Convert a SimpleFeatures object to a non-spatial dataframe.}
\usage{
sf_as_tbl(sf_obj)
}
\arguments{
\item{sf_obj}{A SimpleFeatures object}
}
\value{
The SimpleFeatures object with geometry set to NULL.
}
\description{
Convert a SimpleFeatures object to a non-spatial dataframe.
}
