% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wastd_api.R
\name{skip_test_if_odkc_offline}
\alias{skip_test_if_odkc_offline}
\title{Check whether API token is set and WAStD is online, else skip test}
\usage{
skip_test_if_odkc_offline()
}
\description{
Check whether API token is set and WAStD is online, else skip test
}
