% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_count_heatmap}
\alias{survey_count_heatmap}
\title{Plot the survey count from the output of \code{parse_surveys} as heatmap.}
\usage{
survey_count_heatmap(surveys, placename = "", prefix = "")
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}
}
\description{
Plot the survey count from the output of \code{parse_surveys} as heatmap.
}
