% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_ground_covered}
\alias{survey_ground_covered}
\title{Return the number of surveys for a given site_id and season}
\usage{
survey_ground_covered(surveys, sid, kms, seas)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{sid}{The ID of a site, e.g. 22}

\item{kms}{The numbers of kilometers covered in a survey, e.g. 1.6}

\item{seas}{The season as integer, e.g. 2018}
}
\description{
\lifecycle{stable}
}
\examples{
data("wastd_data")
one_season <- unique(wastd_data$surveys$season)[1]
one_site_id <- unique(wastd_data$surveys$site_id)[1]
some_kms <- 10
survey_ground_covered(wastd_data$surveys, one_site_id, some_kms, one_season)
}
\seealso{
Other wastd_surveys: 
\code{\link{duplicate_surveys}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()}
}
\concept{wastd_surveys}
