% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveys.R
\name{survey_hours_per_person}
\alias{survey_hours_per_person}
\title{Sum the hours surveyed per person by season from the output of \code{parse_surveys}.}
\usage{
survey_hours_per_person(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\value{
A tibble with columns reporter, season, hours_surveyed, sorted by most to fewest hours.
}
\description{
Sum the hours surveyed per person by season from the output of \code{parse_surveys}.
}
