% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_season_stats}
\alias{survey_season_stats}
\title{Generate a season summary from the output of \code{parse_surveys}}
\usage{
survey_season_stats(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\description{
\lifecycle{stable}
}
\details{
Surveys, gruped by \code{season},
summarised as first and last day of surveys, season length in days,
number and total hours of surveys.
}
\seealso{
Other wastd_surveys: 
\code{\link{duplicate_surveys}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()}
}
\concept{wastd_surveys}
