% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_show_detail}
\alias{survey_show_detail}
\title{Select main survey attributes.}
\usage{
survey_show_detail(value)
}
\arguments{
\item{value}{The output of \code{parse_surveys()}}
}
\description{
Select main survey attributes.
}
