% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveys.R
\name{surveys_per_site_name_and_date}
\alias{surveys_per_site_name_and_date}
\title{Count number of surveys per site_name and date from the output of \code{parse_surveys}.}
\usage{
surveys_per_site_name_and_date(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\value{
A tibble with columns date, site_name, n (number of surveys)
}
\description{
Count number of surveys per site_name and date from the output of \code{parse_surveys}.
}
