% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tag_json}
\alias{tag_json}
\title{300 animal encounters}
\format{A \code{wastd_api_response} with 300 encounters
\itemize{
  \item TODO
}}
\usage{
tag_json
}
\description{
A \code{wastd_api_response} with 300 animal encounters from
\code{\link{get_wastd("animal-encounters")}}
}
\keyword{datasets}
