% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tags}
\alias{tags}
\title{300 animal encounters (taggings, strandings and others)}
\format{A \code{tbl_df} with columns:
\itemize{
  \item datetime (dttm)
  \item longitude (dbl)
  \item latitude (dbl)
  \item crs (chr)
  \item location_accuracy (dbl)
  \item date (date)
  \item species (chr)
  \item health (chr)
  \item sex (chr)
  \item behaviour (chr)
  \item habitat (chr)
  \item activity (chr)
  \item nesting_event (chr)
  \item checked_for_injuries (chr)
  \item scanned_for_pit_tags (chr)
  \item checked_for_flipper_tags (chr)
  \item cause_of_death (chr)
  \item cause_of_death_confidence (chr)
  \item absolute_admin_url (chr)
  \item obs (list)
  \item source (chr)
  \item source_id (chr)
  \item encounter_type (chr)
  \item status (chr)
}}
\source{
https://strandings.dpaw.wa.gov.au/api/1/animal-encounters/?taxon=Cheloniidae&limit=10&format=json
}
\usage{
tags
}
\description{
A parsed \code{wastd_api_response} with 300 animal encounters from
\code{get_wastd("animal-encounters")}
}
\examples{
  head(tags)
}
\keyword{datasets}
