% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdate_as_fdate.R
\name{tdate_as_fdate}
\alias{tdate_as_fdate}
\title{True date as days since fiscal year start}
\usage{
tdate_as_fdate(value)
}
\arguments{
\item{value}{A date}
}
\description{
True date as days since fiscal year start
}
