% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tne}
\alias{tne}
\title{TurtleNestEncounter (tracks and nests) WAStD API response}
\format{An object of class \code{wastd_api_response} of length 5.}
\source{
https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=100&format=json
}
\usage{
tne
}
\description{
This API response is parsed into wastr's data "nests".
}
\examples{
# Prove that turtle_nest_encounters_hatched parses to nests
library(dplyr)
data(tne)
data(nests)
parsed_nests <- parse_turtle_nest_encounters(tne)
}
\keyword{datasets}
