% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{track_json}
\alias{track_json}
\title{300 track / nest encounters}
\format{A \code{wastd_api_response} with 300 encounters
\itemize{
  \item TODO
}}
\usage{
track_json
}
\description{
A \code{wastd_api_response} with 300 turtle nest encounters from
\code{\link{get_wastd("turtle-nest-encounters")}}
}
\keyword{datasets}
