% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{track_success}
\alias{track_success}
\title{Calculate nesting success as number of tracks with nests vs other tracks}
\usage{
track_success(tracks)
}
\arguments{
\item{tracks}{The output of \code{parse_turtle_nest_encounters} and \code{add_nest_labels}.}
}
\description{
Calculate nesting success as number of tracks with nests vs other tracks
}
