% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{track_success_by_species}
\alias{track_success_by_species}
\title{The nesting success grouped by season and species}
\usage{
track_success_by_species(track_success)
}
\arguments{
\item{track_success}{The output of \code{\link{track_success}}}
}
\description{
The nesting success grouped by season and species
}
