% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{10 turtle nest encounters (tracks and nests)}
\format{A \code{tbl_df} with columns:
\itemize{
  \item datetime (dttm)
  \item longitude (dbl)
  \item latitude (dbl)
  \item crs (chr)
  \item location_accuracy (dbl)
  \item date (date)
  \item species (chr)
  \item nest_age (chr)
  \item nest_type (chr)
  \item habitat (chr)
  \item disturbance <chr>
  \item absolute_admin_url <chr>
  \item obs <list>
  \item hatching_success <dbl>
  \item emergence_success <dbl>
  \item clutch_size <dbl>
  \item source <chr>
  \item source_id <chr>
  \item encounter_type <chr>
  \item status <chr>
}}
\source{
https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=300&format=json
}
\usage{
tracks
}
\description{
A parsed \code{wastd_api_response} with 300 turtle nest encounters from
\code{get_wastd("turtle-nest-encounters")}
}
\examples{
  head(tracks)
}
\keyword{datasets}
