% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_as_geojson.R
\name{tracks_as_geojson}
\alias{tracks_as_geojson}
\title{Convert tracks to a GeoJSON string}
\usage{
tracks_as_geojson(tracks)
}
\description{
Turns calendar_date_awst (required) into start and end (start + 1 day) dates,
converts dataframe to GeoJSON string.
}
