% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_ts.R
\name{tracks_ts}
\alias{tracks_ts}
\title{Plot of nesting_type_by_season_day_species over time}
\usage{
tracks_ts(data, surveys, placename = "", prefix = "")
}
\arguments{
\item{data}{The output of \code{parse_turtle_nest_encounters}.}

\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}
}
\value{
A ggplot2 object. Saves the plot to .png.
}
\description{
Plot of nesting_type_by_season_day_species over time
}
