% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tsc_data}
\alias{tsc_data}
\title{TSC data, parsed, first ten records each}
\format{
An object of class \code{list} of length 14.
}
\usage{
tsc_data
}
\description{
TSC data, parsed, first ten records each
}
\seealso{
Other included: 
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wamtram_data}()},
\code{\link{print.wastd_data}()},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
