% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tun.R
\name{tun}
\alias{tun}
\title{A wrapper for \code{tidyr::unnest_wider}}
\usage{
tun(data, col)
}
\arguments{
\item{data}{The data frame to be unnested.}

\item{col}{The column name to be unnested as string.}
}
\value{
The data frame \code{data} with column \code{col} replaced by its unnested
components.
}
\description{
This function saves repetition by currying tidyr::unnest_wider with default
parameters \verb{names_repair = "universal", names_sep = "_"}.
}
