% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_odkc_to_wastd.R
\name{upload_odkc_to_wastd}
\alias{upload_odkc_to_wastd}
\title{Top level helper to upload all Turtle Nesting Census data to WAStD}
\usage{
upload_odkc_to_wastd(
  data,
  update_existing = FALSE,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{ODKC data transformed into WAStD format and split into create,
update, skip.}

\item{update_existing}{Whether to update existing but unchanged records in
WAStD Default: FALSE.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A list of results from `wastd_create_update_skip`.
}
\description{
Encounters and their related observations are uploaded to WAStD:
}
\details{
* New encounters will be created
* Existing but unchanged (status=new) encounters will be updated
  if `update_existing=TRUE`, else skipped.
* Existing and value-added encounters (status > new) will be skipped.s
}
