% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_geojson.R
\name{upsert_geojson}
\alias{upsert_geojson}
\title{Upsert GeoJSON into WAStD API endpoints}
\usage{
upsert_geojson(
  gj_featurecollection,
  serializer = "names",
  chunksize = 1000,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = get_wastdr_verbose()
)
}
\arguments{
\item{gj_featurecollection}{(list) A GeoJSON featurecollection as list}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{chunksize}{(int) The number of features to upload simultaneously,
default: 1000.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\description{
Upsert GeoJSON into WAStD API endpoints
}
\concept{wacensus}
