% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urlize.R
\name{urlize}
\alias{urlize}
\title{Convert between url-safe dash-separated-names and Human Readable Title Case}
\usage{
urlize(human_string)
}
\arguments{
\item{human_string}{(chr) A Human Readable Title Cased String}
}
\value{
A url-safe-string
}
\description{
Convert between url-safe dash-separated-names and Human Readable Title Case
}
\examples{
urlize("file name 1")
urlize("Natator depressus")
}
