% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_wastd_data.R
\name{verify_wastd_data}
\alias{verify_wastd_data}
\title{Abort if the parameter x is not of class \code{wastd_data}}
\usage{
verify_wastd_data(x)
}
\arguments{
\item{x}{An object of S3 class \code{wastd_data}, e.g. the output of
\code{wastdr::download_wastd_turtledata()}, optionally filtered to a locality
by \code{wastdr::filter_wastd_turtledata(area_name="Thevenard Island")}.}
}
\description{
Abort if the parameter x is not of class \code{wastd_data}
}
\examples{
\dontrun{
data(wastd_data)
verify_wastd_data(wastd_data)
testthat::expect_error(verify_wastd_data(NULL))
testthat::expect_error(verify_wastd_data(c(1,2,3)))
}
}
