% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_ae}
\alias{wastd_ae}
\title{WAStD AnimalEncounters, parsed, test data captured by author}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 43 columns.}
\usage{
wastd_ae
}
\description{
WAStD AnimalEncounters, parsed, test data captured by author
}
\keyword{datasets}
