% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_api.R
\name{wastd_api}
\alias{wastd_api}
\title{Return GeoJSON features from a WAStD API endpoint as data.table or list}
\usage{
wastd_api(serializer, base_url = "https://strandings.dpaw.wa.gov.au/api/1/",
  query = list(taxon = "Cheloniidae", limit = 10000, format = "json"),
  wastd_api_token = Sys.getenv("WASTD_APITOKEN"), simplify = TRUE)
}
\arguments{
\item{serializer}{(character) WAStD API serializer name (required)}

\item{base_url}{(character) Name of the new resource
(default: "https://strandings.dpaw.wa.gov.au/api/1/")}

\item{query}{(list) API query parameters for format, limit, filtering
(default: list(taxon="Cheloniidae", limit=10000, format="json"))}

\item{wastd_api_token}{(character) The WAStD API token
(default: Sys.getenv("WASTD_APITOKEN"))}

\item{simplify}{(Boolean) Whether to flatten nested data frames into a single
data frame with repeating unnamed groups as lists (default: TRUE), or as
list of lists (simplify=FALSE)}
}
\value{
An S3 object of class "wastd_api" containing:
  content: The retrieved GeoJSON features as data.table or list
  serializer: The called serializer, e.g. "animal-encounters"
  response: The API HTTP response with all metadata
}
\description{
Return GeoJSON features from a WAStD API endpoint as data.table or list
}
\details{
Call the WAStD API serializer's list view with given GET parameters,
parse the response as text into a GeoJSON FeatureCollection.
Parse the FeatureCollection using jsonlite::fromJSON and return its features
as nested data.frame (simplify=TRUE) or as list of lists (simplify=FALSE).
TODO: use pagination, see
https://cran.r-project.org/web/packages/jsonlite/vignettes/json-paging.html
}
\examples{
\dontrun{
track_records <- wastd_api("turtle-nest-encounters")

tag_records <- wastd_api("animal-encounters")

nest_json <- wastd_api("turtle-nest-encounters",
                       query=list(
                         nest_type="hatched-nest",
                         limit=10000,
                         format="json"),
                       simplify=FALSE)
}
}
