% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_area_raw}
\alias{wastd_area_raw}
\title{WAStD Areas, unparsed}
\format{An object of class \code{wastd_api_response} of length 5.}
\usage{
wastd_area_raw
}
\description{
WAStD Areas, unparsed
}
\keyword{datasets}
