% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_bulk_post.R
\name{wastd_bulk_post}
\alias{wastd_bulk_post}
\title{Post a list of records to a WAStD or TSC API endpoint}
\usage{
wastd_bulk_post(
  data,
  serializer,
  encode = "json",
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = FALSE
)
}
\arguments{
\item{data}{A tibble of a data with columns equal to the serializer's
fields.}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{encode}{The parameter \code{encode} for \code{link{httr::POST}},
default: "json".
Other options: \code{c("multipart", "form", "json", "raw")}.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
The list of \code{wastd_api_response}s from \code{\link{wastd_POST}}
}
\description{
\lifecycle{maturing}
}
\examples{
\dontrun{
odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne() \%>\%
  wastd_bulk_post("turtle-nest-encounters")
}
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_sites}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{handle_http_status}()},
\code{\link{wastd_GET}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_create_update_skip}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
