% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_create_update_skip.R
\name{wastd_create_update_skip}
\alias{wastd_create_update_skip}
\title{Create, update, skip records to a WAStD serializer.}
\usage{
wastd_create_update_skip(
  data_create,
  data_update,
  data_skip,
  update_existing = FALSE,
  label = "records",
  serializer = "encounters",
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data_create}{Records to create.}

\item{data_update}{Records to update.}

\item{data_skip}{Records to skip.}

\item{update_existing}{Whether to update existing but unchanged records in
WAStD. Default: FALSE.}

\item{label}{The model label for verbose messages, default: `records`.}

\item{serializer}{The WAStD API serializer, default: `encounters`.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A list of three objects:

* `created` The WAStD API response for all created records
* `updated` The WAStD API response for all updated records
* `skipped` The original records for all skipped records
}
\description{
Handles empty data, verbose notifications.
}
