% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_data}
\alias{wastd_data}
\title{Turtle Nesting Census data from WAStD
Names are hidden, first 10 records of encounters, all sites.}
\format{An object of class \code{wastd_data} of length 13.}
\usage{
wastd_data
}
\description{
Turtle Nesting Census data from WAStD
Names are hidden, first 10 records of encounters, all sites.
}
\keyword{datasets}
