% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_parse.R
\name{wastd_parse}
\alias{wastd_parse}
\title{Parse a \code{wastd_api_response} into a tibble}
\usage{
wastd_parse(wastd_api_response, payload = "features")
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} from WAStD}

\item{payload}{(chr) The name of the key containing the parsed
\code{httr::content()} from the WAStD API call}
}
\value{
A tibble with one row per record and columns corresponding to each
  record's fields.
}
\description{
From a \code{wastd_api_response}, turn the key \code{payload}
  (default: "features") into a \code{tibble:tibble}, and
  \code{tidyr::unnest_wider} the tibble into columns equivalent to the fields
  of the WAStD API serializer.
  If GeoJSON is found, the keys `geometry` (including `coordinates`) and
  `properties` are unnested.
}
