% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_tne}
\alias{wastd_tne}
\title{WAStD TurtleNestEncounters, parsed, test data captured by author}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10 rows and 43 columns.}
\usage{
wastd_tne
}
\description{
WAStD TurtleNestEncounters, parsed, test data captured by author
}
\keyword{datasets}
