% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_tne_raw}
\alias{wastd_tne_raw}
\title{WAStD TurtleNestEncounters, unparsed, test data captured by author}
\format{
An object of class \code{wastd_api_response} of length 5.
}
\usage{
wastd_tne_raw
}
\description{
WAStD TurtleNestEncounters, unparsed, test data captured by author
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
