% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr.R
\docType{package}
\name{wastdr-package}
\alias{wastdr}
\alias{wastdr-package}
\title{wastdr: WA Sea Turtle Database 'WAStD' API Wrapper}
\description{
\code{\link{wastdr}} is an R Client for the ODK Central API.
The 'WA Sea Turtle Database' 'WAStD' is a 'Django' web app with a 'RESTful'
'API' returning 'GeoJSON' 'FeatureCollections'.
\code{\link{wastdr}} provides helpers to retrieve data from the 'API',
and to turn the received 'GeoJSON' into R 'tibbles'.
\code{\link{wastdr}} will grow alongside implemented
use cases of retrieving and analysing 'WAStD' data.

  Please see the `wastdr` website for full documentation:
  * <https://dbca-wa.github.io/wastdr/>

  `ruODK` is "pipe-friendly" and re-exports `%>%`, but does not
  require their use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbca-wa/wastdr}
  \item Report bugs at \url{https://github.com/dbca-wa/wastdr/issues}
}

}
\author{
\strong{Maintainer}: Florian W. Mayer \email{Florian.Mayer@dbca.wa.gov.au}

}
\keyword{internal}
