% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_msg.R
\name{wastdr_msg_abort}
\alias{wastdr_msg_abort}
\title{rlang::abort() with a red error message with a cross symbol.}
\usage{
wastdr_msg_abort(message)
}
\arguments{
\item{message}{<chr> A message to print}
}
\description{
rlang::abort() with a red error message with a cross symbol.
}
\examples{
\dontrun{
wastdr_msg_abort("This is an error, abort.")
}
}
\seealso{
Other helpers: 
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()}
}
\concept{helpers}
