% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_msg.R
\name{wastdr_msg_info}
\alias{wastdr_msg_info}
\title{Print a blue info message with an info symbol.}
\usage{
wastdr_msg_info(message)
}
\arguments{
\item{message}{<chr> A message to print}
}
\description{
\lifecycle{stable}
}
\examples{
wastdr_msg_info("This is an info message.")
}
\seealso{
Other helpers: 
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()}
}
\concept{helpers}
