% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_msg.R
\name{wastdr_msg_success}
\alias{wastdr_msg_success}
\title{Print a green success message with a tick symbol.}
\usage{
wastdr_msg_success(message)
}
\arguments{
\item{message}{<chr> A message to print}
}
\description{
Print a green success message with a tick symbol.
}
\examples{
wastdr_msg_success("This is a success message.")
}
\seealso{
Other helpers: 
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_warn}()}
}
\concept{helpers}
