% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_msg.R
\name{wastdr_msg_warn}
\alias{wastdr_msg_warn}
\title{rlang::warn() with a yellow warning message with a warning symbol.}
\usage{
wastdr_msg_warn(message)
}
\arguments{
\item{message}{<chr> A message to print}
}
\description{
\lifecycle{stable}
}
\examples{
\dontrun{
wastdr_msg_warn("This is a warning.")
}
}
\seealso{
Other helpers: 
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()}
}
\concept{helpers}
