% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_settings.R
\name{wastdr_settings}
\alias{wastdr_settings}
\alias{get_wastd_url}
\alias{get_wastdr_api_url}
\alias{get_wastdr_api_token}
\alias{get_wastdr_api_un}
\alias{get_wastdr_api_pw}
\alias{get_wastdr_verbose}
\title{Get or set wastdr WAStD settings}
\usage{
wastdr_settings()

get_wastd_url()

get_wastdr_api_url()

get_wastdr_api_token()

get_wastdr_api_un()

get_wastdr_api_pw()

get_wastdr_verbose()
}
\value{
\code{wastdr_settings} prints your base url and WAStD API key.
\code{ckanr_setup} sets your production and test settings, while
}
\description{
Get or set wastdr WAStD settings
}
\seealso{
\code{\link{wastdr_setup}},
\code{\link{get_wastd_url}},
\code{\link{get_wastdr_api_url}},
\code{\link{get_wastdr_api_token}},
\code{\link{get_wastdr_api_un}}, and
\code{\link{get_wastdr_api_pw}}.
}
\concept{wastr settings}
