% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_settings.R
\name{wastdr_setup}
\alias{wastdr_setup}
\title{Configure default WAStD settings}
\usage{
wastdr_setup(api_url = "https://strandings.dpaw.wa.gov.au/api/1/",
  api_token = NULL)
}
\arguments{
\item{api_url}{A WAStD API URL (optional),
default: "https://strandings.dpaw.wa.gov.au/api/1/"}

\item{api_token}{A CKAN API token (required, character)}
}
\description{
Configure default WAStD settings
}
\details{
\code{wastdr_setup} sets WAStD API connection details. \code{wastdr}'s
functions default to use the default URL, but require an API key to work.
}
\examples{
# WAStD users can run:
wastdr_setup(api_url = "https://strandings.dpaw.wa.gov.au/api/1/",
             api_token = "c12345asdfqwer")

# Not specifying the default WAStD API URL will reset the WAStD URL to its
# default "https://strandings.dpaw.wa.gov.au/api/1/":
wastdr_setup(api_token = "c12345asdfqwer")
}
