% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.R
\name{add_dates}
\alias{add_dates}
\title{Add turtle dates to a dataframe with datetime col \code{observation_start_time}.}
\usage{
add_dates(data, date_col = "observation_start_time", parse_date = TRUE)
}
\arguments{
\item{data}{a dataframe with datetime col \code{observation_start_time}, e.g. the
output of ODKC form "turtle track or nest", "predator or disturbance", or
"marine wildlife incident". If the date_col is plain text, it will be turned
into a tz-aware datetime, else it is expected to be POSIXct / POSIXt.}

\item{date_col}{(chr) The column name of the datetime to annotate.
Default: \code{"observation_start_time"}.}

\item{parse_date}{(lgl) Whether the date_col needs to be parsed from character
into a date format (TRUE, default) or already comes as a POSIXct/POSIXt.}
}
\value{
The initial dataframe plus new columns
\itemize{
\item calendar_date_awst (POSIXct) The calendar date in GMT+08 (AWST) as
POSIXct, an ISO datetime in GMT+00.
\item calendar_date_awst_text (chr) The calendar date in GMT+08 (AWST) as
character to prevent spreadsheet programs from corrupting the values.
\item calendar_year (int) The calendar year in GMT+08 (AWST) as integer.
\item turtle_date (POSIXct) The turtle date,
see \code{\link{datetime_as_turtle_date}}.
\item season (int) The season, see \code{\link{datetime_as_season}}.
\item season_week (int) The season week,
see \code{\link{datetime_as_seasonweek}}.
\item iso_week (int) The season week,
see \code{\link{datetime_as_isoweek}}.
}
}
\description{
\lifecycle{stable}
}
\seealso{
Other helpers: 
\code{\link{absolute_angle}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{leaflet_basemap}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{read_table}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{verify_odkc_turtledata}()},
\code{\link{verify_wamtram_data}()},
\code{\link{verify_wastd_data}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
