% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dates.R
\name{add_dates_sve}
\alias{add_dates_sve}
\title{Add turtle dates to a dataframe with datetime col `survey_end_time`.}
\usage{
add_dates_sve(data)
}
\arguments{
\item{data}{a dataframe with datetime col `survey_end_time`, e.g. the
output of ODKC form "site visit end".}
}
\value{
The initial dataframe plus new columns

  * calendar_date_awst
  * turtle_date
  * season
  * season_week
  * iso_week
}
\description{
\lifecycle{stable}
}
