% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_nest_labels.R
\name{add_nest_labels}
\alias{add_nest_labels}
\title{Add labels for species and nest/track type to parsed
\code{turtle-nest-encounters}.}
\usage{
add_nest_labels(nests)
}
\arguments{
\item{nests}{(tibble) A dataframe or tibble of parsed
\code{turtle-nest-encounters}.}
}
\value{
The dataframe with two new columns, \code{species_colours} and
\code{nest_type_text}.
}
\description{
Adds two columns, two new columns, \code{species_colours} and
\code{nest_type_text}, to parsed \code{turtle-nest-encounters} containing
\code{WAStD} \code{species} and \code{nest_type}. The new columns can be used
to create \code{leaflet} markers with
\code{leaflet::makeAwesomeIcon(
  text = ~nest_type_text, markerColor = ~species_colours)}.
}
\details{
\lifecycle{stable}
}
\seealso{
Other wastd: 
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_emergence_success}()},
\code{\link{ggplot_hatching_success}()},
\code{\link{ggplot_nesting_success_per_area_season_species_pct}()},
\code{\link{ggplot_nesting_success_per_area_season_species}()},
\code{\link{ggplot_sighting_status_per_area_season_species}()},
\code{\link{ggplot_total_emergences_per_area_season_species}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{map_dist}()},
\code{\link{map_fanangles}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{map_wastd}()},
\code{\link{nesting_success_per_area_day_species}()},
\code{\link{nesting_success_per_area_season_species}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{print.wastd_api_response}()},
\code{\link{sighting_status_per_area_season_species}()},
\code{\link{sighting_status_per_site_season_species}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{total_emergences_per_area_season_species}()},
\code{\link{total_emergences_per_site_season_species}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
