% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_encounters}
\alias{animal_encounters}
\title{AnimalEncounter WAStD API response}
\format{An object of class \code{wastd_api_response} of length 3.}
\source{
https://strandings.dpaw.wa.gov.au/api/1/animal-encounters/?taxon=Cheloniidae&limit=10&format=json
}
\usage{
animal_encounters
}
\description{
AnimalEncounter WAStD API response
}
\examples{
    data(animal_encounters)
    data(tags)
    ae <- get_wastd("animal-encounters",
        query = list(taxon = "Cheloniidae", limit = 10, format = "json"))
    ta <- parse_animal_encounters(ae)
    testthat::expect_equal(ae, animal_encounters)
    testthat::expect_equal(ta, tags)
}
\keyword{datasets}
