% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal_encounters}
\alias{animal_encounters}
\title{AnimalEncounter WAStD API response}
\format{An object of class \code{wastd_api_response} of length 3.}
\source{
https://strandings.dpaw.wa.gov.au/api/1/animal-encounters/?taxon=Cheloniidae&format=json
}
\usage{
animal_encounters
}
\description{
This API response is parsed into wastr's data "tags"
}
\examples{
# Prove that animal_encounters parses to animals

library(dplyr)
data("animal_encounters")
data("animals")
fresh_animals <- parse_animal_encounters(animal_encounters)
}
\keyword{datasets}
