% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_as_season.R
\name{datetime_as_season}
\alias{datetime_as_season}
\title{Calculate the "turtle season" (FY) from a given datetime}
\usage{
datetime_as_season(datetime)
}
\arguments{
\item{datetime}{(dttm) A datetime}
}
\value{
The season as int, e.g. 2017
}
\description{
\lifecycle{stable}
}
\details{
Return the earliest year of the season start.

In practice, the turtle date is the date component (year, month, day) of the
datetime minus 12 hours.
}
\examples{
# These datetimes are turtle season 2017:
datetime_as_season(httpdate_as_gmt08("2017-06-30T15:59:59Z")) # 2016
datetime_as_season(httpdate_as_gmt08("2017-06-30T16:00:00Z")) # 2017
datetime_as_season(httpdate_as_gmt08("2017-08-30T06:38:43Z")) # 2017
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates_sve}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{sf_as_tbl}()},
\code{\link{skip_test_if_odkc_offline}()},
\code{\link{skip_test_if_wastd_offline}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
