% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{disturbance_by_season}
\alias{disturbance_by_season}
\title{Summarise disturbance by season and cause}
\usage{
disturbance_by_season(value)
}
\arguments{
\item{value}{The ouput of \code{wastd_GET("disturbance-observations") %>%
parse_disturbance_observations()}}
}
\description{
Summarise disturbance by season and cause
}
