% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_minimal_tsc_turtledata.R
\name{download_minimal_tsc_turtledata}
\alias{download_minimal_tsc_turtledata}
\title{Download a minimal dataset of turtle observations from TSC}
\usage{
download_minimal_tsc_turtledata(
  source = "odk",
  year = NULL,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{source}{The TSC Encounter source, default: "odk".}

\item{year}{The earliest calendar year to be returned. Default: NULL.
This value, if not NULL, will be used to construct a query parameter
\code{when_year_gte=2019}}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A list of tibbles:
\itemize{
\item users A tibble of user names, roles, and contact details which can be
  used to resolve submitted user names to TSC user IDs
\item enc A tibble of encounters (source, source ID, QA status) which can be
  used to determine which encounters to create / update / skip.
\item surveys A tibble of surveys reconstructed from source and year.
}
}
\description{
Download a minimal dataset of turtle observations from TSC
}
