% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_odkc_turtledata_2019.R
\name{download_odkc_turtledata_2019}
\alias{download_odkc_turtledata_2019}
\title{Download and save Turtle data from ODK Central.}
\usage{
download_odkc_turtledata_2019(
  datafile = here::here("wa-turtle-programs", "data_odkc.rda"),
  extrafile = here::here("wa-turtle-programs", "data_odkc_extra.rda"),
  local_dir = here::here("wa-turtle-programs", "media"),
  prod = "https://odkcentral.dbca.wa.gov.au",
  uat = "https://odkcentral-uat.dbca.wa.gov.au",
  tz = "Australia/Perth",
  verbose = TRUE
)
}
\arguments{
\item{datafile}{The target filename to save the final data to.}

\item{extrafile}{The target filename to save the UAT data to.}

\item{local_dir}{A local directory to which to save the attachment files.
Attachment filepaths will be relative to the directory.
The directory and its parent folders will be created if not existing.}

\item{prod}{The ODKC PROD server,
default: "https://odkcentral.dbca.wa.gov.au".}

\item{uat}{The ODKC UAT server,
default: "https://odkcentral-uat.dbca.wa.gov.au".}

\item{tz}{The local timezone, default: "Australia/Perth".}
}
\value{
A list of tibbles and sf objects:

  * downloaded_on An UTC POSIXct timestamp of the data snapshot.
  * tracks The turtle tracks and nests from form "Turtle Track or Nest 1.0".
  * tracks_dist Individual disturbances recorded against tracks, one row per
    disturbance.
  * tracks_log Individual nest tags recorded against nests, one row per tag.
  * tracks_fan_outliser Individual hatchling track outliers recorded against
    hatched nests, one row per outlier.
  * dist The disturbance and predation records from form "Predator or
    Disturbance 1.1".
  * mwi Strandings and rescues from the form "Marine Wildlife Incident 0.6 ".
  * mwi_dmg Individual injuries recorded against mwi, one record per injury.
  * svs Survey start points from form "Site Visit Start 1.3".
  * sve Survey end points from form "Site Visit End 1.2".
  * sites An sf object of known TSC sites.
  * areas An sf object of known TSC localities.
}
\description{
Download and save Turtle data from ODK Central.
}
