% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_w2_data.R
\name{download_w2_data}
\alias{download_w2_data}
\title{Download Turtle Tagging data from WAMTRAM2.}
\usage{
download_w2_data(
  ord = c("YmdHMS", "Ymd"),
  tz = "Australia/Perth",
  db_drv = Sys.getenv("W2_DRV"),
  db_srv = Sys.getenv("W2_SRV"),
  db_name = Sys.getenv("W2_DB"),
  db_user = Sys.getenv("W2_UN"),
  db_pass = Sys.getenv("W2_PW"),
  db_port = Sys.getenv("W2_PT"),
  verbose = wastdr::get_wastdr_verbose(),
  dsn = Sys.getenv("W2_DSN"),
  use_rodbc = Sys.getenv("W2_RODBC", FALSE),
  save = NULL,
  compress = "xz"
)
}
\arguments{
\item{ord}{Lubridate orders, default: \code{c("YmdHMS", "Ymd")}.}

\item{tz}{Timezone, default: \code{"Australia/Perth"}.}

\item{db_drv}{Database driver, default: \code{Sys.getenv("W2_DRV")} which should
resolve to e.g. \code{"/usr/lib/x86_64-linux-gnu/odbc/libtdsodbc.so"}.}

\item{db_srv}{Database server, default: \code{Sys.getenv("W2_SRV")}, which should
resolve to a valid server hostname, e.g. \code{"myserver.corporateict.domain"}.}

\item{db_name}{The database name, default: \code{Sys.getenv("W2_DB")}, which should
resolve to a valid database name, e.g. \code{turtle_tagging}.}

\item{db_user}{The read-permitted database user,
default: \code{Sys.getenv("W2_UN")}.}

\item{db_pass}{The database user's password, default: \code{Sys.getenv("W2_PW")}.}

\item{db_port}{The database port, default: \code{Sys.getenv("W2_PT")}, which
should resolve to a numeric port, e.g. \code{1234}.}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{dsn}{The DSN for Windows systems, default: \code{Sys.getenv("W2_DSN")}.}

\item{use_rodbc}{Whether to use the RODBC library (if TRUE, best for Windows
systems), or the odbc/DBI library (if FALSE, default, best for GNU/Linux
systems).}

\item{save}{If supplied, the filepath to save the data object to.}

\item{compress}{The saveRDS compression parameter, default: "xz".
Set to FALSE for faster writes and reads but larger filesize.}
}
\value{
A structure of class "wamtram_data" containing a named list of
sanitised tables from the Turtle Tagging DB:
\itemize{
\item Metadata:
\itemize{
\item downloaded_on
\item w2_tables
}
\item Personnel:
\itemize{
\item persons
}
\item Data entry:
\itemize{
\item data_entry_batches "TRT_ENTRY_BATCHES"
\item data_entry
\item data_entry_operators
}
\item Lookups:
\itemize{
\item lookup_beach_positions
\item lookup_body_parts
\item lookup_conditions
\item lookup_damage_causes
\item lookup_damage_codes
\item lookup_datum_codes
\item lookup_egg_count_methods
\item lookup_id_types
\item lookup_measurement_types
\item lookup_pit_tag_states
\item lookup_sample_tissue_type
\item lookup_tag_states
}
\item Places:
\itemize{
\item sites
}
\item Encounters:
\itemize{
\item enc
\item enc_qa
}
\item Observations:
\itemize{
\item obs_flipper_tags
\item obs_pit_tags
\item obs_damages
\item obs_measurements
\item obs_samples
}
\item Reconstructed:
\itemize{
\item reconstructed_pit_tags
\item reconstructed_tags
\item reconstructed_turtles
}
}
}
\description{
This function requires an installed ODBC driver for MS SQL Server 2012.
The database credentials are handled via environment variables.
}
\details{
In Windows systems, create a user defined DSN with settings
\itemize{
\item name WAMTRAMPROD
\item server kens-mssql-001-prod.corporateict.domain
\item SQL auth usign login ID and password entered by user
\item trust server vertificate (this is where odbc falls over)
}

Add to .Renviron:
W2_RODBC=TRUE
W2_DSN="WAMTRAMPROD"
}
\examples{
\dontrun{
# Credentials are set in .Renviron

wamtram_data <- download_w2_data()


# develop:
ord <- c("YmdHMS", "Ymd")
tz <- "Australia/Perth"
verbose <- wastdr::get_wastdr_verbose()

con <- DBI::dbConnect(
  odbc::odbc(),
  Driver   = Sys.getenv("W2_DRV"),
  Server   = Sys.getenv("W2_SRV"),
  Database = Sys.getenv("W2_DB"),
  UID      = Sys.getenv("W2_UN"),
  PWD      = Sys.getenv("W2_PW"),
  Port     = Sys.getenv("W2_PT")
)
}
}
\seealso{
Other wamtram: 
\code{\link{w2_online}()}
}
\concept{wamtram}
