% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wastd_sites.R
\name{download_wastd_sites}
\alias{download_wastd_sites}
\title{Download Areas of type Site from WAStD}
\usage{
download_wastd_sites(
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose(),
  save = NULL,
  compress = "xz"
)
}
\arguments{
\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{save}{If supplied, the filepath to save the data object to.}

\item{compress}{The saveRDS compression parameter, default: "xz".
Set to FALSE for faster writes and reads but larger filesize.}
}
\value{
A list
\itemize{
\item downloaded_on A timestamp of the download
\item localities A \code{sf} object of WAStD Areas of type Locality
\item sites An \code{sf} object of WAStD Areas of type Sites joined to Localities.
}
}
\description{
Download Areas of type Site from WAStD
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata_area}()},
\code{\link{filter_wastd_turtledata_seasons}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{handle_http_status}()},
\code{\link{wastd_GET}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_bulk_post}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_create_update_skip}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
