% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_wastd_turtledata.R
\name{download_wastd_turtledata}
\alias{download_wastd_turtledata}
\title{Download all turtle data from WAStD}
\usage{
download_wastd_turtledata(max_records = NULL, verbose = get_wastdr_verbose())
}
\arguments{
\item{max_records}{<int> The max number of records to retrieve,
default: NULL (all).}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 class "wastd_data" with items:
\itemize{
  \item downloaded_on An UTC POSIXct timestamp of the data snapshot generated
        by \code{\link{Sys.time}}.
  \item sites An sf object of known WAStD sites.
  \item areas An sf object of known WAStD localities.
  \item surveys A tibble of surveys, reconstructed by WAStD from
        Site Visit Start and End records (where given) and
        (TurtleNest)Encounters.
  \item animals A tibble of AnimalEncounters, which includes live sightings,
        rescues, strandings, and other incidents, and turtle tagging.
        Data is included from 2016 and onwards.
  \item tracks Turtle tracks and nests.
  \item nest_dist Turtle Nest and General disturbance / predation.
  \item nest_tags Turtle Nest Tags.
  \item nest_excavations Turtle Nest Excavations - hatching and emergence
        success.
  \item hatchling_morph Hatchling morphometrics from next exacavations.
  \item nest_fans Hatchling emergence tracks - main fan.
  \item nest_fan_outliers Hatchling emergence tracks - individual outliers.
  \item nest_lightsources - Light sources present during hatchling emergence.
}
}
\description{
\lifecycle{maturing}
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
