% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{exclude_training_species_odkc}
\alias{exclude_training_species_odkc}
\title{Exclude ODKC records with \code{species} "corolla-corolla".}
\usage{
exclude_training_species_odkc(value)
}
\arguments{
\item{value}{The output of \code{parse_disturbance_observations()}.}
}
\description{
Exclude ODKC records with \code{species} "corolla-corolla".
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{disturbance_by_season_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_dead_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{filter_predation_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
