% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_possibly.R
\name{extract_possibly}
\alias{extract_possibly}
\title{Wrap magrittr::extract to return field value or -1}
\usage{
extract_possibly(...)
}
\arguments{
\item{...}{A list and a field name}
}
\description{
Wrap magrittr::extract to return field value or -1
}
\examples{
  lol <- list(field1=1, field2=2, field3=3, field4=4, field5=5)
  lol \%>\% extract_possibly("field1")
  lol \%>\% extract_possibly("field8")
}
\seealso{
\url{https://rdrr.io/cran/purrr/man/safely.html}
}
