% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_mwi.R
\name{filter_dead_odkc}
\alias{filter_dead_odkc}
\title{Filter strandings to mortalities for ODKC data}
\usage{
filter_dead_odkc(data)
}
\arguments{
\item{data}{A dataframe with column "health", e.g. data from form
"Marine Wildlife Incident 0.6".}
}
\value{
The dataframe with rows matching mortal outcomes.
}
\description{
\lifecycle{stable}
}
\seealso{
Other odkc: 
\code{\link{add_nest_labels_odkc}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_alive_odkc}()},
\code{\link{filter_disturbance_odkc}()},
\code{\link{filter_predation_odkc}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
