% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_dist.R
\name{general_disturbance_by_season_odkc}
\alias{general_disturbance_by_season_odkc}
\title{Tally ODKC General disturbances by season, cause, and encounter type}
\usage{
general_disturbance_by_season_odkc(data)
}
\arguments{
\item{data}{ODKC dist}
}
\description{
\lifecycle{stable}
}
\examples{
data("odkc_data")
odkc_data$dist \%>\% general_disturbance_by_season_odkc()
}
\seealso{
Other odkc: 
\code{\link{add_hatching_emergence_success_odkc}()},
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{hatching_emergence_success_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{map_tracks_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
