% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_more.R
\name{get_more}
\alias{get_more}
\title{Decide whether the given download limit has been reached}
\usage{
get_more(total_count, limit)
}
\arguments{
\item{total_count}{<int> The total count of downloaded records}

\item{limit}{<int> or NULL The max count of records to download}
}
\value{
<lgl> TRUE if limit has not been reached or limit is NULL, else FALSE
}
\description{
\lifecycle{stable}
}
\examples{
testthat::expect_true(get_more(0L, NULL))
testthat::expect_true(get_more(1000L, NULL))
testthat::expect_true(get_more(9L, 10))
testthat::expect_false(get_more(10L, 10))
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
