% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_num_field.R
\name{get_num_field}
\alias{get_num_field}
\title{From a list of unnamed lists, or the data.frame equivalent, extract a field or -1}
\usage{
get_num_field(lol, field)
}
\arguments{
\item{lol}{(list) A list of lists}

\item{field}{(chr) The name of a field inside the list of lists}
}
\value{
The field value or -1
}
\description{
The given list of lists is unlisted,
  \code{\link{extract_possibly}} extracts the value of the field or -1,
  another call to unlist deduplicates, and finally the result is cast to numeric.
}
\examples{
nest_json <- get_wastd("turtle-nest-encounters")
nest_hs <- nest_json$content \%>\% {
  tibble::tibble(
    hatching_success = obs \%>\% purrr::map(get_num_field, "hatching_success") \%>\% as.numeric
 )
 }
}
