% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{ggplot_total_emergences_per_area_season_species}
\alias{ggplot_total_emergences_per_area_season_species}
\title{Return a stacked ggplot barchart of emergences}
\usage{
ggplot_total_emergences_per_area_season_species(data)
}
\arguments{
\item{data}{The output of \code{total_emergences_per_area_season_species}, a
summary of \code{wastd_data}.}

\item{species}{The species to filter the data by.
Default: "Natator depressus".
Other options:
"Chelonia mydas" (Green), "Eretmochelys imbricata" (Hawksbill),
"Caretta caretta" (Loggerhead), "Lepidochelys olivacea" (Olive Ridley)}
}
\value{
A plotly figure
}
\description{
Facets: species
}
\examples{
data(wastd_data)
wastd_data \%>\%
  total_emergences_per_area_season_species() \%>\%
  ggplot_total_emergences_per_area_season_species()
}
