% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{ggplot_track_success_by_date}
\alias{ggplot_track_success_by_date}
\title{Plot the track success (absolute numbers) of a given species as time series}
\usage{
ggplot_track_success_by_date(
  data,
  speciesname,
  placename = "",
  prefix = "",
  local_dir = here::here(),
  export = FALSE
)
}
\arguments{
\item{data}{The output of \code{\link{track_success}}}

\item{speciesname}{The species name, e.g. "natator-depressus"}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}

\item{local_dir}{A local directory to save the plot to,
default: \code{here::here()}.}

\item{export}{Whether to export the figure as a file, default: FALSE}
}
\description{
\lifecycle{stable}
}
\examples{
data("wastd_data")
sp <- wastd_data$tracks$species[[1]]
track_success(wastd_data$tracks) \%>\% ggplot_track_success_by_date(sp)
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{duplicate_surveys}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{map_dist}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
