% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gj_linestring_to_st_linestring.R
\name{gj_linestring_to_st_linestring}
\alias{gj_linestring_to_st_linestring}
\title{Convert a GeoJSON linestring into an st_linestring}
\usage{
gj_linestring_to_st_linestring(gj_ls)
}
\arguments{
\item{gj_ls}{A GeoJSON linestring}
}
\value{
The linestring as \code{sf::sfg}
}
\description{
Convert a GeoJSON linestring into an st_linestring
}
\examples{
\dontrun{
gj_ls <- paste0(
  "-14.803659 128.403426 10.9 5.9;",
  "-14.803719 128.40326 1.7 1.9;",
  "-14.803756 128.40317 1.9 1.7;",
  "-14.80383 128.402983 1.7 1.6;",
  "-14.803913 128.402786 1.4 1.6;"
)
x <- gj_linestring_to_st_linestring(gj_ls)
class(x) == c("XYZM", "LINESTRING", "sfg")
}
}
\seealso{
Other helpers: 
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
