% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_http_status.R
\name{handle_http_status}
\alias{handle_http_status}
\title{Emit a specific warning messages depending on HTTP status}
\usage{
handle_http_status(response, verbose = wastdr::get_wastdr_verbose())
}
\arguments{
\item{response}{An HTTP response.}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\description{
Status 401: Instruct how to authenticate.
Status http_error: warn and show status.
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_sites}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{wastd_GET}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_bulk_post}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_create_update_skip}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
