% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpdate_as_gmt08_turtle_date.R
\name{httpdate_as_gmt08_turtle_date}
\alias{httpdate_as_gmt08_turtle_date}
\title{Calculate the "turtle date" from a given datetime}
\usage{
httpdate_as_gmt08_turtle_date(datestring)
}
\arguments{
\item{datestring}{(character) A UTC HTTP date as character string,
e.g. '2016-11-20T21:46:57.163000Z' with time zone included}
}
\value{
A POSIXct date
}
\description{
\lifecycle{stable}
}
\details{
Return the actual date for afternoon and evening datetimes
(noon to midnight) or the date of the previous day for morning datetimes
(midnight to noon).
The turtle date keeps observations made between sunset and sunrise,
as well as on the "morning after" together by assigning the date of the start
of observations.

In practice, the turtle date is the date component (year, month, day) of the
datetime minus 12 hours.

This date convention is widely used in the turtle research community.
}
\examples{
# These datetimes are turtle date 2016-11-20:
httpdate_as_gmt08_turtle_date("2016-11-20T04:00:00Z") # noon AWST
httpdate_as_gmt08_turtle_date("2016-11-21T03:59:59Z") # noon AWST - 1 sec
}
\seealso{
Other helpers: 
\code{\link{absolute_angle}()},
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
