% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpdate_as_season.R
\name{httpdate_as_season}
\alias{httpdate_as_season}
\title{Calculate the "turtle season" (FY) from a given datetime}
\usage{
httpdate_as_season(datestring)
}
\arguments{
\item{datestring}{(character) A UTC HTTP date as character string,
e.g. '2016-11-20T21:46:57.163000Z' with time zone included}
}
\value{
A POSIXct date
}
\description{
Calculate the "turtle season" (FY) from a given datetime
}
\details{
Return the earliest year of the season start.


In practice, the turtle date is the date component (year, month, day) of the
datetime minus 12 hours.

This date convention is widely used in the turtle research community.
}
\examples{
# These datetimes are turtle season 2017:
httpdate_as_season("2017-06-30T15:59:59Z") == 2016
httpdate_as_season("2017-06-30T16:00:00Z") == 2017
}
