% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/httpdate_as_season.R
\name{httpdate_as_season}
\alias{httpdate_as_season}
\title{Calculate the "turtle season" (FY) from a given UTC HTTP datestring}
\usage{
httpdate_as_season(datestring)
}
\arguments{
\item{datestring}{(character) A UTC HTTP date as character string,
e.g. '2016-11-20T21:46:57.163000Z' with time zone included}
}
\value{
The season as int, e.g. 2017
}
\description{
Calculate the "turtle season" (FY) from a given UTC HTTP datestring
}
\details{
Return the earliest year of the season start.
E.g. any date in the fiscal year 2017/18 will be season 2017.
Care has to be taken to calculate the offset through days rather than months,
which can result in invalid dates such as Feb 30 or June 31.
}
\examples{
# These datetimes are turtle season 2017:
httpdate_as_season("2017-06-30T15:59:59Z") # 2016
httpdate_as_season("2017-06-30T16:00:00Z") # 2017
httpdate_as_season("2017-08-30T06:38:43Z") # 2017
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates_sve}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{sf_as_tbl}()},
\code{\link{skip_test_if_odkc_offline}()},
\code{\link{skip_test_if_wastd_offline}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
