% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/humanize.R
\name{humanize}
\alias{humanize}
\title{Convert between url-safe dash-separated-names and Human Readable Title Case}
\usage{
humanize(urlsafe_string)
}
\arguments{
\item{urlsafe_string}{(chr) A url-safe-string}
}
\value{
a Human Readable Title Cased String
}
\description{
\lifecycle{stable}
}
\examples{
humanize("chelonia-mydas")
humanize("natator-depressus")
humanize("successful-crawl")
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates_sve}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{sf_as_tbl}()},
\code{\link{skip_test_if_odkc_offline}()},
\code{\link{skip_test_if_wastd_offline}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
