% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{list_survey_count}
\alias{list_survey_count}
\title{Create a datatable of survey counts from the output of \code{parse_surveys}.}
\usage{
list_survey_count(surveys, placename = "")
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{placename}{(string) The place name, used in labels. Default: ""}
}
\description{
Create a datatable of survey counts from the output of \code{parse_surveys}.
}
