% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_user_mapping.R
\name{make_user_mapping}
\alias{make_user_mapping}
\title{Generate a TSC user mapping from a given list of usernames}
\usage{
make_user_mapping(odkc_data)
}
\description{
Generate a TSC user mapping from a given list of usernames
}
