% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_dist.R
\name{map_dist}
\alias{map_dist}
\title{Map Disturbance observations interactively.}
\usage{
map_dist(dist, fmt = "\%d/\%m/\%Y \%H:\%M")
}
\arguments{
\item{dist}{The output of \code{wastdr::parse_disturbance_observations()}.}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}
}
\value{
A leaflet map
}
\description{
Map Disturbance observations interactively.
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle each disturbance class
separately. The maps auto-zooms to the extent of data given.
}
