% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_nests.R
\name{map_nests}
\alias{map_nests}
\title{Map tagged Turtle Nests interactively.}
\usage{
map_nests(
  data,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  cluster = FALSE
)
}
\arguments{
\item{data}{The output of \code{parse_nesttag_observations}.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}
}
\value{
A leaflet map
}
\description{
Map tagged Turtle Nests interactively.
}
\details{
Creates a Leaflet map with an interactive legend.
The maps auto-zooms to the extent of data given.
}
