% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tracks.R
\name{map_tracks}
\alias{map_tracks}
\title{Map TurtleNestEncounters interactively.}
\usage{
map_tracks(
  tracks,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  cluster = FALSE,
  ts = FALSE
)
}
\arguments{
\item{tracks}{The output of \code{parse_turtle_nest_encounters} and \code{add_nest_labels}.}

\item{sites}{An sf object of sites with \code{site_name} and polygon geom, e.g.
\code{turtleviewer::turtledata$sites}.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}

\item{ts}{Whether to render the data as additional timeseries (warning: slow)}
}
\value{
A leaflet map
}
\description{
\lifecycle{maturing}
}
\details{
Creates a Leaflet map with an interactive legend offering to
toggle each species separately.
The maps auto-zooms to the extent of data given.
}
\examples{
\dontrun{
data("wastd_data")
map_tracks(wastd_data$tracks, sites = wastd_data$sites, cluster = T)
map_tracks(wastd_data$tracks, sites = wastd_data$sites, cluster = F)
}
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{duplicate_surveys}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{map_dist}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
