% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tracks_odkc.R
\name{map_tracks_odkc}
\alias{map_tracks_odkc}
\title{Map tracks directly from ODK data}
\usage{
map_tracks_odkc(
  tracks,
  sites = NULL,
  wastd_url = wastdr::get_wastd_url(),
  fmt = "\%d/\%m/\%Y \%H:\%M",
  tz = "Australia/Perth",
  cluster = FALSE,
  ts = FALSE
)
}
\arguments{
\item{tracks}{The output of \code{parse_turtle_nest_encounters} and \code{add_nest_labels}.}

\item{sites}{An sf object of sites with \code{site_name} and polygon geom, e.g.
\code{turtleviewer::turtledata$sites}.}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}

\item{fmt}{The desired date format, default: "d/m/Y H:M"}

\item{tz}{The lubridate timezone, default: "Australia/Perth}

\item{cluster}{If TRUE, cluster map markers. Default: FALSE.
Note: In some places, the aerial background layer does not provide imagery
at sufficient zoom levels, and therefore restricts the map zoom at levels
where the cluster markers don't expand. Switch to "Place names" to let
cluster markers expand.}

\item{ts}{Whether to render the data as additional timeseries (warning: slow)}
}
\value{
A leaflet map
}
\description{
Map tracks directly from ODK data
}
\details{
Creates a Leaflet map with an interactive legend offering to toggle
each species separately. The maps auto-zooms to the extent of data given.

This map function uses data from ODK Central / ruODK using form
"Turtle Track or Nest 1.0".
}
\examples{
\dontrun{
data("odkc_data")
map_tracks_odkc(odkc_data$tracks, sites = odkc_data$sites, cluster = TRUE)
map_tracks_odkc(odkc_data$tracks, sites = odkc_data$sites, cluster = FALSE)
map_tracks_odkc(odkc_data$tracks, sites = NULL, cluster = TRUE)
map_tracks_odkc(odkc_data$tracks, sites = NULL, cluster = FALSE)
}
}
\seealso{
Other odkc: 
\code{\link{add_hatching_emergence_success_odkc}()},
\code{\link{add_nest_labels_odkc}()},
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{exclude_training_species_odkc}()},
\code{\link{filter_odkc_turtledata}()},
\code{\link{general_disturbance_by_season_odkc}()},
\code{\link{hatching_emergence_success_odkc}()},
\code{\link{map_dist_odkc}()},
\code{\link{map_mwi_odkc}()},
\code{\link{map_sv_odkc}()},
\code{\link{nest_disturbance_by_season_odkc}()},
\code{\link{nesting_type_by_area_season_age_species_odkc}()},
\code{\link{nesting_type_by_area_season_species_odkc}()},
\code{\link{nesting_type_by_season_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_age_species_odkc}()},
\code{\link{nesting_type_by_season_calendarday_species_odkc}()},
\code{\link{nesting_type_by_season_day_species_odkc}()},
\code{\link{nesting_type_by_season_species_odkc}()},
\code{\link{nesting_type_by_season_week_age_species_odkc}()},
\code{\link{nesting_type_by_season_week_site_species_odkc}()},
\code{\link{nesting_type_by_season_week_species_odkc}()},
\code{\link{nesting_type_by_site_season_age_species_odkc}()},
\code{\link{nesting_type_by_site_season_species_odkc}()}
}
\concept{odkc}
