% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wamtram.R
\name{map_wamtram}
\alias{map_wamtram}
\title{Map WAMTRAM data}
\usage{
map_wamtram(data, location = NULL, place = NULL)
}
\arguments{
\item{data}{The output of \code{wastdr::download_w2_data()}}
}
\value{
A leaflet map showing W2 sites and encounters
}
\description{
Map WAMTRAM data
}
\examples{
\dontrun{
data("w2_data")
map_wamtram(wamtram_data)

map_wamtram(wamtram_data, location_code="DH")
map_wamtram(wamtram_data, place_code="DHTB")
}
}
