% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wastd_wamtram_sites.R
\name{map_wastd_wamtram_sites}
\alias{map_wastd_wamtram_sites}
\title{Map WAStD and WAMTRAM sites}
\usage{
map_wastd_wamtram_sites(wastd_areas, wastd_sites, wamtram_sites)
}
\arguments{
\item{wastd_areas}{Areas from \code{wastd_data}.}

\item{wastd_sites}{Sites from \code{wastd_data}.}

\item{wamtram_sites}{Sites from \code{wamtram_data}.}
}
\value{
A leaflet map
}
\description{
This map is useful to compare and update WAStD with and from WAMTRAM sites.
}
\examples{
\dontrun{
library(wastdr)

# From canned data
wastd_sites <- wastdr::download_wastd_sites()
load("inst/wamtram_data.RData")
w2_data <- wamtram_data
map_wastd_wamtram_sites(wastd_sites$localities, wastd_sites$sites, w2_data$sites)


# From live data
ws <- wastdr::download_wastd_sites()

# split rows by place code, map only what place codes are not in wastd yet
map_wastd_wamtram_sites(areas, sites, w2_data$sites)
}
}
