% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{odkc_data}
\alias{odkc_data}
\title{Turtle Nesting Census data from ODK Centrail
First 6 records of each dataset, all sites.}
\format{
An object of class \code{list} of length 18.
}
\usage{
odkc_data
}
\description{
Turtle Nesting Census data from ODK Centrail
First 6 records of each dataset, all sites.
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
