% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{odkc_data}
\alias{odkc_data}
\title{Turtle Nesting Census data from ODK Central.}
\format{
An object of class \code{odkc_turtledata} of length 23.
}
\usage{
odkc_data
}
\description{
This dataset contains the truncated output from
\code{\link{download_odkc_turtledata_2019}}.
All WAStD sites and areas are included, while all other tables are truncated
with \code{head()}.
}
\seealso{
Other included: 
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
