% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_plan.R
\name{odkc_plan}
\alias{odkc_plan}
\title{Top-level Turtle Nesting Census data import Drake Plan}
\usage{
odkc_plan()
}
\description{
* Download all ODKC data including attachments
* Load existing nesting records from WAStD: load only a minimal set of
  source, source ID, QA status to determine later what to
  create / update / skip:
  * does not exist in WAStD: create (POST)
  * exists in WAStD with status "new": update (PATCH)
  * exists in WAStD with status higher than "new": skip (and message)
  * Make (transform) ODKC to WAStD data
  * Load transformed data into WAStD's API (create/update/skip)
  * No QA
}
\examples{
\dontrun{

# Step 1: New users (username, name, phone, email, role)
# 400 for existing, 201 for new
Append new users to spreadsheet: username, name, email, phone, role
users <- here::here("users.csv") \%>\%
  readr::read_csv(col_types = "ccccc") \%>\%
  wastdr::wastd_bulk_post("users",
  #api_url = Sys.getenv("WASTDR_API_DEV_URL"),
  #api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"),
  verbose = TRUE)

users_dev <- here::here("users.csv") \%>\%
  readr::read_csv(col_types = "ccccc") \%>\%
  wastdr::wastd_bulk_post("users",
  api_url = Sys.getenv("WASTDR_API_DEV_URL"),
  api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"),
  verbose = TRUE)

# save point for debug
save(atkn, aurl, vbse, updt, odkc_ex, odkc_tf, odkc_up,
wastd_data, wastd_users, user_mapping, file="odkc_import.RData")
load("odkc_import.RData")

wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_DEV_URL"),
                     api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_TEST_URL"),
                     api_token = Sys.getenv("WASTDR_API_TEST_TOKEN"))
Sys.setenv(ODKC_IMPORT_UPDATE_EXISTING=TRUE)
wastdr::odkc_plan()
drake::vis_drake_graph(wastdr::odkc_plan())
drake::clean()
drake::clean("user_qa") # after updating WAStD user aliases
drake::make(odkc_plan(), lock_envir = FALSE)
}
}
