% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_plan.R
\name{odkc_plan}
\alias{odkc_plan}
\title{Top-level Turtle Nesting Census data import Drake Plan}
\usage{
odkc_plan()
}
\description{
* Download all ODKC data including attachments
* Load existing nesting records from TSC: load only a minimal set of
  source, source ID, QA status to determine later what to create / update / skip:
  * does not exist in TSC: create (POST)
  * exists in TSC with status "new": update (PATCH)
  * exists in TSC with status higher than "new": skip (and message)
  * Make (transform) ODKC to WAStD data
  * Load transformed data into WAStD's API (create/update/skip)
  * No QA
}
\examples{
\dontrun{
odkc_plan()
drake::vis_drake_graph(odkc_plan())
make(odkc_plan())
}
}
