% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_as_wastd_tne.R
\name{odkc_tracks_as_wastd_tne}
\alias{odkc_tracks_as_wastd_tne}
\title{Transform odkc_data$tracks into WAStD TurtleNestEncounters.}
\usage{
odkc_tracks_as_wastd_tne(data, user_mapping)
}
\arguments{
\item{data}{A tibble of parsed ODKC Track or Nests,
e.g. \code{odkc_data$tracks}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
  \code{\link{wastd_POST}("turtle-nest-encounters")}
}
\description{
Transform odkc_data$tracks into WAStD TurtleNestEncounters.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
user_mapping = NULL # see odkc_plan for actual user mapping

odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne(user_mapping) \%>\%
  head(1) \%>\%
  jsonlite::toJSON()

tne <- odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne(user_mapping) \%>\%
  dplyr::mutate(reporter_id=4, observer_id=4) \%>\%  # missing users in local dev
  head()

tne_update <- dplyr::anti_join(x, tsc_data$enc, by="source_id")
tne_create <- dplyr::anti_join(tsc_data$enc, x, by="source_id")

tne_create \%>\%
  wastd_POST("turtle-nest-encounters",
    api_url = Sys.getenv("WASTDR_API_DEV_URL"),
    api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
}
}
