% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_animal_encounters.R
\name{parse_animal_encounters}
\alias{parse_animal_encounters}
\title{Parse a \code{wastd_api_response} of \code{animal-encounters} to tbl_df}
\usage{
parse_animal_encounters(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{animal-encounters}, e.g. \code{get_wastd("animal-encounters")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
\item area_name (chr)
\item area_type (chr)
\item area_id (int)
\item site_name (chr)
\item site_type (chr)
\item site_id (int)
\item survey_id (int)
\item survey_start_time (dttm)
\item survey_end_time (dttm)
\item survey_start_comments (chr)
\item survey_end_comments (chr)
\item datetime (dttm)
\item turtle_date (date)
\item season (int)
\item season_week (int) Number of completed weeks since fiscal year start
\item iso_week (int) Number of completed weeks since calendar year start
\item longitude (chr)
\item latitude (chr)
\item crs (chr)
\item location_accuracy (dbl)
\item name (chr)
\item species (chr)
\item health (chr)
\item sex (chr)
\item maturity (chr)
\item behaviour (chr)
\item habitat (chr)
\item activity (chr)
\item comments (chr)
\item nesting_event (chr)
\item checked_for_injuries (chr)
\item scanned_for_pit_tags (chr)
\item checked_for_flipper_tags (chr)
\item cause_of_death (chr)
\item cause_of_death_confidence (chr)
\item absolute_admin_url (chr)
\item obs (list)
\item source (chr)
\item source_id (chr)
\item encounter_type (chr)
\item status (chr)
}
}
\description{
\lifecycle{stable}
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_disturbance_by_season}()},
\code{\link{ggplot_emergence_success}()},
\code{\link{ggplot_hatching_success}()},
\code{\link{ggplot_hatchling_misorientation}()},
\code{\link{ggplot_nesting_success_per_area_season_species_pct}()},
\code{\link{ggplot_nesting_success_per_area_season_species}()},
\code{\link{ggplot_sighting_status_per_area_season_species}()},
\code{\link{ggplot_total_emergences_per_area_season_species}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{map_dist}()},
\code{\link{map_fanangles}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{map_wastd_wamtram_sites}()},
\code{\link{map_wastd}()},
\code{\link{nesting_success_per_area_day_species}()},
\code{\link{nesting_success_per_area_season_species}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{print.wastd_api_response}()},
\code{\link{sighting_status_per_area_season_species}()},
\code{\link{sighting_status_per_site_season_species}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{total_emergences_per_area_season_species}()},
\code{\link{total_emergences_per_site_season_species}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
