% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_encounterobservations.R
\name{parse_encounterobservations}
\alias{parse_encounterobservations}
\title{Parse a \code{wastd_api_response} of Encounter observations to tbl_df}
\usage{
parse_encounterobservations(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{turtle-nest-*}, e.g.
\code{wastd_GET("turtle-nest-encounters")}}
}
\value{
A \code{tbl_df} with non-repetitive columns of:
\itemize{
\item Encounter
\item Survey
\item Area
\item Site
\item Observation
}
}
\description{
\lifecycle{stable}
}
\details{
WAStD provides all subclasses of \code{wastd.observations.observation} together
with their encounter (survey, site, area, reporter, observer) via its API
as GeoJSON FeatureCollections.

This function acts as a generic parser for
\itemize{
\item "turtle-nest-hatchling-emergences"
\item "turtle-nest-hatchling-emergence-outliers"
\item "turtle-nest-hatchling-emergence-light-sources"
\item "turtle-nest-excavations"
\item "turtle-nest-tag-observations"
\item "turtle-nest-disturbance-observations"
}
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_emergence_success}()},
\code{\link{ggplot_hatching_success}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{map_dist}()},
\code{\link{map_fanangles}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{map_wastd}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
