% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveys.R
\name{parse_surveys}
\alias{parse_surveys}
\title{Parse a \code{wastd_api_response} of \code{surveys} to tbl_df}
\usage{
parse_surveys(wastd_api_response, wastd_url = wastdr::get_wastd_url())
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{surveys}, e.g. \code{get_wastd("surveys")}}

\item{wastd_url}{The base URL for WAStD, default: \code{get_wastd_url()}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
\item area_name (chr) The name of the Locality this Survey took place in.
\item area_id (int) The ID of the Locality this Survey took place in.
\item site_name (chr) The name of the Site this Survey took place in.
\item site_id (int) The ID of the Site this Survey took place in.
\item reporter (chr)
\item reporter_username (chr)
\item reporter_id (chr)
\item start_time (dttm) The actual start time of the survey.
\item end_time (dttm) The actual, automatically guessed (if 6h later), or
human guessed end time.
\item start_comments (chr) Comments at start plus QA messages from username
guessing
\item end_comments (chr) Comments at end plus QA messages from username
guessing. A mismatch
of original usernames can indicate an incorrectly picked Site Visit
End point.
\item turtle_date (dttm) The "turtle date" of the survey.
\item season_week (int) The number of completed weeks since fiscal year
start
\item iso_week (int) The number of completed weeks since calendar year
start
\item season (int) The "turtle season" of the survey, rolling over with
Fiscal Year.
\item source (chr) Where this record was born.
\item source_id (chr) The ODK record UID of the Site Visit Start.
\item end_source_id (chr) The ODK record UID of the Site Visit End.
\item status (chr) The QA status of the Survey. "Proofread" or "Curated"
indicate human QA edits.
\item device_id (chr) The unique ID of the device the Site Visit Start was
captured on.
\item end_device_id (chr) The unique ID of the device the Site Visit End
was captured on.
\item status (chr) QA status of the survey.
\item id (int) WAStD ID of the survey.
\item is_production (bool) Whether this survey is a real (production) or
a training survey.
\item absolute_admin_url (chr) The absolute URL path to edit the survey in
WAStD. Append this to the base \code{WASTD_URL}.
\item start_photo_url (chr) The URL of the start photo, if given.
Requires DBCA SSO when viewed.
\item end_photo_url (chr) The URL of the end photo, if given.
Requires DBCA SSO when viewed.
}
}
\description{
\lifecycle{maturing}
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{duplicate_surveys}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{map_dist}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
