% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_turtle_nest_encounters.R
\name{parse_turtle_nest_encounters}
\alias{parse_turtle_nest_encounters}
\title{Parse a \code{wastd_api_response} of \code{turtle-nest-encounters} to tbl_df}
\usage{
parse_turtle_nest_encounters(wastd_api_response)
}
\arguments{
\item{wastd_api_response}{A \code{wastd_api_response} of
\code{turtle-nest-encounters}, e.g. \code{get_wastd("turtle-nest-encounters")}}
}
\value{
A \code{tbl_df} with columns:
\itemize{
  \item area_name <chr>
  \item area_type <chr>
  \item area_id <int>
  \item site_name <chr>
  \item site_type <chr>
  \item site_id <int>
  \item datetime <dttm>
  \item longitude <dbl>
  \item latitude <dbl>
  \item crs <chr>
  \item location_accuracy <dbl>
  \item date <date>
  \item species <chr>
  \item nest_age <chr>
  \item nest_type <chr>
  \item name <chr>
  \item habitat <chr>
  \item disturbance <chr>
  \item absolute_admin_url <chr>
  \item obs <list>
  \item photos <list>
  \item hatching_success <dbl>
  \item emergence_success <dbl>
  \item clutch_size <dbl>
  \item source <chr>
  \item source_id <chr>
  \item encounter_type <chr>
  \item status <chr>
  \item observer <chr>
  \item reporter <chr>
}
}
\description{
Parse a \code{wastd_api_response} of \code{turtle-nest-encounters} to tbl_df
}
