% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_odkc_turtledata_2019.R
\name{print.odkc_turtledata}
\alias{print.odkc_turtledata}
\title{S3 print method for 'odkc_turtledata'.}
\usage{
\method{print}{odkc_turtledata}(x, ...)
}
\arguments{
\item{x}{An object of class `wastd_data` as returned by
\code{\link{download_odkc_turtledata_2019}}.}

\item{...}{Extra parameters for `print`}
}
\description{
Prints a short representation of data returned by
\code{\link{download_odkc_turtledata_2019}}.
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{odkc_data}},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
