% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_wastd.R
\name{print.wastd_api_response}
\alias{print.wastd_api_response}
\title{S3 print method for 'wastd_api_response'.}
\usage{
\method{print}{wastd_api_response}(x, ...)
}
\arguments{
\item{x}{An object of class `wastd_api_response` as returned by
\code{\link{get_wastd}}.}

\item{...}{Extra parameters for `print`}
}
\description{
Prints a short representation of data returned by
\code{\link{get_wastd}}.
}

