% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_w2_data.R
\name{read_table}
\alias{read_table}
\title{Read a table from a SQL connection using either RODBC or DBI}
\usage{
read_table(connection, table_name, rodbc = FALSE)
}
\arguments{
\item{connection}{A database connection as returned by RODBC or DBI}

\item{table_name}{The table to fetch all records for}

\item{rodbc}{Whether to use RODBC (if TRUE) or DBI (if FALSE, default)}
}
\description{
Read a table from a SQL connection using either RODBC or DBI
}
