% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wastd_data.R
\name{sighting_status_per_site_season_species}
\alias{sighting_status_per_site_season_species}
\title{Calculate sighting status for emergences per site, season, species}
\usage{
sighting_status_per_site_season_species(x)
}
\arguments{
\item{x}{An object of S3 class \code{wastd_data}, e.g. the output of
\code{wastdr::download_wastd_turtledata()}, optionally filtered to a locality
by \code{wastdr::filter_wastd_turtledata(area_name="Thevenard Island")}.}
}
\value{
A tibble with the summary data
}
\description{
Break up total emergences by sighting status:
}
\details{
\itemize{
\item Na - Unidentified (encounter with untagged animal)
\item New - Initial sighting (new tags applied onto untagged animal)
\item Resighting (existing tags, animal resighted at same site, same season)
\item Remigrant (existing tags, animal resighted at different or same site, different season)
}
}
\examples{
data(wastd_data)
wastd_data \%>\%
  sighting_status_per_site_season_species()
}
