% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_wastd_api.R
\name{skip_test_if_odkc_offline}
\alias{skip_test_if_odkc_offline}
\title{Check whether API token is set and WAStD is online, else skip test}
\usage{
skip_test_if_odkc_offline()
}
\description{
\lifecycle{stable}
}
\seealso{
Other helpers: 
\code{\link{add_dates_sve}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{sf_as_tbl}()},
\code{\link{skip_test_if_wastd_offline}()},
\code{\link{tdate_as_fdate}()},
\code{\link{urlize}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
