% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_create_update_skip.R
\name{split_create_update_skip}
\alias{split_create_update_skip}
\title{Top level helper to split all Turtle Nesting Census data per skip logic}
\usage{
split_create_update_skip(odkc_prep, tsc_data)
}
\arguments{
\item{odkc_prep}{ODKC data transformed into TSC format.}

\item{tsc_data}{Minimal TSC data to inform skip logic.}
}
\value{
A list of results from the various uploads.
  Each result is a `wastd_api_response` or a tibble of data (`_skip`).
}
\description{
Encounters and their related observations are uploaded to TSC:
}
\details{
* New encounters will be created
* Existing but unchanged (status=new) encounters will be updated
  if `update_existing=TRUE`, else skipped.
* Existing and value-added encounters (status > new) will be skipped.s
}
