% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{summarise_hatching_and_emergence_success}
\alias{summarise_hatching_and_emergence_success}
\title{\lifecycle{stable}}
\usage{
summarise_hatching_and_emergence_success(data)
}
\arguments{
\item{data}{The output of \code{wastd_data$nest_excavations}}
}
\description{
Utility to summarise a tibble of tracks with nest excavation data.
If the dataset does not contain variables \code{egg_count} or
\code{egg_count_calculated}, they will be added and set to
\code{NA_integer_}. This can happen if the subset of data retrieved from
WAStD happens to contain all NA in either of these variables, as
\code{\link{parse_encounterobservations}} drops columns with all NA.
}
\details{
Calculates:
\itemize{
\item count The count of nests
\item clutch_size_fresh Mean observed clutch size during nesting event
\item clutch_size_{mean, sd, min, max}  Reconstructed clutch size stats
\item hatching_success_{mean, sd, min, max} Hatching success stats
\item emergence_success_{mean, sd, min, max} Emergence success stats
}
}
\examples{
data("wastd_data")
summarise_hatching_and_emergence_success(wastd_data$nest_excavations)
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{map_dist}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{print.wastd_api_response}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()},
\code{\link{tracks_ts}()}
}
\concept{wastd}
