% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_tracks.R
\name{summarise_hatching_and_emergence_success}
\alias{summarise_hatching_and_emergence_success}
\title{Utility to summarise a tibble of tracks with nest excavtion data.}
\usage{
summarise_hatching_and_emergence_success(value)
}
\arguments{
\item{value}{The output of \code{parse_turtle_nest_encounters()}}
}
\description{
Calculates:
\itemize{
\item count The count of nests
\item clutch_size_fresh Mean observed clutch size during nesting event
\item clutch_size_mean Mean of reconstructed clutch size
\item clutch_size_sd SD of reconstructed clutch size
\item clutch_size_min Min of reconstructed clutch size
\item clutch_size_maz Max of reconstructed clutch size
\item hatching_success_{mean, sd, min, max} Hatching success stats
\item emergence_success_{mean, sd, min, max} Emergence success stats
}
}
