% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_count}
\alias{survey_count}
\title{Return the number of surveys for a given site_id and season.}
\usage{
survey_count(surveys, site_id, season)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{site_id}{The ID of a site, e.g. 22}

\item{season}{The desired season's start year, e.g. 2018}
}
\description{
Return the number of surveys for a given site_id and season.
}
