% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_hours_heatmap}
\alias{survey_hours_heatmap}
\title{Plot the surveyed hours from the output of \code{parse_surveys} as heatmap}
\usage{
survey_hours_heatmap(
  surveys,
  local_dir = here::here(),
  placename = "",
  prefix = "",
  export = FALSE
)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{local_dir}{A local directory to save the plot to,
default: \code{here::here()}.}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}

\item{export}{Whether to export the figure as a file, default: FALSE}
}
\description{
\lifecycle{stable}
}
\seealso{
Other wastd_surveys: 
\code{\link{duplicate_surveys}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()}
}
\concept{wastd_surveys}
