% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_surveys.R
\name{survey_hours_per_site_name_and_date}
\alias{survey_hours_per_site_name_and_date}
\title{Sum the hours surveyed per site_name and date from the output of \code{parse_surveys}.}
\usage{
survey_hours_per_site_name_and_date(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\value{
A tibble with columns date, site_name, hours_surveyed
}
\description{
Sum the hours surveyed per site_name and date from the output of \code{parse_surveys}.
}
