% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_surveys.R
\name{survey_season_stats}
\alias{survey_season_stats}
\title{Generate a season summary from the output of \code{parse_surveys}.}
\usage{
survey_season_stats(surveys)
}
\arguments{
\item{surveys}{(tibble) The output of \code{parse_surveys}.}
}
\description{
Generate a season summary from the output of \code{parse_surveys}.
}
\details{
Surveys, gruped by \code{season},
  summarised as first and last day of surveys, season length in days,
  number and total hours of surveys.
}
