% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_as_geojson.R
\name{tracks_as_geojson}
\alias{tracks_as_geojson}
\title{Convert tracks to a GeoJSON string}
\usage{
tracks_as_geojson(tracks)
}
\arguments{
\item{tracks}{The output of \code{parse_turtle_nest_encounters} and \code{add_nest_labels}.}
}
\description{
\lifecycle{stable}
}
\details{
Turns calendar_date_awst (required) into start and end (start + 1 day) dates,
converts dataframe to GeoJSON string.
}
\seealso{
Other helpers: 
\code{\link{absolute_angle}()},
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{gj_linestring_to_st_linestring}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{leaflet_basemap}()},
\code{\link{map_chr_hack}()},
\code{\link{mean_bearing}()},
\code{\link{mis_bearing}()},
\code{\link{odkc_works}()},
\code{\link{read_table}()},
\code{\link{sentencecase}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tun}()},
\code{\link{urlize}()},
\code{\link{verify_wastd_data}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
