% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_ts.R
\name{tracks_ts}
\alias{tracks_ts}
\title{Plot of nesting_type_by_season_day_species over time}
\usage{
tracks_ts(
  data,
  surveys,
  local_dir = here::here(),
  placename = "",
  prefix = "",
  export = FALSE
)
}
\arguments{
\item{data}{The output of \code{parse_turtle_nest_encounters}.}

\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{local_dir}{A local directory to save the plot to,
default: \code{here::here()}.}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}

\item{export}{Whether to export the figure as a file, default: FALSE}
}
\value{
A ggplot2 object. Saves the plot to .png.
}
\description{
Plot of nesting_type_by_season_day_species over time
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_emergence_success}()},
\code{\link{ggplot_hatching_success}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{map_dist}()},
\code{\link{map_fanangles}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{map_wastd}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()}
}
\concept{wastd}
