% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks_ts.R
\name{tracks_ts}
\alias{tracks_ts}
\title{Plot of nesting_type_by_season_day_species over time}
\usage{
tracks_ts(data, surveys, local_dir = ".", placename = "", prefix = "")
}
\arguments{
\item{data}{The output of \code{parse_turtle_nest_encounters}.}

\item{surveys}{(tibble) The output of \code{parse_surveys}.}

\item{local_dir}{A local directory to save the plot to, default: ".".}

\item{placename}{(string) The place name, used in labels.}

\item{prefix}{(string) A prefix for the filename. Default: ""}
}
\value{
A ggplot2 object. Saves the plot to .png.
}
\description{
Plot of nesting_type_by_season_day_species over time
}
\seealso{
Other wastd: 
\code{\link{add_nest_labels}()},
\code{\link{disturbance_by_season}()},
\code{\link{filter_alive}()},
\code{\link{filter_dead}()},
\code{\link{filter_disturbance}()},
\code{\link{filter_predation}()},
\code{\link{ggplot_track_success_by_date}()},
\code{\link{ggplot_track_successrate_by_date}()},
\code{\link{hatching_emergence_success_area}()},
\code{\link{hatching_emergence_success_site}()},
\code{\link{hatching_emergence_success}()},
\code{\link{list_survey_count}()},
\code{\link{list_survey_effort}()},
\code{\link{map_dist}()},
\code{\link{map_mwi}()},
\code{\link{map_nests}()},
\code{\link{map_tracks}()},
\code{\link{nesting_type_by_area_season_age_species}()},
\code{\link{nesting_type_by_area_season_species}()},
\code{\link{nesting_type_by_season_age_species}()},
\code{\link{nesting_type_by_season_calendarday_age_species}()},
\code{\link{nesting_type_by_season_calendarday_species}()},
\code{\link{nesting_type_by_season_day_species}()},
\code{\link{nesting_type_by_season_species}()},
\code{\link{nesting_type_by_season_week_age_species}()},
\code{\link{nesting_type_by_season_week_site_species}()},
\code{\link{nesting_type_by_season_week_species}()},
\code{\link{nesting_type_by_site_season_age_species}()},
\code{\link{nesting_type_by_site_season_species}()},
\code{\link{parse_animal_encounters}()},
\code{\link{parse_area_sf}()},
\code{\link{parse_area}()},
\code{\link{parse_encounterobservations}()},
\code{\link{parse_surveys}()},
\code{\link{parse_turtle_nest_encounters}()},
\code{\link{plot_survey_count}()},
\code{\link{plot_survey_effort}()},
\code{\link{print.wastd_api_response}()},
\code{\link{summarise_hatching_and_emergence_success}()},
\code{\link{summarise_wastd_data_per_day_site}()},
\code{\link{survey_count_heatmap}()},
\code{\link{survey_count}()},
\code{\link{survey_ground_covered}()},
\code{\link{survey_hours_heatmap}()},
\code{\link{survey_hours_per_person}()},
\code{\link{survey_hours_per_site_name_and_date}()},
\code{\link{survey_season_site_stats}()},
\code{\link{survey_season_stats}()},
\code{\link{survey_show_detail}()},
\code{\link{surveys_per_site_name_and_date}()},
\code{\link{track_success_by_species}()},
\code{\link{track_success}()}
}
\concept{wastd}
