% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turtle_nest_encounters}
\alias{turtle_nest_encounters}
\title{TurtleNestEncounter (tracks and nests) WAStD API response}
\format{An object of class \code{wastd_api_response} of length 3.}
\source{
https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=100&format=json
}
\usage{
turtle_nest_encounters
}
\description{
This API response is parsed into wastr's data "tracks".
}
\examples{
# Prove that turtle_nest_encounters_hatched parses to nests
library(dplyr)
data(turtle_nest_encounters)
data(tracks)
fresh_tracks <- parse_turtle_nest_encounters(turtle_nest_encounters)
}
\keyword{datasets}
