% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turtle_nest_encounters_hatched}
\alias{turtle_nest_encounters_hatched}
\title{TurtleNestEncounter (hatched nests) WAStD API response}
\format{\preformatted{List of 3
 $ content   :List of 10
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.6
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 80
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "paper"
  .. .. ..$ source_id         : chr "2015-10-29-09-15-00-114-9335-21-5937-hatched-natator-depressus"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2015 Nest "
  .. .. ..$ status            : chr "proofread"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.6
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2015-10-29T01:15:00Z"
  .. .. ..$ nest_age          : chr "old"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "natator-depressus"
  .. .. ..$ habitat           : chr "in-dune-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : chr "one dead hatchling found inside nest, no comprehensive egg count available"
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/80/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           :List of 1
  .. .. .. ..$ :List of 3
  .. .. .. .. ..$ operator : chr "Florian Mayer"
  .. .. .. .. ..$ status   : chr "proofread"
  .. .. .. .. ..$ timestamp: chr "2016-11-05T04:35:33.810322+00:00"
  .. .. ..$ observation_set   :List of 1
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "data_sheet"
  .. .. .. .. ..$ title           : chr "datasheet"
  .. .. .. .. ..$ attachment      : chr "encounter/2015-10-29-09-15-00-114-9335-21-5937-hatched-natator-depressus/2015-10-29-09-15-00-114-9335-21-5937-h"| __truncated__
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/2015-10-29-09-15-00-114-9335-21-5937-hatched-natator-depressus/2015-1"| __truncated__
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 332
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "odk"
  .. .. ..$ source_id         : chr "uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-18T21:51:01.925000Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "eretmochelys-imbricata"
  .. .. ..$ habitat           : chr "beach-above-high-water"
  .. .. ..$ disturbance       : chr "na"
  .. .. ..$ comments          : NULL
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/332/change/"
  .. .. ..$ photographs       :List of 2
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Nest_yho5It4"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Nest_yho5It4"
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Track"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Track_KbTA2oG"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Track_KbTA2oG"
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   :List of 2
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Nest_yho5It4"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Nest_yho5It4"
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Track"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Track_KbTA2oG"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:0d8e8420-f5bf-437c-b741-17de3c257d4d/Track_KbTA2oG"
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 401
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "odk"
  .. .. ..$ source_id         : chr "uuid:34d6ab5c-b19d-40b2-9f84-5d32b9ca6694"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-19T01:23:44.468000Z"
  .. .. ..$ nest_age          : chr "old"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "natator-depressus"
  .. .. ..$ habitat           : chr "na"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : NULL
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/401/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   : list()
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 492
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "odk"
  .. .. ..$ source_id         : chr "uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-19T21:59:25.763000Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "cheloniidae-fam"
  .. .. ..$ habitat           : chr "in-dune-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : NULL
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/492/change/"
  .. .. ..$ photographs       :List of 2
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Nest_KAdd13c"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Nest_KAdd13c"
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Track"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Track_04QkGBn"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Track_04QkGBn"
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   :List of 2
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Nest_KAdd13c"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Nest_KAdd13c"
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Track"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Track_04QkGBn"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:be1f0fd3-3d57-4144-9506-98335f56e4d8/Track_04QkGBn"
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 178
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "direct"
  .. .. ..$ source_id         : chr "2016-11-20-08-30-00-115-023-21-4612-fresh-chelonia-mydas"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-20T00:30:00Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "chelonia-mydas"
  .. .. ..$ habitat           : chr "beach-edge-of-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : chr "3 rotten eggs added to UH (unhatched eggs)"
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/178/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   :List of 2
  .. .. .. ..$ :List of 6
  .. .. .. .. ..$ observation_name: chr "tagobservation"
  .. .. .. .. ..$ tag_type        : chr "biopsy-sample"
  .. .. .. .. ..$ name            : chr "F5524"
  .. .. .. .. ..$ tag_location    : chr "flipper-front-left"
  .. .. .. .. ..$ status          : chr "removed"
  .. .. .. .. ..$ comments        : chr ""
  .. .. .. ..$ :List of 21
  .. .. .. .. ..$ observation_name    : chr "turtlenestobservation"
  .. .. .. .. ..$ nest_position       : chr "beach-edge-of-vegetation"
  .. .. .. .. ..$ eggs_laid           : logi FALSE
  .. .. .. .. ..$ egg_count           : int 138
  .. .. .. .. ..$ egg_count_calculated: int 123
  .. .. .. .. ..$ no_emerged          : int 58
  .. .. .. .. ..$ no_egg_shells       : int 73
  .. .. .. .. ..$ no_live_hatchlings  : int 9
  .. .. .. .. ..$ no_dead_hatchlings  : int 6
  .. .. .. .. ..$ no_undeveloped_eggs : int 36
  .. .. .. .. ..$ no_unhatched_eggs   : int 14
  .. .. .. .. ..$ no_unhatched_term   : NULL
  .. .. .. .. ..$ no_depredated_eggs  : NULL
  .. .. .. .. ..$ nest_depth_top      : int 850
  .. .. .. .. ..$ nest_depth_bottom   : NULL
  .. .. .. .. ..$ sand_temp           : NULL
  .. .. .. .. ..$ air_temp            : NULL
  .. .. .. .. ..$ water_temp          : NULL
  .. .. .. .. ..$ egg_temp            : NULL
  .. .. .. .. ..$ hatching_success    : int 59
  .. .. .. .. ..$ emergence_success   : int 47
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 522
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "odk"
  .. .. ..$ source_id         : chr "uuid:4a165d98-33b7-485c-888b-7f062ec49cc4"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-20T21:46:57.163000Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "cheloniidae-fam"
  .. .. ..$ habitat           : chr "in-dune-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : NULL
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/522/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   : list()
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 3337
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "direct"
  .. .. ..$ source_id         : chr "2016-11-24-05-25-00-114-9749-21-4571-old-chelonia-mydas"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-23T21:25:00Z"
  .. .. ..$ nest_age          : chr "old"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "chelonia-mydas"
  .. .. ..$ habitat           : chr "in-dune-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : chr ""
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/3337/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   : list()
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 3338
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "direct"
  .. .. ..$ source_id         : chr "2016-11-24-05-30-00-114-9744-21-4577-old-cheloniidae-fam"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-23T21:30:00Z"
  .. .. ..$ nest_age          : chr "old"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "cheloniidae-fam"
  .. .. ..$ habitat           : chr "in-dune-vegetation"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : chr ""
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/3338/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   : list()
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 115
  .. .. .. ..$ : num -21.5
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 3334
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "direct"
  .. .. ..$ source_id         : chr "2016-11-25-10-25-00-115-0166-21-4536-fresh-eretmochelys-imbricata"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -21.5
  .. .. ..$ longitude         : num 115
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-25T02:25:00Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "eretmochelys-imbricata"
  .. .. ..$ habitat           : chr "beach-above-high-water"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : chr ""
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/3334/change/"
  .. .. ..$ photographs       : list()
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   :List of 2
  .. .. .. ..$ :List of 21
  .. .. .. .. ..$ observation_name    : chr "turtlenestobservation"
  .. .. .. .. ..$ nest_position       : chr "beach-above-high-water"
  .. .. .. .. ..$ eggs_laid           : logi TRUE
  .. .. .. .. ..$ egg_count           : int 110
  .. .. .. .. ..$ egg_count_calculated: int 107
  .. .. .. .. ..$ no_emerged          : int 95
  .. .. .. .. ..$ no_egg_shells       : int 98
  .. .. .. .. ..$ no_live_hatchlings  : int 2
  .. .. .. .. ..$ no_dead_hatchlings  : int 1
  .. .. .. .. ..$ no_undeveloped_eggs : int 4
  .. .. .. .. ..$ no_unhatched_eggs   : int 5
  .. .. .. .. ..$ no_unhatched_term   : NULL
  .. .. .. .. ..$ no_depredated_eggs  : NULL
  .. .. .. .. ..$ nest_depth_top      : int 430
  .. .. .. .. ..$ nest_depth_bottom   : int 530
  .. .. .. .. ..$ sand_temp           : NULL
  .. .. .. .. ..$ air_temp            : NULL
  .. .. .. .. ..$ water_temp          : NULL
  .. .. .. .. ..$ egg_temp            : NULL
  .. .. .. .. ..$ hatching_success    : int 91
  .. .. .. .. ..$ emergence_success   : int 88
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "data_sheet"
  .. .. .. .. ..$ title           : chr "datasheet"
  .. .. .. .. ..$ attachment      : chr "encounter/2016-11-25-10-25-00-115-0166-21-4536-fresh-eretmochelys-imbricata/2016-11-25-10-25-00-115-0166-21-453"| __truncated__
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/2016-11-25-10-25-00-115-0166-21-4536-fresh-eretmochelys-imbricata/201"| __truncated__
  ..$ :List of 3
  .. ..$ type      : chr "Feature"
  .. ..$ geometry  :List of 2
  .. .. ..$ type       : chr "Point"
  .. .. ..$ coordinates:List of 2
  .. .. .. ..$ : num 116
  .. .. .. ..$ : num -20.4
  .. ..$ properties:List of 24
  .. .. ..$ pk                : int 1147
  .. .. ..$ site_visit        : NULL
  .. .. ..$ source            : chr "odk"
  .. .. ..$ source_id         : chr "uuid:53bcc7ae-4daa-48a3-839b-c480450d7500"
  .. .. ..$ encounter_type    : chr "nest"
  .. .. ..$ leaflet_title     : chr "2016 Nest "
  .. .. ..$ status            : chr "new"
  .. .. ..$ observer          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ reporter          :List of 4
  .. .. .. ..$ username: chr "test_name"
  .. .. .. ..$ name    : chr "Test Name"
  .. .. .. ..$ email   : chr "test@email.com"
  .. .. .. ..$ phone   : chr ""
  .. .. ..$ latitude          : num -20.4
  .. .. ..$ longitude         : num 116
  .. .. ..$ crs               : chr "WGS 84"
  .. .. ..$ location_accuracy : chr "10"
  .. .. ..$ when              : chr "2016-11-26T10:51:37.972000Z"
  .. .. ..$ nest_age          : chr "fresh"
  .. .. ..$ nest_type         : chr "hatched-nest"
  .. .. ..$ species           : chr "eretmochelys-imbricata"
  .. .. ..$ habitat           : chr "na"
  .. .. ..$ disturbance       : chr "absent"
  .. .. ..$ comments          : NULL
  .. .. ..$ absolute_admin_url: chr "/admin/observations/turtlenestencounter/1147/change/"
  .. .. ..$ photographs       :List of 1
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:53bcc7ae-4daa-48a3-839b-c480450d7500/Nest_glyrOiQ"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:53bcc7ae-4daa-48a3-839b-c480450d7500/Nest_glyrOiQ"
  .. .. ..$ tx_logs           : list()
  .. .. ..$ observation_set   :List of 1
  .. .. .. ..$ :List of 5
  .. .. .. .. ..$ observation_name: chr "mediaattachment"
  .. .. .. .. ..$ media_type      : chr "photograph"
  .. .. .. .. ..$ title           : chr "Nest"
  .. .. .. .. ..$ attachment      : chr "encounter/uuid:53bcc7ae-4daa-48a3-839b-c480450d7500/Nest_glyrOiQ"
  .. .. .. .. ..$ filepath        : chr "/mnt/projects/wastd/wastd/media/encounter/uuid:53bcc7ae-4daa-48a3-839b-c480450d7500/Nest_glyrOiQ"
 $ serializer: chr "turtle-nest-encounters"
 $ response  :List of 10
  ..$ url        : chr "https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=10&format=json&nest_type=hatched-nest"
  ..$ status_code: int 200
  ..$ headers    :List of 13
  .. ..$ date                     : chr "Tue, 02 May 2017 00:51:23 GMT"
  .. ..$ content-type             : chr "application/json"
  .. ..$ transfer-encoding        : chr "chunked"
  .. ..$ connection               : chr "keep-alive"
  .. ..$ server                   : chr "nginx"
  .. ..$ x-xss-protection         : chr "1; mode=block"
  .. ..$ x-content-type-options   : chr "nosniff"
  .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. ..$ vary                     : chr "Accept, Cookie"
  .. ..$ allow                    : chr "GET, POST, OPTIONS"
  .. ..$ x-frame-options          : chr "DENY"
  .. ..$ x-ua-compatible          : chr "chrome=1"
  .. ..$ content-encoding         : chr "gzip"
  .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  ..$ all_headers:List of 2
  .. ..$ :List of 3
  .. .. ..$ status : int 301
  .. .. ..$ version: chr "HTTP/1.1"
  .. .. ..$ headers:List of 11
  .. .. .. ..$ date                     : chr "Tue, 02 May 2017 00:51:23 GMT"
  .. .. .. ..$ content-type             : chr "text/html; charset=utf-8"
  .. .. .. ..$ transfer-encoding        : chr "chunked"
  .. .. .. ..$ connection               : chr "keep-alive"
  .. .. .. ..$ server                   : chr "nginx"
  .. .. .. ..$ x-xss-protection         : chr "1; mode=block"
  .. .. .. ..$ x-content-type-options   : chr "nosniff"
  .. .. .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. .. .. ..$ location                 : chr "/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=10&format=json&nest_type=hatched-nest"
  .. .. .. ..$ x-frame-options          : chr "DENY"
  .. .. .. ..$ x-ua-compatible          : chr "chrome=1"
  .. .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  .. ..$ :List of 3
  .. .. ..$ status : int 200
  .. .. ..$ version: chr "HTTP/1.1"
  .. .. ..$ headers:List of 13
  .. .. .. ..$ date                     : chr "Tue, 02 May 2017 00:51:23 GMT"
  .. .. .. ..$ content-type             : chr "application/json"
  .. .. .. ..$ transfer-encoding        : chr "chunked"
  .. .. .. ..$ connection               : chr "keep-alive"
  .. .. .. ..$ server                   : chr "nginx"
  .. .. .. ..$ x-xss-protection         : chr "1; mode=block"
  .. .. .. ..$ x-content-type-options   : chr "nosniff"
  .. .. .. ..$ strict-transport-security: chr "max-age=60; includeSubDomains"
  .. .. .. ..$ vary                     : chr "Accept, Cookie"
  .. .. .. ..$ allow                    : chr "GET, POST, OPTIONS"
  .. .. .. ..$ x-frame-options          : chr "DENY"
  .. .. .. ..$ x-ua-compatible          : chr "chrome=1"
  .. .. .. ..$ content-encoding         : chr "gzip"
  .. .. .. ..- attr(*, "class")= chr [1:2] "insensitive" "list"
  ..$ cookies    :'data.frame':	0 obs. of  7 variables:
  .. ..$ domain    : logi(0) 
  .. ..$ flag      : logi(0) 
  .. ..$ path      : logi(0) 
  .. ..$ secure    : logi(0) 
  .. ..$ expiration:Classes 'POSIXct', 'POSIXt'  num(0) 
  .. ..$ name      : logi(0) 
  .. ..$ value     : logi(0) 
  ..$ content    : raw [1:14122] 7b 22 74 79 ...
  ..$ date       : POSIXct[1:1], format: "2017-05-02 00:51:23"
  ..$ times      : Named num [1:6] 0.059755 0.000032 0.000036 0.000127 0.204033 ...
  .. ..- attr(*, "names")= chr [1:6] "redirect" "namelookup" "connect" "pretransfer" ...
  ..$ request    :List of 7
  .. ..$ method    : chr "GET"
  .. ..$ url       : chr "https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters?taxon=Cheloniidae&limit=10&format=json&nest_type=hatched-nest"
  .. ..$ headers   : Named chr [1:2] "application/json, text/xml, application/xml, */*" "Token c5575c09f6a0171668f31d5dd5013f02658668bd"
  .. .. ..- attr(*, "names")= chr [1:2] "Accept" "Authorization"
  .. ..$ fields    : NULL
  .. ..$ options   :List of 2
  .. .. ..$ useragent: chr "http://github.com/parksandwildlife/turtle-scripts"
  .. .. ..$ httpget  : logi TRUE
  .. ..$ auth_token: NULL
  .. ..$ output    : list()
  .. .. ..- attr(*, "class")= chr [1:2] "write_memory" "write_function"
  .. ..- attr(*, "class")= chr "request"
  ..$ handle     :Class 'curl_handle' <externalptr> 
  ..- attr(*, "class")= chr "response"
 - attr(*, "class")= chr "wastd_api_response"
}}
\source{
https://strandings.dpaw.wa.gov.au/api/1/turtle-nest-encounters/?taxon=Cheloniidae&limit=10&format=json&nest_type__exact=hatched-nest
}
\usage{
turtle_nest_encounters_hatched
}
\description{
This API response is parsed into wastr's data "nests".
}
\examples{
# Prove that turtle_nest_encounters_hatched parses to nests
library(dplyr)
data(turtle_nest_encounters_hatched)
data(nests)
fresh_nests <- parse_turtle_nest_encounters(turtle_nest_encounters_hatched)
testthat::expect_equal(nrow(fresh_nests), nrow(nests))
# Compare pickled and fresh nests excluding list columns (like obs)
testthat::expect_equal(fresh_nests \%>\% dplyr::select(-obs),
                       nests \%>\% dplyr::select(-obs))
}
\keyword{datasets}

