% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_odkc_to_tsc.R
\name{upload_odkc_to_tsc}
\alias{upload_odkc_to_tsc}
\title{Top level helper to upload all Turtle Nesting Census data to TSC}
\usage{
upload_odkc_to_tsc(odkc_prep, tsc_data, update_existing = FALSE)
}
\arguments{
\item{odkc_prep}{ODKC data transformed into TSC format.}

\item{tsc_data}{Minimal TSC data to inform skip logic.}
}
\value{
A list of results from the various uploads.
  Each result is a `wastd_api_response`.
}
\description{
Top level helper to upload all Turtle Nesting Census data to TSC
}
