% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upsert_geojson.R
\name{upsert_geojson}
\alias{upsert_geojson}
\title{Upsert GeoJSON into WAStD API endpoints}
\usage{
upsert_geojson(
  gj_featurecollection,
  serializer = "names",
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  chunksize = 1000,
  verbose = FALSE
)
}
\arguments{
\item{gj_featurecollection}{(list) A GeoJSON featurecollection as list}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values as per
\code{https://strandings.dpaw.wa.gov.au/api/1/?format=corejson} are:
\itemize{
\item encounters (all encounters, but only core fields)
\item animal-encounters (strandings, tagging)
\item turtle-nest-encounters (tracks and nests)
\item logger-encounters (temp and other loggers)
\item areas (polygons of known areas)
\item media-attachments (photos, data sheets etc)
\item nesttag-observations (sightings of nest tags)
\item tag-observations (tag observations during encounters)
\item taxa (WACensus taxa)
}}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{chunksize}{(int) The number of features to upload simultaneously, default: 1000.}

\item{verbose}{Whether to show debug messages.}
}
\description{
Upsert GeoJSON into WAStD API endpoints
}
