% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_ae}
\alias{wastd_ae}
\title{WAStD AnimalEncounters, parsed, test data captured by author}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 22 rows and 68 columns.
}
\usage{
wastd_ae
}
\description{
WAStD AnimalEncounters, parsed, test data captured by author
}
\seealso{
Other included: 
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wamtram_data}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
