% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_ae_raw}
\alias{wastd_ae_raw}
\title{WAStD AnimalEncounters, unparsed, test data captured by author}
\format{
An object of class \code{wastd_api_response} of length 5.
}
\usage{
wastd_ae_raw
}
\description{
WAStD AnimalEncounters, unparsed, test data captured by author
}
\seealso{
Other included: 
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wamtram_data}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_data}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
