% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_bulk_post.R
\name{wastd_bulk_post}
\alias{wastd_bulk_post}
\title{Post a list of records to a WAStD / TSC API endpoint}
\usage{
wastd_bulk_post(
  data,
  serializer,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = FALSE
)
}
\arguments{
\item{data}{A tibble of a data with columns equal to the serializer's
fields.}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
The list of \code{wastd_api_response}s from \code{\link{wastd_POST}}
}
\description{
\lifecycle{maturing}
}
\examples{
\dontrun{
odkc_data$tracks \%>\%
  odkc_tracks_as_wastd_tne() \%>\%
  wastd_bulk_post("turtle-nest-encounters")
}
}
\concept{("api")}
