% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_chunk_post.R
\name{wastd_chunk_post}
\alias{wastd_chunk_post}
\title{POST data to a WAStD API serializer in chunks}
\usage{
wastd_chunk_post(
  data,
  serializer,
  query = list(),
  chunksize = 1000,
  encode = "json",
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{Data to post to the serializer}

\item{serializer}{A WAStD API serializer, e.g. "encounters"}

\item{query}{An optional query for \code{\link{wastd_POST}},
default: \code{list()}.}

\item{chunksize}{The number of records to post at a time, default: 1000.}

\item{encode}{The parameter \code{encode} for \code{link{httr::POST}},
default: "json".
Other options: \code{c("multipart", "form", "json", "raw")}.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
The \code{wastd_api_response} of the last batch of data.
}
\description{
POST data to a WAStD API serializer in chunks
}
\examples{
\dontrun{
# ODKC data as WAStD data, chunk_post
}
}
