% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_create_update_skip.R
\name{wastd_create_update_skip}
\alias{wastd_create_update_skip}
\title{Create, update, skip records to a WAStD serializer.}
\usage{
wastd_create_update_skip(
  data_create,
  data_update,
  data_skip,
  update_existing = FALSE,
  label = "records",
  serializer = "encounters",
  chunksize = 1000,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data_create}{Records to create.}

\item{data_update}{Records to update.}

\item{data_skip}{Records to skip.}

\item{update_existing}{Whether to update existing but unchanged records in
WAStD. Default: FALSE.}

\item{label}{The model label for verbose messages, default: \code{records}.}

\item{serializer}{The WAStD API serializer, default: \code{encounters}.}

\item{chunksize}{The number of records to post at a time, default: 1000.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A list of three objects:
\itemize{
\item \code{created} The WAStD API response for all created records
\item \code{updated} The WAStD API response for all updated records
\item \code{skipped} The original records for all skipped records
}
}
\description{
Handles empty data, verbose notifications.
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_sites}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata_area}()},
\code{\link{filter_wastd_turtledata_seasons}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{handle_http_status}()},
\code{\link{wastd_GET}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_bulk_post}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
