% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_data}
\alias{wastd_data}
\title{Turtle Nesting Census data from WAStD}
\format{An object of class \code{wastd_data} of length 13.}
\usage{
wastd_data
}
\description{
The output of \code{\link{download_wastd_turtledata}}, but
observer and reporter names are hidden, tibbles are filtered down to the
first 10 records, however all areas and sites are included.
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
