% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_data}
\alias{wastd_data}
\title{Turtle Nesting Census data from WAStD}
\format{
An object of class \code{wastd_data} of length 21.
}
\usage{
wastd_data
}
\description{
This dataset contains the truncated output from
\code{\link{download_wastd_turtledata}}.
Names of observer and reporter are hidden.
All WAStD sites and areas are included, while all other tables are truncated
with \code{head()}.
}
\seealso{
Other included: 
\code{\link{download_odkc_turtledata_2019}()},
\code{\link{download_odkc_turtledata_2020}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{odkc_data}},
\code{\link{print.odkc_turtledata}()},
\code{\link{print.wastd_data}()},
\code{\link{tsc_data}},
\code{\link{wastd_ae_raw}},
\code{\link{wastd_ae}},
\code{\link{wastd_area_raw}},
\code{\link{wastd_surveys_raw}},
\code{\link{wastd_tne_raw}},
\code{\link{wastd_tne}}
}
\concept{included}
\keyword{datasets}
