% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_GET.R
\name{wastd_GET}
\alias{wastd_GET}
\title{Return GeoJSON features from a WAStD API endpoint}
\usage{
wastd_GET(
  serializer,
  query = list(),
  format = "json",
  max_records = NULL,
  chunk_size = 1000,
  api_url = get_wastdr_api_url(),
  api_token = get_wastdr_api_token(),
  api_un = get_wastdr_api_un(),
  api_pw = get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{serializer}{(character) WAStD API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{query}{A list of GET parameters, default: list().
The \code{format} is specified in a separate top-level param.}

\item{format}{The desired API output format, default: "json".}

\item{max_records}{The maximum number of records retrieved.
If left at default (NULL), all records are returned.
Default: NULL.}

\item{chunk_size}{The number of records to retrieve in each paginated
response. A specified but smaller \code{limit} will override
\code{chunk_size}.
Adjust \code{chunk_size} down if getting timeouts from the API.
Default: 1000.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 object of class 'wastd_api_response' containing:

data: The parsed response content

serializer: The called serializer, e.g. 'supra'

url: The called URL

date: The timestamp of the request

status_code: The HTTP status code of the request
}
\description{
Call the WAStD API serializer's list view with given GET
parameters, parse the response's features into a nested list.
This function requires the WAStD API to return the results in a key
\code{features} (if GeoJSON) or \code{data} (if JSON).
}
\details{
\lifecycle{stable}
}
\examples{
\dontrun{
track_records <- wastd_GET("turtle-nest-encounters")
tag_records <- wastd_GET("animal-encounters")
hatched_nest_records <- wastd_GET("turtle-nest-encounters",
  query = list(nest_type = "hatched-nest")
)
}
}
\seealso{
Other api: 
\code{\link{build_auth}()},
\code{\link{download_minimal_wastd_turtledata}()},
\code{\link{download_wastd_sites}()},
\code{\link{download_wastd_turtledata}()},
\code{\link{download_wastd_users}()},
\code{\link{export_wastd_turtledata}()},
\code{\link{filter_wastd_turtledata}()},
\code{\link{handle_http_status}()},
\code{\link{wastd_POST}()},
\code{\link{wastd_bulk_post}()},
\code{\link{wastd_chunk_post}()},
\code{\link{wastd_create_update_skip}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
