% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_POST.R
\name{wastd_POST}
\alias{wastd_POST}
\title{Send a POST request to WAStD's API}
\usage{
wastd_POST(
  data,
  serializer,
  query = list(format = "json"),
  encode = "json",
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{(JSON) A list of lists (JSON) to post to WAStD.}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values see \code{https://tsc.dbca.wa.gov.au/api/1/}}

\item{query}{<list> A list of POST parameters,
default: \code{list(format="json")}.}

\item{encode}{The parameter \code{encode} for \code{link{httr::POST}},
default: "json".
Other options: \code{c("multipart", "form", "json", "raw")}.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 object of class 'wastd_api_response' containing:

  features: The sent GeoJSON features

  serializer: The called serializer, e.g. 'supra'

  response: The API HTTP response with all metadata
}
\description{
\lifecycle{stable}
}
\examples{
\dontrun{
# One by one
gj <- "public:herbie_hbvnames_public" \%>\% kmi_getFeature()
props <- purrr::map(gj[["features"]], "properties")
wastd_POST(props[[1]], serializer = "names")

# All in batch
"public:herbie_hbvnames_public" \%>\%
  kmi_getFeature() \%>\%
  wastd_upsert_geojson(serializer = "names", verbose = T)
}
}
\seealso{
Other api: 
\code{\link{wastd_GET}()},
\code{\link{wastd_parse}()},
\code{\link{wastd_post_one}()}
}
\concept{api}
