% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_POST.R
\name{wastd_POST}
\alias{wastd_POST}
\title{Send a POST request to WAStD's API}
\usage{
wastd_POST(
  data,
  serializer,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  api_un = wastdr::get_wastdr_api_un(),
  api_pw = wastdr::get_wastdr_api_pw(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{(JSON) A list of lists (JSON) to post to WAStD.}

\item{serializer}{(character) WAStD API serializer name (required)
Possible values as per
\code{https://strandings.dpaw.wa.gov.au/api/1/?format=corejson} are:
\itemize{
\item encounters (all encounters, but only core fields)
\item animal-encounters (strandings, tagging)
\item turtle-nest-encounters (tracks and nests)
\item logger-encounters (temp and other loggers)
\item areas (polygons of known areas)
\item media-attachments (photos, data sheets etc)
\item nesttag-observations (sightings of nest tags)
\item tag-observations (tag observations during encounters)
\item taxa (WACensus taxa)
}}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
An S3 object of class 'wastd_api_response' containing:

  features: The sent GeoJSON features

  serializer: The called serializer, e.g. 'supra'

  response: The API HTTP response with all metadata
}
\description{
Send a POST request to WAStD's API
}
\examples{
\dontrun{
# One by one
gj <- "public:herbie_hbvnames_public" \%>\% kmi_getFeature()
props <- purrr::map(gj[["features"]], "properties")
wastd_POST(props[[1]], serializer = "names")

# All in batch
"public:herbie_hbvnames_public" \%>\%
  kmi_getFeature() \%>\%
  wastd_upsert_geojson(serializer = "names", verbose = T)
}
}
\seealso{
Other api: 
\code{\link{wastd_GET}()},
\code{\link{wastd_parse}()}
}
\concept{api}
