% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wastd_tne_raw}
\alias{wastd_tne_raw}
\title{WAStD TurtleNestEncounters, unparsed, test data captured by author}
\format{An object of class \code{wastd_api_response} of length 5.}
\usage{
wastd_tne_raw
}
\description{
WAStD TurtleNestEncounters, unparsed, test data captured by author
}
\keyword{datasets}
