% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_msg.R
\name{wastdr_msg_warn}
\alias{wastdr_msg_warn}
\title{rlang::warn() with a yellow warning message with a warning symbol.}
\usage{
wastdr_msg_warn(message)
}
\arguments{
\item{message}{(chr) A message to print}
}
\description{
rlang::warn() with a yellow warning message with a warning symbol.
}
\examples{
\dontrun{
wastdr_msg_warn("This is a warning.")
}
}
\seealso{
Other helpers: 
\code{\link{add_dates}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{odkc_works}()},
\code{\link{sf_as_tbl}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastd_works}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_settings}()},
\code{\link{wastdr_setup}()}
}
\concept{helpers}
