% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_settings.R
\name{wastdr_setup}
\alias{wastdr_setup}
\title{Configure default WAStD settings}
\usage{
wastdr_setup(
  wastd_url = get_wastd_url(),
  api_url = get_wastdr_api_url(),
  api_token = NULL,
  api_un = get_wastdr_api_un(),
  api_pw = get_wastdr_api_pw()
)
}
\arguments{
\item{wastd_url}{A WAStD URL (optional),
default: "https://tsc.dbca.wa.gov.au"}

\item{api_url}{A WAStD API URL (optional),
default: "https://tsc.dbca.wa.gov.au/api/1/"}

\item{api_token}{A CKAN API token, leading with "Token " (character)}

\item{api_un}{Alternatively, a CKAN API username (character)}

\item{api_pw}{The password to the CKAN username (character)}
}
\description{
Configure default WAStD settings
}
\details{
\code{wastdr_setup} sets WAStD API connection details. \code{wastdr}'s
functions default to use the default URL, but require an API key to work.
}
\examples{
# WAStD users with a DBCA account can access their WAStD authentication
# token on their profile page on WAStD:
wastdr_setup(api_token = "wastd_token")

# Non-DBCA users will have been given a WAStD username and password:
wastdr_setup(
  api_un = "wastd_username",
  api_pw = "wastd_password"
)

# Not specifying the default WAStD API URL will reset the WAStD URL to its
# default "https://tsc.dbca.wa.gov.au/api/1/":
\dontrun{wastdr_setup(api_token = "c12345asdfqwer")}
}
