% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastdr_settings.R
\name{wastdr_setup}
\alias{wastdr_setup}
\title{Configure default WAStD settings}
\usage{
wastdr_setup(
  wastd_url = get_wastd_url(),
  api_url = get_wastdr_api_url(),
  api_token = NULL,
  api_un = get_wastdr_api_un(),
  api_pw = get_wastdr_api_pw()
)
}
\arguments{
\item{wastd_url}{A WAStD URL (optional),
default: "https://tsc.dbca.wa.gov.au"}

\item{api_url}{A WAStD API URL (optional),
default: "https://tsc.dbca.wa.gov.au/api/1/"}

\item{api_token}{A CKAN API token, leading with "Token " (character)}

\item{api_un}{Alternatively, a CKAN API username (character)}

\item{api_pw}{The password to the CKAN username (character)}
}
\description{
Configure default WAStD settings
}
\details{
\code{wastdr_setup} sets WAStD API connection details. \code{wastdr}'s
functions default to use the default URL, but require an API key to work.
}
\examples{
# WAStD users with a DBCA account can access their WAStD authentication
# token on their profile page on WAStD:
wastdr_setup(api_token = "wastd_token")

# Non-DBCA users will have been given a WAStD username and password:
wastdr_setup(
  api_un = "wastd_username",
  api_pw = "wastd_password"
)

# Not specifying the default WAStD API URL will reset the WAStD URL to its
# default "https://tsc.dbca.wa.gov.au/api/1/":
\dontrun{
wastdr_setup(api_token = "c12345asdfqwer")
}
}
\seealso{
Other helpers: 
\code{\link{\%>\%}()},
\code{\link{add_dates_sve}()},
\code{\link{add_dates_svs}()},
\code{\link{add_dates}()},
\code{\link{chr2int}()},
\code{\link{datetime_as_isoweek}()},
\code{\link{datetime_as_seasonweek}()},
\code{\link{datetime_as_season}()},
\code{\link{datetime_as_turtle_date}()},
\code{\link{exclude_training_species}()},
\code{\link{exclude_training_surveys}()},
\code{\link{extract_possibly}()},
\code{\link{fdate_as_tdate}()},
\code{\link{filter_missing_site}()},
\code{\link{filter_missing_survey}()},
\code{\link{filter_surveys_missing_end}()},
\code{\link{filter_surveys_requiring_qa}()},
\code{\link{filter_wastd_season}()},
\code{\link{get_more}()},
\code{\link{get_num_field}()},
\code{\link{httpdate_as_gmt08_turtle_date}()},
\code{\link{httpdate_as_gmt08}()},
\code{\link{httpdate_as_season}()},
\code{\link{humanize}()},
\code{\link{join_tsc_sites}()},
\code{\link{map_chr_hack}()},
\code{\link{sf_as_tbl}()},
\code{\link{skip_test_if_odkc_offline}()},
\code{\link{skip_test_if_wastd_offline}()},
\code{\link{tdate_as_fdate}()},
\code{\link{tracks_as_geojson}()},
\code{\link{urlize}()},
\code{\link{wastdr_msg_abort}()},
\code{\link{wastdr_msg_info}()},
\code{\link{wastdr_msg_noop}()},
\code{\link{wastdr_msg_success}()},
\code{\link{wastdr_msg_warn}()},
\code{\link{wastdr_settings}()}
}
\concept{helpers}
